Depth calculation

Calculate depth functions.

depth(u, X, method = "Projection", threads = -1, ...)

Arguments

u
Numerical vector or matrix whose depth is to be calculated. Dimension has to be the same as that of the observations.

X
The data as a matrix, data frame or list. If it is a matrix or data frame, then each row is viewed as one multivariate observation. If it is a list, all components must be numerical vectors of equal length (coordinates of observations).

method
Character string which determines the depth function. method can be "Projection" (the default), "Mahalanobis", "Euclidean" or "Tukey". For details see depth.

threads
number of threads used in parallel computations. Default value -1 means that all possible cores will be used.

...
parameters specific to method --- see depthEuclid

Details

The Mahalanobis depth $$ {D}_{MAH}(y, {X} ^ {n}) = \frac{ 1 }{ 1 + {{(y - \bar{x})} ^ {T}}{{S} ^ {-1}}(y - \bar{x}) }, $$ where \( S \) denotes the sample covariance matrix \( {X} ^ {n} \).

A symmetric projection depth \( D\left( x, X\right) \) of a point \( x \in {{{R}} ^ {d}} \), \( d \ge 1 \) is defined as $$ D\left( x, X\right)_{PRO} = {{\left[ 1 + su{{p}_{\left\| u \right\| = 1}}\frac{ \left| {{u} ^ {T}}x - Med\left( {{u} ^ {T}}X\right)\right| }{ MAD\left( {{u} ^ {T}}X\right) }\right]} ^ {-1}}, $$ where Med denotes the univariate median, \( MAD\left( Z \right) \) = \( Med\left(\left| Z - Med\left( Z \right)\right|\right) \). Its sample version denoted by \( D\left( x, {X} ^ {n} \right) \) or \( D\left( x, {X} ^ {n} \right) \) is obtained by replacing \( F \) by its empirical counterpart \( {{F}_{n}} \) calculated from the sample \( {X} ^ {n} \) .

Next interesting depth is the weighted \( {L} ^ {p} \) depth. The weighted \( {L} ^ {p} \) depth \( D({x}, F) \) of a point \( {x} \in {R} ^ {d} \), \( d \ge 1 \) generated by \( d \) dimensional random vector \( {X} \) with distribution \( F \), is defined as \( D({x}, F) = \frac{1 }{ 1 + Ew({{\left\| x - X \right\| }_{p}}) }, \) where \( w \) is a suitable weight function on \( [0, \infty) \), and \( {{\left\| \cdot \right\| }_{p}} \) stands for the \( {L} ^ {p} \) norm (when p = 2 we have usual Euclidean norm). We assume that \( w \) is non-decreasing and continuous on \( [0, \infty) \) with \( w(\infty-) = \infty \), and for \( a, b \in {{{R}} ^ {d}} \) satisfying \( w(\left\| a + b \right\|) \le w(\left\| a \right\|) + w(\left\| b \right\|) \). Examples of the weight functions are: \( w(x) = a + bx \), \( a, b > 0 \) or \( w(x) = {x} ^ {\alpha} \). The empirical version of the weighted \( {L} ^ {p} \) depth is obtained by replacing distribution \( F \) of \( {X} \) in \( Ew({{\left\| {x} - {X} \right\| }_{p}}) = \int {w({{\left\| x - t \right\| }_{p}})}dF(t) \) by its empirical counterpart calculated from the sample \( {{{X}} ^ {n}} \)...

The Projection and Tukey's depths are calculated using an approximate algorithm. Calculations of Mahalanobis, Euclidean and \( L ^ p \) depths are exact. Returns the depth of multivariate point u with respect to data set X.

References

Liu, R.Y., Parelius, J.M. and Singh, K. (1999), Multivariate analysis by data depth: Descriptive statistics, graphics and inference (with discussion), Ann. Statist., 27, 783--858.

Mosler K (2013). Depth statistics. In C Becker, R Fried, K S (eds.), Robustness and Complex Data Structures, Festschrift in Honour of Ursula Gather, pp. 17--34. Springer.

Rousseeuw, P.J. and Struyf, A. (1998), Computing location depth and regression depth in higher dimensions, Stat. Comput., 8, 193--203.

Zuo, Y. and Serfling, R. (2000), General Notions of Statistical Depth Functions, Ann. Statist., 28, no. 2, 461--482.

See also

Examples

library(robustbase)

# Calculation of Projection depth
data(starsCYG, package = "robustbase")
depth(t(colMeans(starsCYG)), starsCYG)

#> Depth method:  Projection 
#> [1] 0.3933721

# Also for matrices
depth(starsCYG, starsCYG)

#> Depth method:  Projection 
#>  [1] 0.44549966 0.32661309 0.31866508 0.32661309 0.31385837 0.45420971
#>  [7] 0.10939758 0.31518707 0.20226849 0.51764792 0.05842178 0.41883831
#> [13] 0.46077383 0.18152957 0.30080943 0.35557245 0.23498109 0.23610413
#> [19] 0.27354830 0.05681380 0.33064804 0.29202504 0.29581123 0.35521715
#> [25] 0.65582766 0.39547262 0.43022524 0.64818780 0.31720408 0.05483175
#> [31] 0.33282390 0.29535294 0.57655832 0.05315617 0.30715723 0.29548863
#> [37] 0.34409153 0.57655832 0.36620487 0.36102094 0.43568694 0.58306346
#> [43] 0.45014721 0.52446268 0.36090119 0.52894081 0.32298565

# Projection depth applied to a large bivariate data set
x <- matrix(rnorm(9999), nc = 3)
depth(x, x)

#> Depth method:  Projection 
#>    [1] 0.1845190 0.4281967 0.2671678 0.3372862 0.2418110 0.2124226 0.3519991
#>    [8] 0.3526441 0.4891957 0.3164199 0.4043022 0.1971379 0.7005986 0.2078889
#>   [15] 0.2564805 0.4181578 0.2290001 0.5679235 0.2771307 0.3864750 0.2873990
#>   [22] 0.3541572 0.4260452 0.5365237 0.4133832 0.2941509 0.6257109 0.3269003
#>   [29] 0.3048097 0.2981967 0.2926064 0.4656271 0.3805627 0.3087697 0.2657323
#>   [36] 0.3495137 0.2504115 0.2814374 0.2956905 0.5025439 0.2718784 0.3484203
#>   [43] 0.2225834 0.2693907 0.2233572 0.3621018 0.5311404 0.2063794 0.2653592
#>   [50] 0.2940294 0.2609336 0.3594534 0.1734951 0.2866222 0.2417179 0.2243652
#>   [57] 0.3731939 0.3376548 0.2110528 0.4203788 0.2726943 0.2221815 0.2924362
#>   [64] 0.3661927 0.4946005 0.4796661 0.4543497 0.3585076 0.2556665 0.2537608
#>   [71] 0.5618957 0.3700877 0.2372446 0.2907878 0.4106071 0.2886319 0.2847436
#>   [78] 0.3067565 0.2545424 0.2459805 0.1996566 0.2405760 0.2150456 0.2569123
#>   [85] 0.2526983 0.2601024 0.2885339 0.2539185 0.2332760 0.2364308 0.1897568
#>   [92] 0.2896208 0.1657375 0.6732563 0.3602987 0.2318014 0.3247344 0.2848245
#>   [99] 0.2339838 0.2602378 0.2442183 0.3538574 0.2061432 0.4402646 0.3292287
#>  [106] 0.3895336 0.2009060 0.4829883 0.3305247 0.4746754 0.3543150 0.2414695
#>  [113] 0.2693422 0.4445000 0.6353667 0.1952752 0.3500774 0.2213084 0.3098862
#>  [120] 0.4065290 0.2374105 0.3224084 0.2122482 0.2890680 0.4222351 0.4077482
#>  [127] 0.3688915 0.3226213 0.3023089 0.3974173 0.6488335 0.2447555 0.2596794
#>  [134] 0.2983569 0.5723977 0.3449075 0.2625057 0.2423859 0.2842148 0.4271871
#>  [141] 0.2865055 0.4183988 0.2563233 0.3099550 0.3663075 0.2721641 0.1931834
#>  [148] 0.3320279 0.5683435 0.3742455 0.2394813 0.2729494 0.2502137 0.3213311
#>  [155] 0.2333745 0.2417565 0.2673545 0.2415863 0.1741788 0.5298226 0.4263969
#>  [162] 0.2480876 0.2467550 0.3889226 0.2227919 0.2425621 0.2447751 0.3550741
#>  [169] 0.2728910 0.3494554 0.2514711 0.3759904 0.2452173 0.3233282 0.2924279
#>  [176] 0.3680012 0.3942751 0.2968137 0.1873707 0.2150559 0.4734643 0.3978257
#>  [183] 0.2974623 0.3206902 0.3884352 0.2386943 0.2236503 0.3693675 0.1510487
#>  [190] 0.3036775 0.4004086 0.4894591 0.3099426 0.2338256 0.3645569 0.4457695
#>  [197] 0.3029467 0.2068915 0.2454499 0.2410354 0.3288835 0.3890324 0.2333287
#>  [204] 0.2832336 0.3174661 0.3911480 0.1950230 0.2961249 0.3382049 0.2426289
#>  [211] 0.8029568 0.1882239 0.2576019 0.2834153 0.3675853 0.4576466 0.2768000
#>  [218] 0.2303284 0.3303570 0.3751916 0.3416982 0.3099725 0.2244702 0.3065330
#>  [225] 0.2685666 0.3972128 0.2204936 0.3267812 0.2054834 0.2577135 0.2290163
#>  [232] 0.2740143 0.2382694 0.2648850 0.2147095 0.3277423 0.2515654 0.3182025
#>  [239] 0.2903845 0.2168842 0.2853818 0.4616947 0.1849651 0.1923182 0.2301251
#>  [246] 0.3397628 0.2104452 0.2994762 0.4194235 0.2172601 0.1859370 0.4526879
#>  [253] 0.3088907 0.3156628 0.3344175 0.1727014 0.2975852 0.4088852 0.2996231
#>  [260] 0.6064299 0.6543807 0.5061829 0.3222277 0.3562050 0.2676610 0.6368654
#>  [267] 0.2181023 0.4149742 0.3930541 0.2553745 0.2536013 0.3014653 0.2327171
#>  [274] 0.2848071 0.3051612 0.1563927 0.2638892 0.2575632 0.3019372 0.2236065
#>  [281] 0.3400539 0.2706347 0.2647844 0.3126667 0.4224971 0.2148583 0.3602023
#>  [288] 0.6064147 0.3130268 0.1606245 0.5741952 0.2807773 0.2087249 0.2908831
#>  [295] 0.2541668 0.2902248 0.2872977 0.4690581 0.4305463 0.3907464 0.2475314
#>  [302] 0.2678168 0.2479503 0.2703995 0.3296049 0.2523954 0.2978527 0.3452175
#>  [309] 0.3487774 0.4447782 0.4581007 0.1848621 0.1896305 0.2589421 0.4528320
#>  [316] 0.2816516 0.2236030 0.2417580 0.2042860 0.5656190 0.2516753 0.3737154
#>  [323] 0.3111808 0.7012480 0.3917724 0.2073184 0.5686493 0.2613416 0.4930591
#>  [330] 0.2938282 0.6207992 0.3024579 0.4109024 0.5221103 0.2426047 0.3441208
#>  [337] 0.2917612 0.3128637 0.2739241 0.2594487 0.2159957 0.3218592 0.1834553
#>  [344] 0.4700198 0.2914700 0.3201863 0.2151120 0.2904919 0.3840082 0.1856987
#>  [351] 0.5443213 0.3333104 0.1543795 0.3373947 0.4534800 0.2878866 0.4032158
#>  [358] 0.2744902 0.2942155 0.2333281 0.1538287 0.2336424 0.2384908 0.2971702
#>  [365] 0.5127048 0.2146134 0.2684170 0.2790775 0.2537448 0.3585323 0.5836358
#>  [372] 0.4400867 0.2808924 0.2689352 0.3105498 0.2474155 0.1816489 0.2757146
#>  [379] 0.2554611 0.3518000 0.3742929 0.4060062 0.1585864 0.3593158 0.3073864
#>  [386] 0.3599121 0.3552437 0.2634529 0.4925659 0.2187833 0.5762982 0.3582276
#>  [393] 0.3247672 0.3809009 0.6886522 0.5879270 0.3707317 0.2136411 0.1747349
#>  [400] 0.3402503 0.3941280 0.1919294 0.4171644 0.3410845 0.5720476 0.2837105
#>  [407] 0.3829258 0.2875731 0.1867681 0.3946677 0.2878471 0.5270724 0.4579141
#>  [414] 0.2349905 0.2212902 0.2877501 0.2708827 0.3021194 0.3925118 0.3831293
#>  [421] 0.2984374 0.2303359 0.2559186 0.3712755 0.2376847 0.3106328 0.5065443
#>  [428] 0.2175178 0.2531274 0.3040291 0.2625031 0.2949805 0.2231113 0.3956645
#>  [435] 0.3230945 0.2539729 0.2257215 0.4351627 0.2124238 0.2502858 0.2723718
#>  [442] 0.6150358 0.2747929 0.4740974 0.4027572 0.4280211 0.3160788 0.3128857
#>  [449] 0.3336164 0.3810990 0.2401715 0.3215185 0.2701070 0.2730927 0.3485666
#>  [456] 0.2151129 0.2900858 0.2183880 0.4663757 0.1903897 0.3227707 0.3462432
#>  [463] 0.3151237 0.4223572 0.5373769 0.4464947 0.4473650 0.2639757 0.2498546
#>  [470] 0.2986434 0.3007556 0.4394211 0.2270471 0.2366571 0.3241284 0.3817528
#>  [477] 0.2765440 0.3035182 0.2964275 0.2703425 0.3995810 0.3269261 0.2851160
#>  [484] 0.2819599 0.5310413 0.4064327 0.5939282 0.5101569 0.2449473 0.2998368
#>  [491] 0.4256630 0.3472973 0.3130008 0.2329855 0.3580072 0.1784082 0.3978188
#>  [498] 0.2333162 0.2428879 0.2313156 0.3612910 0.2598073 0.3238021 0.4093852
#>  [505] 0.4609756 0.3983535 0.4448745 0.2486347 0.3802506 0.1962745 0.2061809
#>  [512] 0.3446741 0.1943245 0.2592695 0.2814581 0.2791000 0.2080948 0.2348686
#>  [519] 0.2825802 0.2883923 0.2098888 0.1673010 0.2720162 0.2668597 0.3138271
#>  [526] 0.5178551 0.3616918 0.2005764 0.4157286 0.3894758 0.3234589 0.7134578
#>  [533] 0.3307952 0.2146569 0.3357010 0.1925646 0.2638458 0.4609564 0.3271058
#>  [540] 0.2006200 0.4361238 0.4140430 0.2093084 0.2256918 0.2294029 0.2445863
#>  [547] 0.5045420 0.2321477 0.2802687 0.2592121 0.2335558 0.1974933 0.1875454
#>  [554] 0.5429314 0.3996641 0.3221777 0.3928572 0.2971845 0.2225559 0.2152568
#>  [561] 0.7332289 0.5277803 0.2871948 0.3202552 0.2266471 0.2215787 0.3384800
#>  [568] 0.3163030 0.3572346 0.2633491 0.4269712 0.3396999 0.3864691 0.3356917
#>  [575] 0.2960575 0.3206041 0.4954426 0.2938207 0.4027565 0.4482919 0.3076210
#>  [582] 0.2872941 0.3877871 0.3308872 0.2038889 0.2521345 0.2925993 0.3027994
#>  [589] 0.3322434 0.2938121 0.2893869 0.2555616 0.2186788 0.3551937 0.2649760
#>  [596] 0.2605117 0.1799729 0.3883274 0.2595631 0.3515423 0.2997360 0.2066741
#>  [603] 0.2798252 0.3390954 0.4927706 0.4558939 0.3059125 0.3503318 0.2193071
#>  [610] 0.2080831 0.4655519 0.4662796 0.4616189 0.3082471 0.2587455 0.3138342
#>  [617] 0.2259176 0.2122950 0.2104942 0.3821415 0.2539755 0.2599571 0.3426324
#>  [624] 0.4790012 0.2179924 0.2847884 0.1519327 0.2414956 0.2754552 0.3427686
#>  [631] 0.4015645 0.5793431 0.3403892 0.2370240 0.3374007 0.6333649 0.2058591
#>  [638] 0.7072838 0.4869763 0.3359589 0.3702975 0.3681144 0.1906902 0.5358905
#>  [645] 0.2399921 0.3781384 0.2029036 0.2669208 0.3662491 0.2104685 0.2942182
#>  [652] 0.5568019 0.2333593 0.2343123 0.4690473 0.2695078 0.2751632 0.3626658
#>  [659] 0.2721129 0.6091970 0.5201360 0.2244344 0.3368470 0.2681069 0.4903372
#>  [666] 0.4927584 0.4001810 0.3449188 0.3818706 0.3519026 0.2892658 0.4259874
#>  [673] 0.2161114 0.2405573 0.2395865 0.5010096 0.4395181 0.2055697 0.2888395
#>  [680] 0.3067246 0.3520022 0.3628904 0.3807105 0.7668725 0.2556744 0.2139745
#>  [687] 0.3144216 0.2596447 0.5196315 0.3564033 0.3187612 0.3931125 0.2824913
#>  [694] 0.2105415 0.2634396 0.3777107 0.4646750 0.3077063 0.5754153 0.2684382
#>  [701] 0.3726038 0.2621073 0.4542117 0.2825349 0.4312264 0.2219076 0.1472924
#>  [708] 0.3488597 0.2725298 0.3617334 0.6677637 0.2358770 0.3509483 0.3572861
#>  [715] 0.3429825 0.2988010 0.3758214 0.2379107 0.3183031 0.3727061 0.2524104
#>  [722] 0.2826151 0.2072999 0.3397902 0.2262959 0.2333779 0.2970784 0.2053131
#>  [729] 0.3910786 0.3377695 0.3346534 0.2632837 0.3724075 0.3688745 0.2880989
#>  [736] 0.3132917 0.2677302 0.5062372 0.2395688 0.2923242 0.2296332 0.2108746
#>  [743] 0.3835471 0.4600865 0.1841845 0.4156905 0.3608048 0.2201637 0.3425644
#>  [750] 0.2727573 0.2914810 0.2518544 0.3311985 0.2920304 0.3282528 0.3775384
#>  [757] 0.3771231 0.1601500 0.5370210 0.2656158 0.2951833 0.3081538 0.2245764
#>  [764] 0.2294595 0.2645786 0.4107457 0.1649472 0.1880414 0.2639588 0.2623150
#>  [771] 0.6444112 0.2307711 0.5253228 0.3877254 0.6943472 0.2193599 0.3954462
#>  [778] 0.4815493 0.1674079 0.3314689 0.2866523 0.3296404 0.2534114 0.1926497
#>  [785] 0.3655709 0.3141113 0.2699833 0.2024483 0.3817244 0.2431647 0.2260225
#>  [792] 0.4149990 0.2387331 0.2431047 0.2937638 0.3183596 0.3653719 0.2040596
#>  [799] 0.3619658 0.4749650 0.4878429 0.2719721 0.2518566 0.3755439 0.2445608
#>  [806] 0.3229689 0.2842908 0.2216517 0.2550263 0.3180349 0.4249986 0.2394681
#>  [813] 0.2710541 0.3293096 0.2864617 0.3505302 0.3427341 0.3552116 0.3433531
#>  [820] 0.4065109 0.3331749 0.1964553 0.2046373 0.3425052 0.2523308 0.2829965
#>  [827] 0.2294503 0.2692527 0.2851309 0.2895104 0.2506138 0.2776463 0.3764541
#>  [834] 0.3735188 0.3362292 0.3338995 0.2912642 0.3310920 0.2655369 0.2507705
#>  [841] 0.4807183 0.2225726 0.2901830 0.1967014 0.4365575 0.4589243 0.1923579
#>  [848] 0.2872364 0.2800233 0.3006226 0.3179105 0.1971840 0.3662644 0.3022824
#>  [855] 0.2639980 0.2214784 0.3164268 0.2089243 0.2358112 0.3349552 0.1831239
#>  [862] 0.5101932 0.3218082 0.1832517 0.1335640 0.4896434 0.2024383 0.3687288
#>  [869] 0.3740071 0.1906120 0.3137393 0.5152410 0.4122144 0.3827821 0.5052467
#>  [876] 0.4708607 0.3430801 0.1965741 0.3964855 0.4018256 0.2475760 0.1825187
#>  [883] 0.2212502 0.4968976 0.3617824 0.3830987 0.3291365 0.2860503 0.4731143
#>  [890] 0.4393334 0.2956871 0.4386303 0.3642207 0.2522037 0.2503928 0.6377092
#>  [897] 0.2848181 0.3484466 0.2371647 0.2029375 0.2457139 0.2605492 0.3766246
#>  [904] 0.2133336 0.2990131 0.4002816 0.2820143 0.2655143 0.2382775 0.3806503
#>  [911] 0.2397778 0.3039159 0.2313265 0.2938228 0.4367828 0.3342474 0.3081969
#>  [918] 0.3989755 0.2715901 0.3136425 0.2315973 0.3272022 0.3188122 0.1841695
#>  [925] 0.2888662 0.4103102 0.3425314 0.3470770 0.3164112 0.2673204 0.4297892
#>  [932] 0.2582046 0.3693687 0.2703292 0.2216796 0.2112150 0.3148076 0.2843213
#>  [939] 0.2635657 0.2742230 0.3093001 0.3151468 0.2101611 0.2528829 0.2028352
#>  [946] 0.3029331 0.5492119 0.4374923 0.2757829 0.3359072 0.3258682 0.2156111
#>  [953] 0.2957196 0.2229032 0.3047203 0.2603364 0.3599467 0.2317111 0.3556559
#>  [960] 0.3543343 0.2207337 0.2068746 0.3056550 0.4682181 0.3016401 0.2080445
#>  [967] 0.2830170 0.4809891 0.4496243 0.1950359 0.1539395 0.2722497 0.3369291
#>  [974] 0.3487530 0.3981522 0.4671652 0.5701638 0.3147248 0.4850950 0.3022997
#>  [981] 0.2785153 0.3905227 0.2074414 0.2161605 0.2579213 0.2625315 0.3100725
#>  [988] 0.2504522 0.2931444 0.2551015 0.2728748 0.7284797 0.6824748 0.3364010
#>  [995] 0.3530758 0.4718259 0.2463711 0.5479627 0.4039162 0.1901866 0.3895502
#> [1002] 0.3548389 0.3112224 0.2296930 0.2566060 0.3481145 0.4131906 0.1737938
#> [1009] 0.5024422 0.3611933 0.3243882 0.1851724 0.3838312 0.4831210 0.2647207
#> [1016] 0.2282721 0.2479257 0.3876834 0.4415031 0.2795398 0.2148310 0.3754043
#> [1023] 0.3896995 0.3205376 0.5510850 0.2025063 0.2595707 0.2179225 0.4586191
#> [1030] 0.2739635 0.3283186 0.3554538 0.2567932 0.2465145 0.2546613 0.4802269
#> [1037] 0.2673134 0.3395230 0.4373618 0.2427335 0.5132623 0.2395877 0.4219428
#> [1044] 0.3725741 0.2334042 0.1687137 0.4109431 0.4274316 0.4640325 0.2673489
#> [1051] 0.1321184 0.2417288 0.3673217 0.4002864 0.4629416 0.2612453 0.2011881
#> [1058] 0.4412151 0.2771143 0.2698318 0.2946129 0.3341856 0.3257592 0.3937799
#> [1065] 0.3643093 0.2273057 0.2178895 0.2824689 0.4479125 0.1878365 0.2293263
#> [1072] 0.2983342 0.3340310 0.2268332 0.5113486 0.2641287 0.1781543 0.2273813
#> [1079] 0.2236640 0.2500244 0.3057424 0.3163944 0.4128484 0.3775350 0.2559027
#> [1086] 0.2054981 0.2449295 0.5381957 0.1890424 0.1782422 0.5393488 0.2996896
#> [1093] 0.3394483 0.4293790 0.5148421 0.2463074 0.2438125 0.3230366 0.2042360
#> [1100] 0.3114841 0.3218565 0.3550114 0.2499478 0.2358666 0.3028801 0.5433644
#> [1107] 0.2735789 0.2511152 0.2963880 0.7981887 0.2151066 0.2923062 0.2322761
#> [1114] 0.2065546 0.3475887 0.2382110 0.1711808 0.3290164 0.5151656 0.2851579
#> [1121] 0.3322828 0.3097889 0.1980647 0.2655839 0.3618321 0.6603680 0.2261203
#> [1128] 0.2556369 0.3824720 0.2264052 0.5076516 0.2152310 0.2186271 0.5789007
#> [1135] 0.2240254 0.2370619 0.5261403 0.3937345 0.3385604 0.3180937 0.2940913
#> [1142] 0.4147576 0.3719912 0.3415142 0.3319917 0.3133987 0.7005231 0.3394390
#> [1149] 0.2765780 0.4297331 0.1845750 0.2405610 0.3583838 0.3090038 0.3012045
#> [1156] 0.2722577 0.3474558 0.3066836 0.2205917 0.4491203 0.2318882 0.5591190
#> [1163] 0.4908551 0.2298415 0.3599012 0.1634432 0.2946863 0.3448525 0.3355382
#> [1170] 0.3354535 0.2727264 0.3968510 0.2626067 0.4987276 0.3408297 0.5737319
#> [1177] 0.5291651 0.3463392 0.2349544 0.3579899 0.2520587 0.4481943 0.3489238
#> [1184] 0.3087858 0.4642110 0.1678387 0.2449066 0.2148517 0.4236064 0.3575358
#> [1191] 0.3656842 0.3576133 0.3656219 0.5841919 0.3324927 0.4996753 0.1818449
#> [1198] 0.5569206 0.3315933 0.2402779 0.1899398 0.4352930 0.2672546 0.3923878
#> [1205] 0.3433817 0.3138567 0.2384863 0.5926082 0.2916964 0.1104774 0.2152923
#> [1212] 0.2862356 0.3315506 0.1591519 0.2394468 0.5680645 0.2221866 0.2093054
#> [1219] 0.3576065 0.2829812 0.3466170 0.1921185 0.3443413 0.2142392 0.2125545
#> [1226] 0.3615814 0.3643085 0.2976906 0.2977897 0.1452998 0.2726321 0.3100749
#> [1233] 0.3587263 0.5311910 0.3291628 0.3003663 0.3670987 0.2513652 0.5934677
#> [1240] 0.1870722 0.4015148 0.5088329 0.3849243 0.4617732 0.3464704 0.2710396
#> [1247] 0.4319179 0.3415921 0.2167236 0.3148018 0.4573245 0.2207888 0.2049336
#> [1254] 0.4717294 0.3888645 0.3542153 0.3459308 0.5217741 0.2180884 0.4259269
#> [1261] 0.3201493 0.4420934 0.3041888 0.3700455 0.2805180 0.2188501 0.3460521
#> [1268] 0.4351297 0.1580111 0.3935534 0.5728517 0.2873709 0.2469293 0.3964502
#> [1275] 0.2512581 0.4339819 0.2683077 0.1667046 0.2241136 0.3571773 0.2620149
#> [1282] 0.2890632 0.3086179 0.3292495 0.1765833 0.2624515 0.2640110 0.2964935
#> [1289] 0.3591332 0.2740629 0.2673894 0.3733610 0.4494807 0.2440683 0.2914432
#> [1296] 0.2049841 0.2677745 0.3833574 0.3743701 0.2833338 0.3547053 0.2676420
#> [1303] 0.5038150 0.3424803 0.3709113 0.2394253 0.3827053 0.3073331 0.2522473
#> [1310] 0.5239183 0.2055339 0.2327169 0.2499153 0.1897050 0.2008785 0.3952846
#> [1317] 0.3114149 0.3879658 0.2342201 0.2473706 0.2293508 0.4649044 0.3455254
#> [1324] 0.2609602 0.4870300 0.2467354 0.4661696 0.2021997 0.2624659 0.4340581
#> [1331] 0.2912060 0.2914532 0.2923149 0.2350157 0.3125977 0.3840469 0.2442495
#> [1338] 0.3293657 0.2102493 0.2619184 0.1912385 0.3076946 0.2445438 0.2824811
#> [1345] 0.5800648 0.2604005 0.4719672 0.2124092 0.2553624 0.4403236 0.2824697
#> [1352] 0.4070533 0.3452911 0.2766260 0.6105073 0.2615159 0.3217890 0.4046828
#> [1359] 0.3638092 0.2652429 0.2711709 0.3058326 0.2010089 0.3865913 0.2737985
#> [1366] 0.2463085 0.3620941 0.4242616 0.3096394 0.4804602 0.3745824 0.4088166
#> [1373] 0.2114409 0.2833421 0.1937616 0.1977998 0.4840101 0.5218439 0.2373243
#> [1380] 0.2537919 0.5046667 0.2984904 0.3730334 0.3596095 0.3704431 0.2935712
#> [1387] 0.3057732 0.5354006 0.2104748 0.2686573 0.3146130 0.3779621 0.2087111
#> [1394] 0.2609755 0.4633453 0.2098622 0.1899902 0.1758974 0.4599840 0.3210853
#> [1401] 0.4292473 0.4436852 0.3053497 0.1782777 0.1854092 0.2229950 0.3315579
#> [1408] 0.2733019 0.2415391 0.2381531 0.2982605 0.2177458 0.3688838 0.4204961
#> [1415] 0.1882420 0.4036021 0.1744720 0.2073551 0.3947822 0.1650470 0.3021758
#> [1422] 0.3006048 0.2293788 0.2713717 0.3047907 0.3006964 0.2902372 0.2521826
#> [1429] 0.2132367 0.2761974 0.4271376 0.2228558 0.3459703 0.3363765 0.3353675
#> [1436] 0.3783491 0.1926505 0.5127061 0.2707267 0.2960845 0.2039026 0.3073640
#> [1443] 0.3441596 0.3378504 0.3690943 0.6837033 0.5189992 0.3391679 0.2195739
#> [1450] 0.2749018 0.2622255 0.6402831 0.2001274 0.2781096 0.2033213 0.2477434
#> [1457] 0.3024306 0.1989920 0.1758428 0.4098769 0.2570748 0.3215105 0.3382002
#> [1464] 0.1952209 0.2176968 0.2566089 0.4007920 0.3796001 0.3029278 0.3749769
#> [1471] 0.2655301 0.2762929 0.3112988 0.5122746 0.3741231 0.4654014 0.2273457
#> [1478] 0.3050394 0.2694280 0.1988702 0.2865459 0.2929662 0.3227264 0.2279658
#> [1485] 0.3680578 0.1990295 0.3226736 0.3679572 0.3434205 0.2376873 0.2425705
#> [1492] 0.2323998 0.3590438 0.2488296 0.2827249 0.4841580 0.4074755 0.5174460
#> [1499] 0.5584574 0.4197372 0.2513929 0.3878790 0.2481656 0.4011314 0.3078600
#> [1506] 0.4278477 0.2631427 0.2395162 0.2140985 0.2027003 0.2456875 0.5211151
#> [1513] 0.3164565 0.2731108 0.2428834 0.2872701 0.3618824 0.2014746 0.4497045
#> [1520] 0.3231459 0.4695005 0.3098161 0.3709541 0.2539970 0.2585376 0.2681091
#> [1527] 0.3198075 0.4725366 0.3960544 0.2491795 0.2287414 0.5061646 0.2866436
#> [1534] 0.2454283 0.2103246 0.2342357 0.1817572 0.3113669 0.2532205 0.4101766
#> [1541] 0.5299864 0.5840637 0.2687905 0.4972431 0.3389442 0.2175552 0.2746299
#> [1548] 0.2352760 0.2353005 0.4765076 0.2530409 0.3752603 0.2838215 0.2886788
#> [1555] 0.4498585 0.2392281 0.5066703 0.2174028 0.5873218 0.3439133 0.7815302
#> [1562] 0.2816847 0.1641046 0.3084680 0.3760738 0.3652658 0.3396080 0.2636820
#> [1569] 0.4741950 0.2531981 0.2141637 0.2285800 0.3405982 0.3255265 0.3710033
#> [1576] 0.2607191 0.3918767 0.2780168 0.4377521 0.3380443 0.2229261 0.2932113
#> [1583] 0.2320963 0.2972415 0.2762011 0.3154477 0.3352886 0.2618537 0.1634018
#> [1590] 0.2482070 0.2860992 0.2578971 0.3980176 0.5153112 0.3018068 0.2743282
#> [1597] 0.1821846 0.2553281 0.2188561 0.4684300 0.2883647 0.3767686 0.2917686
#> [1604] 0.3120504 0.2472158 0.6102650 0.4879758 0.2850939 0.2621440 0.3442061
#> [1611] 0.2858638 0.4551507 0.7440670 0.2858149 0.2695547 0.2383990 0.3539349
#> [1618] 0.2508582 0.4879210 0.2198682 0.3483689 0.2405424 0.3210554 0.3711567
#> [1625] 0.2300910 0.2712792 0.4502078 0.4406538 0.2550529 0.4104685 0.3496115
#> [1632] 0.4095202 0.4028976 0.2173278 0.2328842 0.3700423 0.5125387 0.3246571
#> [1639] 0.3503431 0.3449884 0.2823954 0.3256290 0.3326570 0.2271842 0.2172974
#> [1646] 0.3569679 0.3543131 0.3378006 0.2617588 0.2012529 0.4551490 0.3031147
#> [1653] 0.3004402 0.2430289 0.3761436 0.3510515 0.4211888 0.3572726 0.2469519
#> [1660] 0.2712124 0.1970666 0.4870812 0.3897634 0.3139365 0.1919107 0.2777171
#> [1667] 0.2282928 0.2147005 0.2737114 0.2652896 0.2836280 0.2265763 0.3075711
#> [1674] 0.2708310 0.4602777 0.2030366 0.3928786 0.3937453 0.3243513 0.2459951
#> [1681] 0.4000680 0.3054004 0.3007810 0.4244889 0.3111853 0.2264180 0.3701864
#> [1688] 0.3026603 0.4205097 0.2798203 0.2066358 0.6731840 0.2709030 0.3421759
#> [1695] 0.4407452 0.4310589 0.2000350 0.1915818 0.2298483 0.3298992 0.3624580
#> [1702] 0.2922820 0.2272240 0.2928048 0.2789635 0.3214079 0.3002759 0.3109944
#> [1709] 0.3746584 0.2267186 0.2051326 0.2156319 0.2132366 0.1800907 0.2877480
#> [1716] 0.2568655 0.4507470 0.2838201 0.5593194 0.1679704 0.4073077 0.2140493
#> [1723] 0.2407462 0.2724719 0.3429140 0.3440063 0.3027999 0.3570922 0.3081523
#> [1730] 0.2377082 0.3351722 0.3188214 0.3890370 0.1908039 0.2187485 0.4785962
#> [1737] 0.2921662 0.2381289 0.2739057 0.5428815 0.3589890 0.3564114 0.2769966
#> [1744] 0.3611553 0.3522608 0.1750928 0.3493966 0.4163638 0.2214838 0.2332940
#> [1751] 0.3574765 0.3764293 0.3593483 0.2550692 0.2879140 0.3098474 0.2244653
#> [1758] 0.3004108 0.2666990 0.3099868 0.2231692 0.2618214 0.2980253 0.2501112
#> [1765] 0.2324352 0.2045365 0.4762358 0.2931988 0.3089537 0.2972295 0.4730173
#> [1772] 0.3862841 0.2606026 0.4775710 0.3579726 0.2260144 0.5704577 0.3550063
#> [1779] 0.4116131 0.3522126 0.3138342 0.3188560 0.2598966 0.6160988 0.2620873
#> [1786] 0.2372883 0.2693253 0.3203084 0.3434774 0.4263362 0.3082671 0.5319264
#> [1793] 0.2744446 0.6248705 0.3087913 0.3260202 0.3730787 0.5982974 0.3175224
#> [1800] 0.6032451 0.3347978 0.3404623 0.2348994 0.4693262 0.3029762 0.2947577
#> [1807] 0.4799348 0.3812944 0.4506740 0.4129626 0.2153125 0.2400186 0.3095980
#> [1814] 0.2559652 0.4652931 0.2578630 0.3842485 0.3238862 0.2045375 0.2744718
#> [1821] 0.2250413 0.3042543 0.4563073 0.2640189 0.3130648 0.3660629 0.3257045
#> [1828] 0.2924368 0.2363796 0.2213260 0.3634638 0.4879451 0.1779338 0.4655030
#> [1835] 0.3770433 0.3812724 0.2891360 0.3136019 0.2277803 0.2224319 0.2217312
#> [1842] 0.3939059 0.6175169 0.5821024 0.2143888 0.2519451 0.5430767 0.2379721
#> [1849] 0.4231837 0.1771852 0.3339954 0.5775539 0.5470299 0.3091276 0.5877655
#> [1856] 0.3723566 0.2458239 0.2784044 0.2782215 0.3291063 0.3199005 0.3267038
#> [1863] 0.1634349 0.3292717 0.4268380 0.2488489 0.2648856 0.3845173 0.2219199
#> [1870] 0.2843098 0.3162625 0.1808196 0.3482419 0.4222136 0.1972990 0.3038803
#> [1877] 0.2769139 0.6353390 0.2007467 0.3263850 0.3767745 0.1607602 0.4142862
#> [1884] 0.2506907 0.3208672 0.2500266 0.2419813 0.3842417 0.2557726 0.3445937
#> [1891] 0.4165337 0.3248085 0.4317331 0.2629425 0.2635671 0.2380206 0.2655803
#> [1898] 0.5043997 0.2170147 0.3122776 0.2597672 0.2871876 0.2174007 0.3690314
#> [1905] 0.2446045 0.3819187 0.2632964 0.2282166 0.5290234 0.3400092 0.5455749
#> [1912] 0.4199456 0.4270915 0.3286729 0.5267827 0.3286298 0.3939052 0.4380936
#> [1919] 0.2481726 0.2557119 0.3157925 0.3063575 0.2937485 0.2023991 0.3648335
#> [1926] 0.4909274 0.3300541 0.4499648 0.3103681 0.3589145 0.2319627 0.3414812
#> [1933] 0.5038788 0.3487161 0.3026154 0.3409745 0.2392346 0.4860295 0.3014324
#> [1940] 0.2966093 0.4365319 0.2395169 0.3305571 0.2665921 0.3488528 0.4219798
#> [1947] 0.2318101 0.2980070 0.2244814 0.5914325 0.2547037 0.3532720 0.6140690
#> [1954] 0.3372859 0.2067984 0.5227720 0.2856206 0.4335327 0.3553006 0.1641135
#> [1961] 0.2622418 0.2276450 0.2339628 0.4567757 0.3556198 0.2287688 0.3292841
#> [1968] 0.3309543 0.3623291 0.4957533 0.6055064 0.3492281 0.2177296 0.2097007
#> [1975] 0.2758198 0.2416209 0.3142700 0.7933285 0.2644364 0.2774267 0.3961932
#> [1982] 0.3929286 0.3363932 0.2215337 0.5353431 0.2734191 0.2367519 0.2251507
#> [1989] 0.3823404 0.4999189 0.2636780 0.2669565 0.5617020 0.2471966 0.1782152
#> [1996] 0.5287214 0.4373181 0.2527571 0.2135219 0.3502196 0.3217769 0.3398903
#> [2003] 0.3546610 0.5836691 0.2958751 0.3876910 0.3505502 0.2317509 0.3280523
#> [2010] 0.2414477 0.3475156 0.1772087 0.3040374 0.2709506 0.2541105 0.2607515
#> [2017] 0.3587758 0.4897961 0.2790282 0.2872644 0.2218778 0.3001245 0.2496756
#> [2024] 0.2906182 0.1808384 0.4317618 0.2119375 0.4294364 0.2487654 0.2789390
#> [2031] 0.3143112 0.3004021 0.4302000 0.3253399 0.5186404 0.3700392 0.2504095
#> [2038] 0.2287889 0.3030375 0.2803576 0.4687216 0.2047147 0.2202613 0.3631799
#> [2045] 0.2602477 0.2705806 0.2008613 0.2116541 0.2227914 0.3077995 0.2200995
#> [2052] 0.3092053 0.2261963 0.3061989 0.4474347 0.1975970 0.2567549 0.2884990
#> [2059] 0.3739894 0.2288463 0.5724529 0.2862474 0.2707974 0.2305781 0.4206271
#> [2066] 0.2634575 0.1638783 0.2363289 0.2889587 0.3002289 0.2417706 0.3858058
#> [2073] 0.4907716 0.1646102 0.2772726 0.3642554 0.2870591 0.3204028 0.2025379
#> [2080] 0.4591848 0.3239368 0.4212389 0.3858196 0.4897962 0.2653024 0.3767138
#> [2087] 0.1669533 0.2606092 0.2367066 0.2297026 0.3778148 0.3238724 0.2227403
#> [2094] 0.3776970 0.6615489 0.2539261 0.2511138 0.2143515 0.2477329 0.3950309
#> [2101] 0.5449026 0.2498471 0.4286904 0.4037613 0.2325646 0.2340072 0.1953494
#> [2108] 0.2842328 0.4304549 0.4415862 0.2823356 0.5289478 0.6313503 0.4695004
#> [2115] 0.2640548 0.3324614 0.1692881 0.2548364 0.2338703 0.4707194 0.2735853
#> [2122] 0.5948256 0.2597053 0.2090647 0.3242956 0.2675646 0.2605636 0.3739120
#> [2129] 0.4166638 0.4596909 0.2321939 0.3647776 0.3786094 0.3774909 0.2406404
#> [2136] 0.4243978 0.2308152 0.1908304 0.1904918 0.2684304 0.3429116 0.3263338
#> [2143] 0.3936195 0.2971451 0.5345083 0.4663920 0.4044163 0.4768147 0.2936836
#> [2150] 0.2541102 0.2049511 0.2332706 0.4193514 0.3048227 0.2787819 0.4060052
#> [2157] 0.2026853 0.2207140 0.3506147 0.2011680 0.2214760 0.6156407 0.2592316
#> [2164] 0.2707131 0.4956261 0.4812241 0.3551794 0.1918033 0.5559466 0.3810354
#> [2171] 0.3652526 0.2919145 0.2987750 0.2314831 0.2503153 0.3712622 0.4430383
#> [2178] 0.5952294 0.6158497 0.2994620 0.3108740 0.3752342 0.3123826 0.2099157
#> [2185] 0.2922656 0.4435031 0.3429004 0.3674647 0.4008119 0.2458852 0.3806910
#> [2192] 0.5635968 0.2190118 0.3185797 0.2808278 0.4614769 0.2767132 0.2352880
#> [2199] 0.2794241 0.2626771 0.2163359 0.2237616 0.2905790 0.3264086 0.2239298
#> [2206] 0.2426666 0.2090300 0.6749552 0.4612932 0.2891894 0.3299861 0.3838102
#> [2213] 0.2228736 0.4951938 0.2340152 0.2386900 0.4795381 0.3611388 0.1655996
#> [2220] 0.2766868 0.2614053 0.6897366 0.2415136 0.1824792 0.3840207 0.3454313
#> [2227] 0.3164250 0.2079593 0.3629415 0.4884767 0.5008875 0.5014420 0.7250756
#> [2234] 0.3719303 0.2305648 0.3715723 0.2486060 0.4715163 0.3385375 0.2962191
#> [2241] 0.3388160 0.2562495 0.4225379 0.3415698 0.3420026 0.2303461 0.1815540
#> [2248] 0.3846201 0.2870506 0.3479203 0.6356232 0.3589903 0.2579441 0.2989402
#> [2255] 0.3052698 0.2566997 0.1908140 0.4310486 0.4234968 0.1817033 0.6609156
#> [2262] 0.1819792 0.3339959 0.4589577 0.3934273 0.3028599 0.1981275 0.4444892
#> [2269] 0.2515481 0.4072415 0.2468634 0.2462757 0.1864991 0.3893058 0.3724642
#> [2276] 0.3319574 0.2728422 0.5180789 0.6502577 0.1736614 0.3051100 0.3534178
#> [2283] 0.3097001 0.2278612 0.5148043 0.4374500 0.2842792 0.3016485 0.1648009
#> [2290] 0.1891716 0.2306525 0.2613564 0.2317891 0.1642571 0.2120948 0.4379285
#> [2297] 0.2031649 0.2579093 0.2434324 0.4593736 0.2738044 0.3327342 0.2368744
#> [2304] 0.2850950 0.3141222 0.2074880 0.2800528 0.3266582 0.2434737 0.2847612
#> [2311] 0.4965827 0.2798810 0.2830007 0.3210639 0.4688413 0.2784926 0.1865773
#> [2318] 0.2170013 0.2944518 0.3411597 0.2475276 0.1950842 0.2690378 0.2426038
#> [2325] 0.4568118 0.6549017 0.2375922 0.5252015 0.2719656 0.5300795 0.3606523
#> [2332] 0.2808416 0.3773306 0.4925232 0.2517941 0.3778228 0.1742047 0.2947538
#> [2339] 0.4838265 0.2681992 0.2718658 0.4043723 0.2463888 0.3019340 0.3008904
#> [2346] 0.3488911 0.2091915 0.3960770 0.3122001 0.3407247 0.2206429 0.2881743
#> [2353] 0.2224965 0.1848433 0.2262704 0.2196424 0.1905480 0.2814345 0.2328166
#> [2360] 0.2477725 0.3903993 0.2021842 0.2835090 0.3803103 0.2610026 0.3362533
#> [2367] 0.2700813 0.2587835 0.2720616 0.3169974 0.2320855 0.2695803 0.3077399
#> [2374] 0.2060072 0.3555899 0.2186968 0.2682711 0.2670921 0.2829367 0.2361628
#> [2381] 0.2387151 0.2452422 0.1768174 0.3815148 0.2419003 0.3011733 0.2788402
#> [2388] 0.4731458 0.3573215 0.3702481 0.2222615 0.2371904 0.2078009 0.5210372
#> [2395] 0.3136107 0.2613093 0.4286956 0.3325475 0.3768834 0.2181198 0.2298472
#> [2402] 0.3819776 0.3425140 0.4545313 0.2446113 0.3568080 0.1948706 0.3367402
#> [2409] 0.2423918 0.3753274 0.2390362 0.3109519 0.1837402 0.4944174 0.3656793
#> [2416] 0.2607765 0.2942916 0.2096124 0.2133837 0.2533344 0.4594600 0.3710966
#> [2423] 0.4335837 0.2784241 0.2227401 0.2508059 0.1931564 0.2761946 0.2584252
#> [2430] 0.1783060 0.3080853 0.3574848 0.2337341 0.2357566 0.2549946 0.2393865
#> [2437] 0.4878289 0.3685450 0.2722284 0.2980684 0.2442929 0.3059006 0.4418003
#> [2444] 0.2574205 0.3715919 0.3006200 0.3836127 0.1899613 0.2744423 0.2123748
#> [2451] 0.5457034 0.3951206 0.4470631 0.3217447 0.3307279 0.3510644 0.2267870
#> [2458] 0.2171325 0.2571577 0.3853892 0.2434053 0.4565264 0.4256877 0.1915349
#> [2465] 0.2634195 0.3859627 0.2634326 0.2029136 0.3457438 0.1701192 0.3957202
#> [2472] 0.2727054 0.3013274 0.4389657 0.2218273 0.3876918 0.4525757 0.3377606
#> [2479] 0.3330484 0.2782973 0.2157588 0.3703891 0.4470175 0.2074436 0.2046787
#> [2486] 0.4907888 0.3097261 0.5212512 0.4138404 0.1563285 0.5842553 0.3449814
#> [2493] 0.3429384 0.6155772 0.2574165 0.4339382 0.3416804 0.3540819 0.2693101
#> [2500] 0.2436591 0.3071342 0.3320109 0.4238310 0.3072178 0.4075072 0.2914693
#> [2507] 0.4501878 0.4287953 0.2640914 0.2306169 0.2736826 0.4016058 0.2589101
#> [2514] 0.2641891 0.2202408 0.3289214 0.4902704 0.4297731 0.3036983 0.3882083
#> [2521] 0.4069889 0.3851344 0.3324488 0.4140419 0.3668179 0.2355691 0.3791678
#> [2528] 0.2136879 0.1866565 0.2816393 0.3687195 0.3245160 0.3221043 0.1771506
#> [2535] 0.4520438 0.2968061 0.4955634 0.4824196 0.4107505 0.4798303 0.2031501
#> [2542] 0.3377678 0.2150174 0.2796469 0.3074474 0.2946553 0.1960718 0.2518030
#> [2549] 0.2958218 0.1556397 0.2293742 0.1649812 0.2874709 0.2734536 0.1634846
#> [2556] 0.2896043 0.2482524 0.2071839 0.1962017 0.2149400 0.2216502 0.1756552
#> [2563] 0.2172950 0.4179776 0.6223875 0.2652011 0.3144133 0.2552306 0.1664520
#> [2570] 0.3085469 0.2588227 0.2467946 0.3651840 0.3651700 0.3498065 0.4054232
#> [2577] 0.2657672 0.3283813 0.2264680 0.4445936 0.3171458 0.1934680 0.2383467
#> [2584] 0.2860981 0.2894172 0.2699680 0.2583006 0.2705175 0.2530287 0.3972126
#> [2591] 0.5435520 0.4855204 0.3587494 0.3996958 0.2260879 0.2197769 0.2594434
#> [2598] 0.3020793 0.2924372 0.2043287 0.1994431 0.2770318 0.2024668 0.3338906
#> [2605] 0.3097303 0.3911600 0.3449698 0.4063694 0.3970697 0.4553196 0.4025428
#> [2612] 0.2288131 0.4577930 0.2276148 0.2853652 0.3187794 0.2573249 0.2083206
#> [2619] 0.2786684 0.2708246 0.4127365 0.2924412 0.3205822 0.2634448 0.2716850
#> [2626] 0.4218719 0.5081660 0.3825249 0.4864975 0.5079506 0.2326521 0.4105233
#> [2633] 0.4051909 0.3082929 0.3807984 0.2732932 0.2208591 0.2439094 0.2492240
#> [2640] 0.2250241 0.4098877 0.4332246 0.5293048 0.3804846 0.1789224 0.2246633
#> [2647] 0.4003542 0.2613256 0.3706863 0.2245842 0.3856102 0.2604077 0.2827167
#> [2654] 0.1489000 0.2607130 0.3332451 0.3723369 0.2972373 0.3810342 0.3407876
#> [2661] 0.2575977 0.2687323 0.3851139 0.4392650 0.3337158 0.3330453 0.1852304
#> [2668] 0.3169069 0.2807929 0.3361443 0.2763600 0.1611729 0.4571663 0.3076586
#> [2675] 0.2831137 0.2564171 0.3199614 0.2666499 0.3417537 0.3913850 0.3737178
#> [2682] 0.1683471 0.4516687 0.2535092 0.6866799 0.3541073 0.2001731 0.3833209
#> [2689] 0.3091150 0.3491310 0.3068206 0.2890370 0.3933304 0.3996248 0.3007119
#> [2696] 0.2405941 0.4534817 0.3321364 0.3225142 0.3218004 0.4789889 0.2972454
#> [2703] 0.3807287 0.3511555 0.3092616 0.2638896 0.2060822 0.2942041 0.5959486
#> [2710] 0.1901116 0.4253372 0.3445824 0.5651279 0.2011201 0.4629651 0.4670010
#> [2717] 0.2392097 0.3317102 0.2522540 0.5129387 0.2966541 0.3161887 0.4685547
#> [2724] 0.1956456 0.2530029 0.3482359 0.3287590 0.2063144 0.3143865 0.3938259
#> [2731] 0.2573001 0.2390762 0.2126780 0.2726195 0.6235667 0.3110725 0.4283196
#> [2738] 0.1969347 0.3207724 0.1876883 0.2607335 0.3473004 0.2798459 0.2820088
#> [2745] 0.2819709 0.4361056 0.2456683 0.3854378 0.2692862 0.2967195 0.2284147
#> [2752] 0.3222426 0.2419183 0.2164805 0.4255749 0.2850737 0.3416259 0.3349958
#> [2759] 0.3579540 0.2268482 0.5329533 0.2189345 0.3589389 0.5078343 0.2943232
#> [2766] 0.3543030 0.2797212 0.3720295 0.4437590 0.5385155 0.3227683 0.2409621
#> [2773] 0.2264046 0.1837283 0.3021165 0.3157684 0.2606756 0.2314002 0.4204999
#> [2780] 0.2134687 0.2909298 0.2401235 0.2865543 0.3028288 0.2804870 0.4337663
#> [2787] 0.2500238 0.2623359 0.2389264 0.4015181 0.4839021 0.3214475 0.2177266
#> [2794] 0.3799992 0.2761003 0.2943310 0.2345989 0.2893670 0.2917129 0.4290270
#> [2801] 0.2396292 0.5257323 0.2793267 0.3471880 0.3028253 0.2241107 0.2371410
#> [2808] 0.4635103 0.5529435 0.2754914 0.2694122 0.3861950 0.4091916 0.2305270
#> [2815] 0.2124526 0.2696502 0.1914533 0.1725336 0.3241524 0.2083380 0.2644011
#> [2822] 0.2148696 0.2082301 0.3717697 0.3053725 0.2709253 0.4434493 0.3630627
#> [2829] 0.5705134 0.2430019 0.4186210 0.3618622 0.1885020 0.2261115 0.2438270
#> [2836] 0.3037449 0.1761589 0.4872171 0.3511612 0.2447982 0.4234942 0.4564205
#> [2843] 0.4009442 0.2268967 0.4585231 0.2866600 0.2676587 0.3342488 0.4029779
#> [2850] 0.2972499 0.3365201 0.3635031 0.2939214 0.2907915 0.3267733 0.2928737
#> [2857] 0.5305244 0.3606378 0.3256951 0.4189644 0.2206086 0.4851386 0.3103186
#> [2864] 0.3131774 0.1952879 0.1745631 0.3183167 0.3242951 0.2768910 0.4416644
#> [2871] 0.1947523 0.2446856 0.1826249 0.3830670 0.2751553 0.3217398 0.2585018
#> [2878] 0.2263092 0.4087442 0.2509253 0.2322125 0.3415222 0.2493554 0.2037249
#> [2885] 0.3460260 0.2253361 0.3767199 0.2622362 0.3077706 0.3228150 0.3964326
#> [2892] 0.1550812 0.2915551 0.2737022 0.3345199 0.2536424 0.3050764 0.1973037
#> [2899] 0.2909883 0.2393946 0.3365707 0.2649501 0.2793423 0.4346710 0.4572141
#> [2906] 0.4173863 0.3367047 0.2512409 0.3127603 0.2610380 0.3509091 0.2982885
#> [2913] 0.2455368 0.3000940 0.3208792 0.3710724 0.4145141 0.3365866 0.3140724
#> [2920] 0.1655844 0.1852751 0.1562317 0.4107807 0.5071735 0.4068781 0.3541979
#> [2927] 0.4753499 0.3689284 0.5685056 0.2848320 0.5612377 0.2329812 0.7958530
#> [2934] 0.2722693 0.2480230 0.4778503 0.1707955 0.3256535 0.2729344 0.2140062
#> [2941] 0.3922562 0.2096236 0.1949170 0.3816053 0.5849685 0.1965734 0.3386293
#> [2948] 0.2517282 0.2894552 0.2890571 0.2654482 0.1593286 0.5325428 0.4241742
#> [2955] 0.3497037 0.2147281 0.3510077 0.4431410 0.2490963 0.2204986 0.3577891
#> [2962] 0.2196651 0.2135868 0.2099562 0.2581916 0.3590020 0.2656495 0.4341085
#> [2969] 0.2043262 0.2126300 0.1913330 0.2888396 0.3397703 0.4148433 0.2589559
#> [2976] 0.4041617 0.2005756 0.2697029 0.3714347 0.3428619 0.5370354 0.2516212
#> [2983] 0.3607161 0.2600817 0.2599704 0.2463804 0.4011690 0.3348924 0.2825432
#> [2990] 0.2528929 0.1977642 0.5575950 0.2267719 0.2266881 0.4805654 0.2261466
#> [2997] 0.4217994 0.1441267 0.4629884 0.2359803 0.4067675 0.3918758 0.3097829
#> [3004] 0.1810435 0.2765490 0.4483623 0.2116008 0.4023388 0.2684075 0.2753736
#> [3011] 0.4070295 0.3201589 0.2542200 0.2443781 0.3305204 0.1931192 0.2243143
#> [3018] 0.4380352 0.2801784 0.3589488 0.1809156 0.3026784 0.2080861 0.5054327
#> [3025] 0.4690330 0.6227447 0.1907726 0.3756260 0.3451103 0.2404894 0.2186619
#> [3032] 0.3867416 0.4222063 0.2187058 0.3713753 0.3660840 0.3046185 0.5933895
#> [3039] 0.3067168 0.4963798 0.6992669 0.3117014 0.4766606 0.2055223 0.4620919
#> [3046] 0.2911432 0.2079612 0.4569850 0.3236385 0.1620727 0.2850292 0.3593712
#> [3053] 0.3925220 0.3738031 0.3328846 0.3831762 0.2669793 0.2795279 0.3490040
#> [3060] 0.2956190 0.3017764 0.1906483 0.3197845 0.3589961 0.3159960 0.4288370
#> [3067] 0.4371363 0.3560856 0.5414062 0.3957763 0.2367413 0.5780394 0.2633463
#> [3074] 0.3110543 0.3775952 0.1549590 0.3955893 0.3642916 0.2162201 0.2977686
#> [3081] 0.2772502 0.2448215 0.3372376 0.3343155 0.2461332 0.1811220 0.1769485
#> [3088] 0.5565876 0.2369688 0.4101293 0.2977378 0.3484446 0.1801283 0.4741425
#> [3095] 0.3886186 0.2256809 0.1927454 0.3152238 0.4070092 0.3375975 0.3408904
#> [3102] 0.3623135 0.2484053 0.1476922 0.2526556 0.3676142 0.3134795 0.2947997
#> [3109] 0.3560953 0.2919333 0.5032097 0.4731120 0.2393655 0.3167215 0.3832891
#> [3116] 0.2345245 0.3878196 0.2817841 0.2622762 0.5622974 0.1663150 0.2060774
#> [3123] 0.2051466 0.2524579 0.2249483 0.5318554 0.3889608 0.2852554 0.3187970
#> [3130] 0.3178016 0.2307216 0.2436579 0.2165305 0.3715987 0.2592628 0.2640770
#> [3137] 0.3522286 0.3284165 0.2696408 0.3369520 0.6328359 0.2504612 0.3272990
#> [3144] 0.2546823 0.2929984 0.3344895 0.3415197 0.2835964 0.2037173 0.2475492
#> [3151] 0.3069376 0.2315845 0.2310816 0.2543671 0.1887700 0.5101961 0.4074084
#> [3158] 0.2488593 0.7135820 0.4847869 0.4310926 0.4705191 0.2647749 0.3660924
#> [3165] 0.2690077 0.3473085 0.5070209 0.2880285 0.4262768 0.2220712 0.2606828
#> [3172] 0.3033074 0.4311620 0.2660034 0.2005204 0.2331756 0.4015947 0.3475318
#> [3179] 0.3886279 0.1603419 0.3308186 0.6399779 0.4895434 0.2381298 0.3707560
#> [3186] 0.2784171 0.7223235 0.3906410 0.2948605 0.3109894 0.3328547 0.6283451
#> [3193] 0.3887480 0.3270290 0.2861079 0.2807902 0.2280678 0.2501204 0.2564980
#> [3200] 0.3072437 0.2706490 0.2501492 0.3901876 0.4006160 0.2507442 0.3454984
#> [3207] 0.3829813 0.3551766 0.5825122 0.2104725 0.5970451 0.2967437 0.4255888
#> [3214] 0.2416196 0.3524738 0.2959024 0.2335882 0.2969770 0.2902305 0.2077232
#> [3221] 0.2721395 0.2953447 0.4219350 0.4423988 0.3636118 0.3197760 0.3354311
#> [3228] 0.3656403 0.3231673 0.3176237 0.2951931 0.3422028 0.3495166 0.2261956
#> [3235] 0.2724869 0.2988926 0.4826329 0.2530803 0.5007518 0.2504681 0.2729980
#> [3242] 0.3891352 0.6348657 0.3741989 0.2723099 0.2631423 0.1708245 0.4829360
#> [3249] 0.4190209 0.2623321 0.2032789 0.3658409 0.2814271 0.2307469 0.2128314
#> [3256] 0.2953061 0.3319742 0.3559372 0.2301079 0.2800751 0.2941573 0.5156704
#> [3263] 0.1698919 0.4759764 0.2589043 0.4757074 0.1914135 0.2725280 0.2698559
#> [3270] 0.2762297 0.3934244 0.1972973 0.3169668 0.5531765 0.2411978 0.3930406
#> [3277] 0.2717051 0.3354781 0.2913599 0.3419667 0.3057882 0.3405413 0.2781018
#> [3284] 0.2518337 0.4343693 0.5321245 0.1964160 0.2988601 0.2831163 0.2760821
#> [3291] 0.2696358 0.3892457 0.2724411 0.2133545 0.5308460 0.6148797 0.3431819
#> [3298] 0.3541961 0.4057568 0.2864683 0.3016147 0.2351557 0.3612809 0.2817543
#> [3305] 0.2230177 0.2874638 0.2928724 0.2780837 0.2710795 0.3367344 0.4639559
#> [3312] 0.6121633 0.4558466 0.3446094 0.3371563 0.2597563 0.2189797 0.2240098
#> [3319] 0.3396792 0.2738380 0.2909605 0.2879688 0.3029177 0.3693661 0.3025594
#> [3326] 0.2926893 0.2870809 0.2837273 0.2996648 0.2057611 0.4023039 0.2205124
#> [3333] 0.2428753