Euclidean Depth

Computes the euclidean depth of a point or vectors of points with respect to a multivariate data set.

depthEuclid(u, X)

Arguments

u
Numerical vector or matrix whose depth is to be calculated. Dimension has to be the same as that of the observations.

X
The data as a matrix, data frame or list. If it is a matrix or data frame, then each row is viewed as one multivariate observation. If it is a list, all components must be numerical vectors of equal length (coordinates of observations).

Details

Calculation of Euclidean depth is exact.

Returns the depth of multivariate point u with respect to data set X.

Examples

x <- matrix(rnorm(9999), nc = 3)
depthEuclid(x, x)

#> Depth method:  Euclidean 
#>    [1] 0.26289741 0.16500824 0.30349370 0.05182999 0.47951420 0.32438951
#>    [7] 0.28226141 0.39290098 0.12320303 0.15238689 0.26163806 0.11998650
#>   [13] 0.38649278 0.44946078 0.24176958 0.32548274 0.65780707 0.70407980
#>   [19] 0.13498709 0.43714247 0.28311128 0.23909916 0.62538460 0.35152846
#>   [25] 0.98359149 0.22744376 0.20591076 0.16021059 0.28919714 0.26211055
#>   [31] 0.38521224 0.50203108 0.16550721 0.37573896 0.60621788 0.16507504
#>   [37] 0.16318689 0.48103402 0.64351756 0.41451917 0.33637739 0.52516648
#>   [43] 0.09691137 0.41284133 0.12428002 0.36292114 0.67609381 0.28438470
#>   [49] 0.33779059 0.65789283 0.22872364 0.22031138 0.16322752 0.15981612
#>   [55] 0.18790230 0.14937805 0.15124421 0.13819751 0.32140275 0.19996813
#>   [61] 0.08393143 0.32881701 0.27937117 0.61709792 0.14971459 0.14814284
#>   [67] 0.21545528 0.23881535 0.57822767 0.12638994 0.10820011 0.58761168
#>   [73] 0.46098079 0.67411131 0.18431781 0.29534254 0.35459131 0.25443885
#>   [79] 0.12188711 0.33880958 0.15566513 0.14819278 0.25514449 0.24437454
#>   [85] 0.36128980 0.17117001 0.12902515 0.22853211 0.28150043 0.39955752
#>   [91] 0.65999548 0.40805933 0.26417615 0.65283879 0.91411529 0.20906293
#>   [97] 0.83826446 0.13649959 0.14965008 0.16393067 0.19337348 0.37170498
#>  [103] 0.36436617 0.78992644 0.16003478 0.28866686 0.20057349 0.17982396
#>  [109] 0.14859398 0.68561476 0.27998132 0.17668590 0.18849732 0.23783124
#>  [115] 0.75196816 0.07895674 0.23470218 0.59629386 0.37965357 0.19779302
#>  [121] 0.51372802 0.23047040 0.36475144 0.58502194 0.13655955 0.16240902
#>  [127] 0.47351585 0.59492344 0.51991628 0.38840251 0.19062565 0.43654122
#>  [133] 0.25917218 0.34913902 0.27257398 0.45346680 0.64379242 0.17428218
#>  [139] 0.37581647 0.39403883 0.16056053 0.55074612 0.19945376 0.28749102
#>  [145] 0.72503608 0.13932191 0.49555855 0.14026217 0.56417151 0.47152255
#>  [151] 0.23575394 0.44689805 0.32235827 0.29275759 0.37187467 0.30980884
#>  [157] 0.21406178 0.23173487 0.11471936 0.38374575 0.29746077 0.76350354
#>  [163] 0.53171252 0.17051029 0.32522737 0.23237677 0.32967315 0.21029495
#>  [169] 0.19185958 0.24891128 0.25738559 0.13502491 0.25561350 0.37250552
#>  [175] 0.29514081 0.57103354 0.26341158 0.07179232 0.23616005 0.24028048
#>  [181] 0.54158557 0.19782539 0.22178330 0.64487733 0.34422874 0.71743093
#>  [187] 0.31941067 0.37095587 0.45602259 0.24149930 0.33583522 0.10818888
#>  [193] 0.55969610 0.36839911 0.36169300 0.22396220 0.22195203 0.76993814
#>  [199] 0.61869156 0.77259254 0.57491676 0.27305812 0.23329631 0.22163731
#>  [205] 0.36549824 0.73778230 0.27339317 0.32171478 0.27708077 0.83140791
#>  [211] 0.28949616 0.08870509 0.16284976 0.62499983 0.24163988 0.23669148
#>  [217] 0.44831375 0.42776791 0.23733550 0.15875575 0.12162826 0.73121597
#>  [223] 0.56849951 0.41038800 0.65206862 0.14356891 0.31393115 0.29902304
#>  [229] 0.55761152 0.61128424 0.17364224 0.16594859 0.48207229 0.38525297
#>  [235] 0.64655694 0.81964884 0.14470045 0.12483990 0.71791766 0.42085658
#>  [241] 0.92078673 0.21466619 0.33056760 0.28808459 0.46554159 0.23051068
#>  [247] 0.12561972 0.14474485 0.67215891 0.38130095 0.40087305 0.51066795
#>  [253] 0.25565024 0.27517143 0.38626289 0.89920837 0.36384858 0.54249497
#>  [259] 0.48365887 0.32555855 0.34184131 0.25848430 0.27089381 0.89463745
#>  [265] 0.24494539 0.15337184 0.08427270 0.11458270 0.24493656 0.16678063
#>  [271] 0.64623043 0.51277157 0.68973705 0.29267220 0.19653494 0.45181637
#>  [277] 0.46941944 0.65043666 0.59376119 0.69582023 0.65045728 0.82637849
#>  [283] 0.34742655 0.70135485 0.18528904 0.63937846 0.26569774 0.77460927
#>  [289] 0.20457613 0.45241925 0.36307703 0.21239978 0.53761141 0.18492877
#>  [295] 0.47714299 0.26257949 0.14774893 0.20537719 0.20733184 0.22042022
#>  [301] 0.16895731 0.19947283 0.57760989 0.37079782 0.17942883 0.15057182
#>  [307] 0.21598797 0.23765834 0.71616014 0.94914371 0.13944723 0.52987222
#>  [313] 0.14684774 0.33675331 0.39828340 0.12218030 0.30106008 0.30295347
#>  [319] 0.28151435 0.07282652 0.75387846 0.21640180 0.60666882 0.39895979
#>  [325] 0.19434478 0.53833155 0.20317072 0.70127871 0.08136259 0.47308157
#>  [331] 0.23826803 0.40977290 0.19846083 0.70211104 0.39003418 0.21842782
#>  [337] 0.21784963 0.42487647 0.80960071 0.30286063 0.47591686 0.14491510
#>  [343] 0.26645356 0.27941642 0.22163144 0.32708114 0.11457229 0.23709134
#>  [349] 0.13570336 0.43381715 0.79698953 0.31592660 0.29502813 0.37105324
#>  [355] 0.21475653 0.13764147 0.23333480 0.44529206 0.43974483 0.19514136
#>  [361] 0.58918665 0.36542502 0.18208127 0.40886169 0.51751306 0.58550295
#>  [367] 0.26572417 0.44293900 0.24510031 0.74416780 0.54260104 0.65001319
#>  [373] 0.78753603 0.24698573 0.18889487 0.33951204 0.39643974 0.38216560
#>  [379] 0.80901873 0.42905465 0.36768593 0.11111374 0.48579147 0.27053395
#>  [385] 0.18794050 0.35679669 0.50970222 0.21063079 0.32839474 0.36168740
#>  [391] 0.22849760 0.13888518 0.15079773 0.38787397 0.26985203 0.41395195
#>  [397] 0.57970383 0.55598513 0.32986395 0.37528273 0.32307235 0.80613209
#>  [403] 0.15161565 0.51771426 0.34467490 0.26796154 0.30283185 0.18823713
#>  [409] 0.12364224 0.27865860 0.19731889 0.41120833 0.23139337 0.87998588
#>  [415] 0.20103549 0.57164461 0.48817171 0.28057809 0.31681914 0.09044520
#>  [421] 0.26049287 0.26205619 0.71105260 0.18861614 0.57779319 0.06839631
#>  [427] 0.26174682 0.09674180 0.13421758 0.19396591 0.17378084 0.82483171
#>  [433] 0.44331880 0.17664555 0.86338058 0.33183471 0.26351012 0.45904935
#>  [439] 0.86575729 0.29100790 0.18301734 0.25074373 0.13687087 0.14997008
#>  [445] 0.58371539 0.33442325 0.16856418 0.20422478 0.21189554 0.49686403
#>  [451] 0.69438898 0.22274104 0.25563632 0.19158637 0.39764902 0.20385287
#>  [457] 0.17205627 0.91027056 0.43847117 0.34841623 0.25726776 0.21979282
#>  [463] 0.40748506 0.20945399 0.36464183 0.15608154 0.21356331 0.30545650
#>  [469] 0.41183116 0.55316568 0.33809581 0.24678725 0.64753398 0.46573903
#>  [475] 0.45150310 0.20218292 0.10310356 0.52540036 0.84556495 0.27508773
#>  [481] 0.57536533 0.14728055 0.25675749 0.15945149 0.24121115 0.17952007
#>  [487] 0.13322564 0.44809391 0.70169225 0.31581427 0.56488833 0.17240006
#>  [493] 0.76906380 0.55923189 0.39228451 0.26167917 0.59699879 0.32820670
#>  [499] 0.21904575 0.61537906 0.39616774 0.23874378 0.84194557 0.28986463
#>  [505] 0.68017003 0.62223595 0.41594840 0.19860655 0.56457775 0.23032639
#>  [511] 0.19898131 0.09613898 0.23034247 0.40964390 0.34583138 0.24965928
#>  [517] 0.30088985 0.59422420 0.38668767 0.22415860 0.05761584 0.37552134
#>  [523] 0.56765221 0.11322334 0.76006671 0.20182717 0.28542685 0.14544783
#>  [529] 0.19144688 0.40046792 0.34762235 0.20702452 0.26742040 0.27318462
#>  [535] 0.44387624 0.24482066 0.54547121 0.23602787 0.64319705 0.44424483
#>  [541] 0.41077820 0.09757494 0.60220333 0.79217886 0.35086987 0.44408509
#>  [547] 0.52498269 0.53469854 0.22123549 0.27226098 0.30923306 0.30537125
#>  [553] 0.13326261 0.29070053 0.42002753 0.88344882 0.36731927 0.10388901
#>  [559] 0.37044533 0.13261909 0.34659885 0.19220920 0.97410215 0.29317135
#>  [565] 0.45763050 0.22794050 0.63299358 0.68380442 0.14183920 0.24656911
#>  [571] 0.29133360 0.58169728 0.12284546 0.28692445 0.54853255 0.66560554
#>  [577] 0.52285625 0.41648252 0.19779267 0.31924616 0.20470269 0.25353848
#>  [583] 0.41861605 0.16356704 0.23166264 0.29961215 0.15141337 0.22157654
#>  [589] 0.37760156 0.52628137 0.59799848 0.55606554 0.63466427 0.89545990
#>  [595] 0.33574789 0.10760948 0.84634200 0.15392675 0.14135838 0.39423026
#>  [601] 0.22056318 0.09961896 0.44249669 0.22056404 0.41330548 0.18400053
#>  [607] 0.10361051 0.63063316 0.15019126 0.31989547 0.35480451 0.18596191
#>  [613] 0.47710087 0.46576287 0.11274947 0.30173683 0.30142289 0.09614906
#>  [619] 0.24530012 0.39089046 0.35273660 0.83447875 0.36822720 0.40598205
#>  [625] 0.20034682 0.27848045 0.15037775 0.27689018 0.48466051 0.51298793
#>  [631] 0.24722077 0.16518041 0.32825617 0.82134807 0.78193002 0.17936449
#>  [637] 0.13383164 0.44215486 0.23568571 0.23080089 0.10762997 0.27043399
#>  [643] 0.33538911 0.36776147 0.52274471 0.24268514 0.63008051 0.53370364
#>  [649] 0.23091793 0.44068717 0.23552269 0.53286433 0.20998173 0.78584373
#>  [655] 0.43126647 0.41416417 0.15661620 0.28505093 0.42159623 0.18838712
#>  [661] 0.49175126 0.18328513 0.58631942 0.37399173 0.17950016 0.35727145
#>  [667] 0.28119803 0.07955125 0.13152219 0.28357379 0.57382511 0.35379517
#>  [673] 0.27816091 0.13114253 0.59526311 0.27241012 0.60727678 0.23440918
#>  [679] 0.22296180 0.47985284 0.82916101 0.11606383 0.20498245 0.23454506
#>  [685] 0.09344711 0.16562968 0.96826336 0.30204102 0.19502911 0.29745895
#>  [691] 0.35796734 0.13363581 0.31316502 0.34171863 0.13344625 0.38920233
#>  [697] 0.31986512 0.43605546 0.37665743 0.24008674 0.37939303 0.34650306
#>  [703] 0.09880228 0.62532317 0.25315879 0.34212100 0.19221005 0.28448749
#>  [709] 0.24991670 0.30553865 0.27349984 0.41241457 0.34270141 0.40930715
#>  [715] 0.30640462 0.48399064 0.34556223 0.14917508 0.49880300 0.19526344
#>  [721] 0.56022144 0.26196276 0.44018572 0.12874985 0.49137374 0.12518962
#>  [727] 0.39327974 0.12206010 0.25400295 0.25865947 0.24580441 0.76894238
#>  [733] 0.72442597 0.70617129 0.07801343 0.24786672 0.52262366 0.18875951
#>  [739] 0.37176573 0.32147939 0.31456908 0.31363280 0.14968923 0.19157536
#>  [745] 0.29345338 0.45899617 0.50505514 0.15527728 0.25916046 0.44646174
#>  [751] 0.35235644 0.29674597 0.18877156 0.21061508 0.45120952 0.15996686
#>  [757] 0.51620488 0.21434983 0.24560955 0.10221850 0.20588126 0.17844242
#>  [763] 0.63246256 0.78196190 0.49822555 0.36357682 0.33200930 0.21309873
#>  [769] 0.16708697 0.08881225 0.20555056 0.35835161 0.35826908 0.26657924
#>  [775] 0.22691718 0.21466527 0.70528051 0.17972905 0.11009954 0.17942589
#>  [781] 0.22241788 0.14440333 0.08129939 0.15715838 0.15321017 0.15337998
#>  [787] 0.16410674 0.28019279 0.27035691 0.33606123 0.56193560 0.16707908
#>  [793] 0.22921311 0.32992411 0.39996164 0.25670058 0.22219735 0.44355915
#>  [799] 0.24208862 0.60463417 0.15004988 0.58226502 0.44964705 0.40824201
#>  [805] 0.07754796 0.30433701 0.16651163 0.32676195 0.17451282 0.14976609
#>  [811] 0.36091937 0.39823760 0.19581835 0.32747310 0.49664143 0.34304801
#>  [817] 0.39486110 0.07495350 0.40219264 0.30974841 0.13509138 0.87513819
#>  [823] 0.24580037 0.66758876 0.24661197 0.40006529 0.24501098 0.16162765
#>  [829] 0.18991282 0.32221635 0.22624499 0.29504161 0.31490172 0.49477757
#>  [835] 0.16849002 0.30625978 0.30589588 0.50111959 0.50253646 0.46942556
#>  [841] 0.05493877 0.17109088 0.27036397 0.23489846 0.71245724 0.37655480
#>  [847] 0.13402351 0.07662874 0.10461270 0.84837957 0.13830672 0.61045450
#>  [853] 0.32959832 0.78623861 0.63893666 0.46133176 0.60044968 0.34517539
#>  [859] 0.25533498 0.35810795 0.27022439 0.33430041 0.16471287 0.29547941
#>  [865] 0.49001857 0.19612597 0.21075417 0.33681674 0.24149772 0.45573745
#>  [871] 0.21337843 0.52976313 0.31098329 0.14712374 0.22517645 0.09480894
#>  [877] 0.14887317 0.41889550 0.87970899 0.13910429 0.34669129 0.30691029
#>  [883] 0.50829332 0.11984134 0.34284194 0.20765233 0.17175524 0.24199129
#>  [889] 0.19904008 0.14463058 0.65718845 0.26758952 0.64549319 0.45356670
#>  [895] 0.34017413 0.51883347 0.28083548 0.60956086 0.12888855 0.25670505
#>  [901] 0.43336192 0.55920401 0.14300323 0.22672595 0.32660901 0.64246293
#>  [907] 0.16719116 0.28811802 0.14126352 0.24821490 0.41357552 0.51153522
#>  [913] 0.20423902 0.20418072 0.71319285 0.48603147 0.68650049 0.17942828
#>  [919] 0.26429745 0.45853002 0.29518465 0.10264846 0.14841040 0.79154497
#>  [925] 0.43985041 0.40542211 0.23714801 0.27389871 0.32224917 0.23813464
#>  [931] 0.41203980 0.47305563 0.31034597 0.07223305 0.33031463 0.20874380
#>  [937] 0.19106206 0.25505279 0.25816927 0.23694965 0.50525975 0.39704215
#>  [943] 0.12124649 0.22915855 0.32019862 0.35826599 0.22529559 0.21824819
#>  [949] 0.20562707 0.18464324 0.24899064 0.50526168 0.22310710 0.43771183
#>  [955] 0.47946650 0.59679687 0.28717998 0.16694443 0.23787879 0.29509555
#>  [961] 0.12942761 0.77689019 0.10047548 0.61023712 0.46273238 0.28250467
#>  [967] 0.39744768 0.38873753 0.14983519 0.51616586 0.27001997 0.43600198
#>  [973] 0.42807755 0.35107249 0.95181414 0.15375124 0.38871618 0.51782587
#>  [979] 0.58179528 0.13524246 0.18370264 0.52612546 0.30566213 0.64501356
#>  [985] 0.36762937 0.16085340 0.27849322 0.35435752 0.59023435 0.67291586
#>  [991] 0.26859875 0.16008509 0.25922336 0.65224686 0.19544146 0.25901957
#>  [997] 0.36969692 0.12670622 0.59881156 0.17556604 0.16844832 0.27463787
#> [1003] 0.09090942 0.14254683 0.26049083 0.27345628 0.10423300 0.13298200
#> [1009] 0.15370988 0.19545415 0.76825893 0.24622959 0.23863644 0.26157249
#> [1015] 0.55593182 0.45615798 0.17220013 0.28611508 0.21501526 0.09611997
#> [1021] 0.20987686 0.29047151 0.72123253 0.42058272 0.32955651 0.16351150
#> [1027] 0.55756921 0.70974760 0.09203726 0.27166633 0.11306045 0.04587936
#> [1033] 0.43516138 0.33853893 0.22160931 0.41018033 0.41025614 0.78610881
#> [1039] 0.49436514 0.67534880 0.22962964 0.18955782 0.10832449 0.40986964
#> [1045] 0.24043180 0.59428429 0.11400736 0.27509867 0.41901690 0.51778354
#> [1051] 0.20411195 0.16634075 0.34004707 0.34388193 0.21051737 0.11989122
#> [1057] 0.28699053 0.19558839 0.33899236 0.42853849 0.29641470 0.55779536
#> [1063] 0.17869205 0.22832169 0.15441751 0.16123584 0.42351564 0.27754244
#> [1069] 0.15582471 0.10799223 0.09453062 0.19102634 0.14788599 0.49076958
#> [1075] 0.71869542 0.40147327 0.36788767 0.12193549 0.48033983 0.48072718
#> [1081] 0.31769801 0.44886931 0.11617407 0.14374641 0.50088212 0.23763377
#> [1087] 0.12437544 0.26080367 0.48713792 0.10621015 0.26843415 0.14377850
#> [1093] 0.21322400 0.25488525 0.17750370 0.59736627 0.43158879 0.70585484
#> [1099] 0.31508175 0.25268903 0.19213272 0.48981273 0.34895125 0.63019235
#> [1105] 0.17451594 0.39633576 0.27647529 0.40998495 0.39412243 0.17795413
#> [1111] 0.59914690 0.10134492 0.16555811 0.08964434 0.49420301 0.56618206
#> [1117] 0.06828015 0.43443842 0.38721109 0.23377843 0.24182443 0.72053090
#> [1123] 0.23836016 0.13932490 0.48047258 0.28351433 0.56064677 0.90617946
#> [1129] 0.25507645 0.71199141 0.63209845 0.43690712 0.71717734 0.19698049
#> [1135] 0.67302953 0.13802105 0.27768607 0.18009536 0.12020838 0.72231380
#> [1141] 0.13750288 0.21909236 0.34246902 0.15371398 0.32892171 0.12067959
#> [1147] 0.82518465 0.35873555 0.23765295 0.49915347 0.27696718 0.33216395
#> [1153] 0.31328299 0.78755722 0.27919398 0.73040354 0.24373037 0.26663462
#> [1159] 0.12678825 0.50104539 0.15870597 0.34073690 0.16235001 0.11391080
#> [1165] 0.58985577 0.42706307 0.36134865 0.49785585 0.22858970 0.33969927
#> [1171] 0.47060851 0.08917038 0.45183509 0.57119698 0.29073709 0.30903800
#> [1177] 0.31944742 0.24709111 0.09910619 0.69506621 0.29928599 0.23800070
#> [1183] 0.45672366 0.42773256 0.32443146 0.50300294 0.42484371 0.23579467
#> [1189] 0.16934166 0.26850770 0.17815889 0.17715412 0.29996534 0.36844229
#> [1195] 0.17173470 0.45848688 0.46372202 0.11159534 0.09217896 0.56857233
#> [1201] 0.38353873 0.21274721 0.27886526 0.27031386 0.32838123 0.32449582
#> [1207] 0.91027144 0.24269532 0.25505858 0.26400207 0.15570139 0.45289894
#> [1213] 0.23532599 0.73342501 0.24650741 0.11818902 0.34719422 0.12725424
#> [1219] 0.14700889 0.19512985 0.68640342 0.71385732 0.15574563 0.48795953
#> [1225] 0.30518109 0.29707032 0.66220875 0.34791017 0.40079066 0.66287491
#> [1231] 0.34544168 0.08680361 0.65406226 0.23075680 0.11588207 0.09510994
#> [1237] 0.60590115 0.50539854 0.15701265 0.28469284 0.32311168 0.40412677
#> [1243] 0.26627592 0.51462787 0.23914772 0.98997324 0.31175249 0.86555653
#> [1249] 0.24894682 0.34004085 0.43778180 0.54179930 0.46444220 0.50710320
#> [1255] 0.25898583 0.16148014 0.15933379 0.45053704 0.19615228 0.19467854
#> [1261] 0.26844068 0.18857435 0.19341205 0.79052911 0.21251095 0.82336990
#> [1267] 0.12875911 0.28649739 0.17022558 0.33736036 0.29998356 0.41472445
#> [1273] 0.26235082 0.10142351 0.29189036 0.46616118 0.32379794 0.28731540
#> [1279] 0.50800489 0.19348366 0.58830055 0.17494500 0.58840080 0.12985548
#> [1285] 0.14575258 0.13965633 0.10246161 0.37298021 0.17085933 0.19370826
#> [1291] 0.14957585 0.16351415 0.09440027 0.47025163 0.36886659 0.32694900
#> [1297] 0.32589165 0.15868942 0.34751274 0.16747182 0.32457153 0.35192945
#> [1303] 0.49135440 0.15979832 0.44528832 0.28600028 0.49548850 0.19845095
#> [1309] 0.10240437 0.19142997 0.54840109 0.70977760 0.18714719 0.39442315
#> [1315] 0.46603156 0.13355859 0.57894783 0.16514322 0.29751124 0.35704099
#> [1321] 0.15029810 0.09541673 0.16934021 0.09034561 0.29466303 0.28748528
#> [1327] 0.11045125 0.68777080 0.38494444 0.27751066 0.83346745 0.21271075
#> [1333] 0.09978378 0.33041283 0.42740603 0.29704401 0.21027556 0.28958604
#> [1339] 0.52017896 0.57007877 0.47641164 0.49166343 0.29774476 0.17052807
#> [1345] 0.25947591 0.08943600 0.17768926 0.13422896 0.49706811 0.28256124
#> [1351] 0.19722751 0.25451248 0.29623056 0.28476080 0.42984814 0.24349989
#> [1357] 0.29906812 0.41944408 0.49355179 0.20957232 0.75973542 0.24839195
#> [1363] 0.38948528 0.64992425 0.59857731 0.33001322 0.18964683 0.18245496
#> [1369] 0.20807698 0.27382522 0.16873819 0.32541218 0.35548324 0.31363786
#> [1375] 0.41961805 0.56956552 0.17773255 0.71006689 0.11443131 0.32592704
#> [1381] 0.42206779 0.62170458 0.11668286 0.46478294 0.16897330 0.34330402
#> [1387] 0.34160089 0.30612698 0.33094515 0.35933659 0.46936816 0.74550837
#> [1393] 0.19040727 0.22089382 0.16510123 0.32897651 0.12705493 0.29130449
#> [1399] 0.36446921 0.11678777 0.44331770 0.23693669 0.29752647 0.29709501
#> [1405] 0.14283127 0.09325975 0.26938580 0.25615078 0.27593034 0.42313249
#> [1411] 0.38060049 0.18161998 0.40402902 0.30415246 0.29302871 0.31648110
#> [1417] 0.24845607 0.35957549 0.94142474 0.19705850 0.64152770 0.30323063
#> [1423] 0.82879619 0.43769343 0.29702786 0.57897531 0.26948478 0.24085412
#> [1429] 0.44311316 0.33817062 0.50556453 0.31907994 0.15233780 0.43416179
#> [1435] 0.24921442 0.48242861 0.84819799 0.19077315 0.48305955 0.62526740
#> [1441] 0.44788674 0.32984838 0.30881082 0.35399321 0.23622443 0.22300878
#> [1447] 0.83669674 0.35836417 0.14510570 0.24748610 0.58076626 0.38734845
#> [1453] 0.50799780 0.27237348 0.35354591 0.08082764 0.23986287 0.70751410
#> [1459] 0.29494051 0.16697218 0.18804510 0.27178129 0.18136313 0.62452813
#> [1465] 0.35670702 0.15386684 0.17298105 0.15643773 0.25417794 0.19754581
#> [1471] 0.22623743 0.16164564 0.15442788 0.14930163 0.60498236 0.25603554
#> [1477] 0.77646903 0.46125624 0.20167114 0.15478496 0.46307172 0.13914312
#> [1483] 0.56677744 0.28955228 0.14714199 0.32919060 0.06212788 0.39540133
#> [1489] 0.27767756 0.16134072 0.32176213 0.91448366 0.45724590 0.16578760
#> [1495] 0.24635546 0.38157431 0.46607462 0.30046372 0.70375907 0.38286486
#> [1501] 0.14813608 0.18994309 0.17655937 0.48371464 0.27990189 0.50973953
#> [1507] 0.16068341 0.88681891 0.25049727 0.77921605 0.19517735 0.56332846
#> [1513] 0.68293274 0.20829113 0.23607172 0.26681578 0.38128904 0.33602025
#> [1519] 0.24034423 0.79110658 0.46753638 0.31453757 0.18708814 0.16109954
#> [1525] 0.09250487 0.54657091 0.12831675 0.18289366 0.73952704 0.21601941
#> [1531] 0.28531787 0.19637451 0.24095678 0.12150403 0.51159823 0.15419666
#> [1537] 0.18184162 0.89603438 0.20874321 0.69476175 0.10675390 0.24651339
#> [1543] 0.26201779 0.12861761 0.16302947 0.42543041 0.43794142 0.22456917
#> [1549] 0.40546553 0.14992388 0.16472796 0.49487413 0.44551404 0.29485050
#> [1555] 0.40029510 0.16466425 0.92244595 0.49528128 0.11335277 0.22421294
#> [1561] 0.43161607 0.37292387 0.32286283 0.50645115 0.28032529 0.67002799
#> [1567] 0.48080961 0.16158723 0.17338503 0.26052930 0.20748182 0.35486022
#> [1573] 0.50757730 0.20544762 0.38007290 0.20031239 0.40030505 0.10102985
#> [1579] 0.31976584 0.20459121 0.28720719 0.55501983 0.21366595 0.31804962
#> [1585] 0.41751415 0.26900534 0.38777801 0.34078625 0.15686950 0.21767216
#> [1591] 0.26383660 0.73681448 0.57712461 0.61797259 0.52610395 0.82873268
#> [1597] 0.20957329 0.24597260 0.22535619 0.46346681 0.19985983 0.13479857
#> [1603] 0.42638471 0.08982070 0.14267158 0.29844302 0.41529995 0.12530028
#> [1609] 0.14884601 0.42485395 0.31133607 0.20081939 0.32666892 0.39717025
#> [1615] 0.24064462 0.14950181 0.22325025 0.29377889 0.22080242 0.27053410
#> [1621] 0.24056812 0.12468318 0.40671349 0.21575321 0.53520328 0.21968949
#> [1627] 0.97686827 0.28072872 0.28167832 0.38620974 0.27631699 0.13680230
#> [1633] 0.12039558 0.30916109 0.27576856 0.23779720 0.51168568 0.96643684
#> [1639] 0.77430095 0.26707303 0.35703313 0.23958112 0.30475075 0.62142331
#> [1645] 0.17397422 0.78494412 0.21888787 0.33560659 0.09662564 0.57025725
#> [1651] 0.25215655 0.20242780 0.13424797 0.24339637 0.19099030 0.14984986
#> [1657] 0.24534963 0.40751536 0.46482835 0.17838721 0.24474219 0.18720237
#> [1663] 0.54338831 0.21135029 0.32553249 0.21965561 0.24088247 0.44665285
#> [1669] 0.18491890 0.27447158 0.43420508 0.13472963 0.36494513 0.52568181
#> [1675] 0.51026237 0.25282606 0.40798067 0.51973590 0.17962500 0.79064839
#> [1681] 0.49743770 0.56736758 0.35366047 0.67367988 0.40125558 0.37208131
#> [1687] 0.17069337 0.18914855 0.20701575 0.49954426 0.19253248 0.44565410
#> [1693] 0.20137270 0.15201795 0.62053106 0.39999470 0.15526016 0.25144272
#> [1699] 0.47043580 0.28164449 0.42439080 0.27792381 0.24429040 0.07933588
#> [1705] 0.14899411 0.35077290 0.32927694 0.26536173 0.66772164 0.25341310
#> [1711] 0.63855424 0.07960260 0.20090165 0.37617208 0.23978448 0.60713429
#> [1717] 0.11924977 0.30082524 0.86269447 0.22879746 0.34864771 0.25602866
#> [1723] 0.10788598 0.20236738 0.33319339 0.68567136 0.25337846 0.17456063
#> [1729] 0.35660205 0.15388158 0.23213793 0.60239191 0.08745373 0.27588599
#> [1735] 0.39034294 0.58012565 0.14943575 0.22288735 0.26143818 0.63608043
#> [1741] 0.37824272 0.22508718 0.44619335 0.50561279 0.23383024 0.36499821
#> [1747] 0.39699996 0.11841605 0.83440046 0.19313279 0.54063791 0.49908407
#> [1753] 0.18829676 0.46695443 0.40343749 0.22636035 0.22909565 0.24599073
#> [1759] 0.57321704 0.15871693 0.40442307 0.21112538 0.47577587 0.14851956
#> [1765] 0.24985886 0.29549197 0.24584804 0.30302627 0.03720663 0.23211014
#> [1771] 0.66728572 0.31046682 0.20553812 0.15036334 0.37784447 0.76647209
#> [1777] 0.44386093 0.22580411 0.36522306 0.15134324 0.26056608 0.84083105
#> [1783] 0.15922918 0.30193698 0.26662827 0.54333501 0.33895030 0.51271451
#> [1789] 0.24469372 0.34328508 0.14703336 0.17038362 0.76526441 0.48772079
#> [1795] 0.59013510 0.86131217 0.47515735 0.12039931 0.36200331 0.32277930
#> [1801] 0.29505935 0.42790009 0.93332428 0.51592667 0.13956902 0.25956901
#> [1807] 0.16416821 0.46177237 0.28548444 0.22915109 0.17352875 0.44934959
#> [1813] 0.18564404 0.17559338 0.48727987 0.27031445 0.21377920 0.09600008
#> [1819] 0.67683754 0.50374076 0.21748718 0.24989613 0.17055179 0.31519312
#> [1825] 0.67264773 0.40124614 0.16797584 0.24568995 0.22523644 0.68597518
#> [1831] 0.30177083 0.26672331 0.40359841 0.28366914 0.20591586 0.33060346
#> [1837] 0.32346017 0.83815596 0.17933057 0.17033586 0.49260101 0.50906665
#> [1843] 0.30460791 0.32588595 0.89939266 0.40630700 0.12703764 0.46642574
#> [1849] 0.55134219 0.75408060 0.13368481 0.29685344 0.37827865 0.63031048
#> [1855] 0.16977358 0.91160671 0.46541932 0.16327379 0.10735810 0.11552663
#> [1861] 0.30830394 0.45309578 0.45683661 0.28228769 0.09039985 0.79043557
#> [1867] 0.18628936 0.29362214 0.64452399 0.40882627 0.16327450 0.19566306
#> [1873] 0.09818961 0.36894328 0.47734278 0.42718038 0.24940786 0.26026934
#> [1879] 0.33561179 0.21414667 0.33016766 0.13465811 0.58277241 0.51240665
#> [1885] 0.18531326 0.40307056 0.51274532 0.17904103 0.24504054 0.55537732
#> [1891] 0.37788512 0.53809645 0.20878885 0.17740980 0.63020721 0.61950293
#> [1897] 0.12655789 0.30106179 0.39847801 0.14028264 0.52064954 0.33301530
#> [1903] 0.52689437 0.30025431 0.26865425 0.24967499 0.43219386 0.28787112
#> [1909] 0.20817430 0.14850427 0.24731813 0.23330591 0.14670968 0.23977084
#> [1915] 0.13680392 0.14660669 0.47452632 0.16952888 0.14495604 0.30825260
#> [1921] 0.20699036 0.35473864 0.33323165 0.62536535 0.18826109 0.66240730
#> [1927] 0.38846732 0.89751674 0.08873358 0.16882270 0.25088460 0.66252715
#> [1933] 0.39384309 0.15882550 0.34474967 0.35148505 0.39742619 0.32007745
#> [1939] 0.16913141 0.29698015 0.46723944 0.39455137 0.60294855 0.30834367
#> [1945] 0.54790159 0.28537138 0.13911708 0.08364646 0.45984083 0.30314429
#> [1951] 0.20947582 0.24864471 0.14535340 0.58336244 0.33642462 0.40519318
#> [1957] 0.95477787 0.22834646 0.36922996 0.10565253 0.39221011 0.30844253
#> [1963] 0.10322444 0.48777021 0.17542033 0.26496095 0.42806470 0.57376753
#> [1969] 0.27805694 0.37219730 0.05827624 0.65636013 0.29421157 0.71768825
#> [1975] 0.11378768 0.36953118 0.17127625 0.68563038 0.23438498 0.32944553
#> [1981] 0.45440014 0.42524112 0.41499353 0.49589944 0.31390373 0.25581032
#> [1987] 0.29014355 0.14111475 0.29796465 0.25026072 0.17435986 0.38877684
#> [1993] 0.47672512 0.37079047 0.12276376 0.29324561 0.10401886 0.30473704
#> [1999] 0.20872782 0.26290921 0.33591802 0.92137311 0.14234610 0.23942178
#> [2005] 0.36041789 0.10217599 0.49436303 0.15382550 0.31902718 0.55814644
#> [2011] 0.48100474 0.30473804 0.32159620 0.11045330 0.25491568 0.15684114
#> [2017] 0.32057533 0.61906651 0.34480905 0.15700995 0.20374483 0.28579861
#> [2023] 0.63969488 0.28748766 0.25814215 0.22510456 0.35409211 0.69293160
#> [2029] 0.46530768 0.15805185 0.16620327 0.30220398 0.17029261 0.55740076
#> [2035] 0.44185450 0.59469723 0.43136743 0.70955014 0.23542017 0.40884870
#> [2041] 0.29274629 0.35806224 0.25328792 0.24248262 0.12088865 0.11393725
#> [2047] 0.36644220 0.28781214 0.81779935 0.20340955 0.36477598 0.58883279
#> [2053] 0.11601208 0.29476183 0.53157747 0.32997576 0.19846004 0.39251910
#> [2059] 0.13031984 0.18584217 0.17703939 0.58361251 0.30510083 0.12381000
#> [2065] 0.19194472 0.50720627 0.68395618 0.18346816 0.46720921 0.20313752
#> [2071] 0.44742840 0.69775034 0.36513031 0.19137437 0.16584476 0.17320328
#> [2077] 0.10822292 0.84643406 0.55611496 0.17182477 0.26739066 0.72732090
#> [2083] 0.20992797 0.69895780 0.25570481 0.31391842 0.26119952 0.56146853
#> [2089] 0.10750162 0.96229190 0.16384858 0.13480276 0.11694893 0.38934394
#> [2095] 0.89157808 0.31484151 0.72021587 0.43832393 0.14778296 0.30833429
#> [2101] 0.44322811 0.19366067 0.93387447 0.39447296 0.19209535 0.42669608
#> [2107] 0.21442680 0.27182414 0.16379104 0.34892889 0.34910120 0.11802056
#> [2113] 0.37454380 0.39671308 0.14814457 0.14908793 0.44628521 0.10636603
#> [2119] 0.52964213 0.22543157 0.21563383 0.27911098 0.43908195 0.55601747
#> [2125] 0.28760463 0.11421824 0.22926710 0.37061754 0.43069865 0.66017434
#> [2131] 0.33596908 0.31018509 0.84489343 0.28967349 0.16441339 0.31988477
#> [2137] 0.20583132 0.72644711 0.14020961 0.65358116 0.30656359 0.15434804
#> [2143] 0.15487703 0.29238004 0.31031689 0.36519228 0.53791121 0.36220878
#> [2149] 0.31279071 0.47016789 0.21210117 0.45981238 0.54781160 0.42371615
#> [2155] 0.89571582 0.13926035 0.18216881 0.38145657 0.22734653 0.34382921
#> [2161] 0.22091795 0.14656323 0.17816848 0.17012635 0.29158745 0.28394093
#> [2167] 0.43110710 0.21856774 0.11708469 0.26010703 0.16542902 0.29805750
#> [2173] 0.67398936 0.59006138 0.39659529 0.11479672 0.35905624 0.22445961
#> [2179] 0.12915057 0.66915888 0.45269619 0.96238803 0.59512532 0.21315616
#> [2185] 0.37607537 0.68164471 0.44215598 0.34261436 0.63861592 0.48555364
#> [2191] 0.48570094 0.34602535 0.79693223 0.28600563 0.10249221 0.21716973
#> [2197] 0.59978676 0.38161944 0.29757812 0.45059084 0.18441400 0.38809882
#> [2203] 0.33761369 0.43191008 0.57632547 0.30680377 0.14357337 0.13982408
#> [2209] 0.23903478 0.37802504 0.30403927 0.34721105 0.10364473 0.56975165
#> [2215] 0.42386926 0.43346905 0.18709681 0.16104004 0.32065103 0.18859310
#> [2221] 0.39190010 0.40032260 0.37426482 0.53657914 0.26970290 0.36678362
#> [2227] 0.31803964 0.23800372 0.34399691 0.22117708 0.08282791 0.16745441
#> [2233] 0.28354883 0.41363685 0.26559201 0.66724832 0.22191131 0.30371921
#> [2239] 0.66067196 0.61038831 0.12758851 0.34137471 0.08336978 0.71791127
#> [2245] 0.22482745 0.31531462 0.31430223 0.15156489 0.29265742 0.58963728
#> [2251] 0.36830591 0.68954336 0.36098423 0.22397689 0.80025402 0.21711321
#> [2257] 0.12877457 0.21561604 0.20012501 0.22546592 0.23722098 0.10836404
#> [2263] 0.86289083 0.49953677 0.47247248 0.30283150 0.64966332 0.24202175
#> [2269] 0.11124215 0.27686219 0.16220784 0.77914694 0.31499402 0.32036671
#> [2275] 0.17737858 0.39006836 0.13186620 0.22743066 0.22775678 0.06588281
#> [2281] 0.14319388 0.85086075 0.28562610 0.54481145 0.34455265 0.29851658
#> [2287] 0.13081950 0.27632126 0.22128527 0.20130532 0.71797643 0.74943574
#> [2293] 0.24435511 0.45961236 0.19298205 0.44511897 0.26022869 0.47725153
#> [2299] 0.82460774 0.52831762 0.77709947 0.77450224 0.56301669 0.12843009
#> [2305] 0.16353841 0.81226194 0.35631712 0.29118695 0.35424062 0.25786608
#> [2311] 0.38664628 0.60816095 0.67219650 0.33670731 0.30625942 0.81378984
#> [2317] 0.11684345 0.67548547 0.50947159 0.11915339 0.30088526 0.22391440
#> [2323] 0.17303850 0.26959121 0.53201793 0.17575463 0.53365125 0.51254084
#> [2329] 0.33348809 0.19670635 0.43134440 0.39004776 0.13972411 0.29586136
#> [2335] 0.17644738 0.39301279 0.43231800 0.14575306 0.13094875 0.39849038
#> [2341] 0.26051100 0.33826250 0.17682261 0.36845888 0.23799557 0.56341742
#> [2347] 0.62655605 0.10171164 0.21200796 0.11059279 0.35963387 0.20312042
#> [2353] 0.64241884 0.29942611 0.19084154 0.35532016 0.30277444 0.68114650
#> [2359] 0.75981525 0.26027685 0.13593979 0.26960369 0.37531290 0.49328432
#> [2365] 0.43410303 0.20624812 0.56215188 0.41754285 0.08750615 0.28340310
#> [2371] 0.07279654 0.84927996 0.35571269 0.37239612 0.42725424 0.15652217
#> [2377] 0.39917381 0.51231055 0.34503386 0.58306384 0.47682154 0.15924306
#> [2383] 0.32922499 0.50661384 0.32312162 0.19383353 0.39791676 0.32009031
#> [2389] 0.41550713 0.47732729 0.81802757 0.52489589 0.16620037 0.27813164
#> [2395] 0.37080727 0.19950099 0.33389319 0.13339156 0.49982912 0.71611158
#> [2401] 0.65303501 0.28793467 0.36102007 0.19344674 0.51918140 0.12786901
#> [2407] 0.09009902 0.22545054 0.38384268 0.22024686 0.19718023 0.31037028
#> [2413] 0.29202460 0.76865584 0.20399458 0.19941945 0.32118516 0.26120560
#> [2419] 0.14305610 0.25406140 0.21341421 0.48838959 0.18007473 0.57129770
#> [2425] 0.24353421 0.30161445 0.28798733 0.19220300 0.26116456 0.36613404
#> [2431] 0.30918842 0.52777165 0.09969069 0.19641183 0.12389810 0.43004031
#> [2437] 0.26044082 0.21531530 0.29544922 0.12007471 0.22450863 0.31856027
#> [2443] 0.42231824 0.37643444 0.58570528 0.23271180 0.33704064 0.19664939
#> [2449] 0.64997337 0.12358172 0.29536524 0.49186850 0.48058713 0.26058133
#> [2455] 0.51765956 0.37407536 0.16041643 0.33734494 0.17245319 0.18553366
#> [2461] 0.37017217 0.33618119 0.23449765 0.76713068 0.52381922 0.20668422
#> [2467] 0.13953855 0.56565214 0.21292223 0.66720872 0.10010659 0.33322174
#> [2473] 0.13441808 0.35262424 0.67290562 0.08572806 0.36599209 0.10388656
#> [2479] 0.16241892 0.23271553 0.41263928 0.12194909 0.75731040 0.32343672
#> [2485] 0.21895023 0.12411300 0.24514951 0.15223715 0.71686549 0.42730263
#> [2491] 0.09484399 0.66306894 0.29917377 0.11238658 0.15399360 0.17392011
#> [2497] 0.51671100 0.08509184 0.76924582 0.36943833 0.20557206 0.31463880
#> [2503] 0.79279832 0.65474899 0.68239917 0.11142493 0.22723982 0.24425346
#> [2509] 0.17855008 0.59067826 0.12508352 0.68500588 0.16168268 0.39975139
#> [2515] 0.12580995 0.21661330 0.21613806 0.32131562 0.47062822 0.39971766
#> [2521] 0.22071747 0.28708016 0.49160268 0.10085508 0.38431426 0.24015521
#> [2527] 0.41168980 0.17922557 0.90436062 0.10633639 0.20071000 0.35849642
#> [2533] 0.53491895 0.35841483 0.24435748 0.33441667 0.13677425 0.29165514
#> [2539] 0.29654463 0.16690607 0.22445801 0.75737728 0.24841009 0.35840815
#> [2545] 0.41182878 0.20399829 0.12595667 0.63582354 0.10491383 0.23509422
#> [2551] 0.66844738 0.12775403 0.33771434 0.38863314 0.23413001 0.37143248
#> [2557] 0.23486724 0.48737474 0.24391932 0.17250674 0.14182666 0.13466756
#> [2563] 0.29221215 0.41654170 0.43088068 0.33435637 0.25814404 0.22300971
#> [2569] 0.31196589 0.22764713 0.25251748 0.17742444 0.09347100 0.30613964
#> [2575] 0.21354745 0.22928908 0.32692816 0.18646731 0.28692478 0.31678280
#> [2581] 0.32078059 0.17312058 0.15469807 0.26558393 0.44042379 0.34807699
#> [2587] 0.26449573 0.19960756 0.45440003 0.12860052 0.12200414 0.14224497
#> [2593] 0.43604986 0.22731688 0.19347700 0.16915192 0.65160301 0.20281963
#> [2599] 0.34784736 0.22445013 0.10483367 0.21403574 0.21603774 0.28943655
#> [2605] 0.85035627 0.20557378 0.09117493 0.18485502 0.18028918 0.24443370
#> [2611] 0.13426928 0.18504149 0.37078253 0.14161209 0.34469398 0.21595051
#> [2617] 0.20471192 0.50918961 0.23111489 0.14122636 0.45313838 0.09200036
#> [2623] 0.35798175 0.19832487 0.22194894 0.19049745 0.52690347 0.26175968
#> [2629] 0.10019593 0.65771526 0.36388917 0.42538399 0.45243442 0.34336757
#> [2635] 0.97647155 0.23209490 0.17365911 0.17944650 0.39326456 0.13804554
#> [2641] 0.26692391 0.76991453 0.23953373 0.24782637 0.94683755 0.46547528
#> [2647] 0.35194652 0.08343386 0.12054814 0.28293039 0.23160184 0.21974911
#> [2653] 0.40335967 0.28084409 0.18323110 0.76942399 0.43770772 0.25703920
#> [2659] 0.45503919 0.33794184 0.64690860 0.28779460 0.37320670 0.40421530
#> [2665] 0.69919340 0.23547377 0.19036835 0.46659270 0.21788389 0.58009549
#> [2671] 0.18719533 0.25383441 0.22149682 0.15542381 0.15772948 0.72465717
#> [2677] 0.35163208 0.29562244 0.59377503 0.31064121 0.38865005 0.17435550
#> [2683] 0.15405932 0.18527428 0.38674542 0.14444174 0.36451593 0.44374819
#> [2689] 0.92946484 0.53910224 0.07177270 0.50163137 0.24183831 0.32577133
#> [2695] 0.30565037 0.14685509 0.24256649 0.36685459 0.49884209 0.17512617
#> [2701] 0.12797509 0.21281069 0.15135470 0.59543144 0.49414858 0.27655262
#> [2707] 0.39637766 0.29289367 0.09496895 0.94329734 0.17290389 0.52565506
#> [2713] 0.17178251 0.50802483 0.53228482 0.96738284 0.15693783 0.22681144
#> [2719] 0.17065639 0.19480048 0.28999702 0.48116019 0.11079552 0.13625347
#> [2725] 0.29184584 0.26395798 0.40699522 0.15155859 0.12357633 0.44991230
#> [2731] 0.58582938 0.14795274 0.41249690 0.17308834 0.32898704 0.45027211
#> [2737] 0.29962315 0.20381539 0.44731412 0.27653205 0.28009082 0.76293147
#> [2743] 0.14671317 0.59284262 0.24185345 0.74186891 0.58279584 0.84726538
#> [2749] 0.31908224 0.17235849 0.61644314 0.50232773 0.35996810 0.30643585
#> [2755] 0.62015194 0.07610334 0.40858097 0.18670143 0.22012415 0.24133282
#> [2761] 0.35001454 0.18761322 0.55806197 0.17976897 0.29121884 0.51129610
#> [2767] 0.31884135 0.10209063 0.58841151 0.25641559 0.26589067 0.12748411
#> [2773] 0.16296489 0.39065669 0.41712567 0.61088762 0.16473498 0.08512943
#> [2779] 0.62107264 0.44506274 0.15156012 0.18635890 0.17061900 0.45460246
#> [2785] 0.23262737 0.37713654 0.37871876 0.36477514 0.09152202 0.20968695
#> [2791] 0.23910927 0.24470498 0.19969321 0.34838369 0.23787740 0.61304805
#> [2797] 0.34955277 0.18348738 0.35298124 0.08313908 0.54200365 0.35663130
#> [2803] 0.24563036 0.29836501 0.38152054 0.15589106 0.84777359 0.59083497
#> [2809] 0.26690507 0.25172444 0.18501660 0.37361269 0.35460123 0.21612665
#> [2815] 0.19861364 0.20929167 0.93753455 0.50869463 0.62549658 0.24863184
#> [2821] 0.28199587 0.53613737 0.10940310 0.15493992 0.48403627 0.82722339
#> [2827] 0.42851711 0.23524185 0.59605028 0.16843513 0.72350120 0.27533245
#> [2833] 0.26300331 0.16304988 0.34558689 0.29945289 0.25222337 0.17311819
#> [2839] 0.29082802 0.66419500 0.21381332 0.26288715 0.30660114 0.61511146
#> [2845] 0.18219247 0.17605088 0.17364040 0.54006841 0.18439677 0.68953129
#> [2851] 0.39486616 0.20750705 0.74784538 0.17874959 0.96742817 0.69881911
#> [2857] 0.50292154 0.09407658 0.09466234 0.56251550 0.25208580 0.14005896
#> [2863] 0.42482855 0.38679575 0.44449196 0.16394831 0.19227909 0.63778475
#> [2869] 0.40342959 0.37814092 0.18646893 0.19278608 0.56473430 0.26057642
#> [2875] 0.28650079 0.09535334 0.17900116 0.26981363 0.33108146 0.64137302
#> [2881] 0.30272171 0.34856309 0.72927341 0.65450377 0.15653975 0.16029724
#> [2887] 0.32570308 0.62596480 0.40608007 0.67778522 0.58396891 0.09074788
#> [2893] 0.30937100 0.55792793 0.94027234 0.70236775 0.31951839 0.31984257
#> [2899] 0.70770718 0.23576081 0.16186515 0.21316348 0.55175157 0.22141774
#> [2905] 0.20888917 0.11365142 0.21241101 0.24764559 0.35622086 0.07545234
#> [2911] 0.21702960 0.55184372 0.28171728 0.30691787 0.27093980 0.29414394
#> [2917] 0.14520078 0.26174083 0.05958685 0.33386783 0.21178014 0.20728393
#> [2923] 0.39204972 0.23804753 0.17618825 0.19158823 0.14030756 0.22668620
#> [2929] 0.82135095 0.49596844 0.56416491 0.57910903 0.55269013 0.44047413
#> [2935] 0.30759325 0.14325596 0.29082095 0.11508301 0.52880925 0.46207025
#> [2941] 0.18678264 0.67647824 0.41700427 0.21369420 0.19919808 0.28131274
#> [2947] 0.23083272 0.63135474 0.46415606 0.27898512 0.55602410 0.66912654
#> [2953] 0.56257161 0.15734754 0.11131888 0.49436757 0.19058407 0.28654672
#> [2959] 0.74739114 0.45926280 0.21182284 0.21717470 0.16291299 0.48652840
#> [2965] 0.58214147 0.17788765 0.48891628 0.55300234 0.25165504 0.51429870
#> [2971] 0.32567394 0.28928610 0.34261978 0.36231041 0.26934365 0.35954230
#> [2977] 0.86809383 0.12140009 0.15577944 0.12803001 0.09443357 0.14584774
#> [2983] 0.31433906 0.67242346 0.47951842 0.82963267 0.20590843 0.16897157
#> [2989] 0.10571817 0.26151156 0.14890084 0.21133433 0.21476351 0.24628190
#> [2995] 0.82506073 0.16205439 0.25698892 0.54895335 0.28793294 0.46362977
#> [3001] 0.23248315 0.35794249 0.24926258 0.12298135 0.33988674 0.64311942
#> [3007] 0.28429514 0.53264438 0.39148061 0.51809161 0.42471930 0.17943901
#> [3013] 0.35938581 0.70760431 0.31967889 0.43935950 0.16298077 0.50402174
#> [3019] 0.26192003 0.46964897 0.43720034 0.48712820 0.38168960 0.21433207
#> [3025] 0.29213382 0.30352760 0.22702538 0.35855354 0.41165734 0.16921450
#> [3031] 0.33828167 0.23404842 0.20450892 0.55304619 0.65550439 0.59632322
#> [3037] 0.15548457 0.38991967 0.25654557 0.11326884 0.22609295 0.35476688
#> [3043] 0.11094264 0.40380896 0.30504495 0.18626915 0.72654730 0.12268035
#> [3049] 0.38494952 0.09258753 0.29534757 0.42860938 0.66036042 0.37275981
#> [3055] 0.09684197 0.26853651 0.19376227 0.27505395 0.14457146 0.18366730
#> [3061] 0.18073696 0.15555821 0.77830841 0.15921324 0.36829193 0.19833929
#> [3067] 0.15676994 0.21083027 0.12037821 0.23568798 0.37059754 0.62908340
#> [3073] 0.38814450 0.60666511 0.46627469 0.41507422 0.25130071 0.13011061
#> [3079] 0.15121032 0.19790143 0.34655262 0.61631061 0.53172985 0.80947994
#> [3085] 0.55160950 0.50796827 0.24178992 0.27269247 0.10092659 0.19369288
#> [3091] 0.27286061 0.43255758 0.56163604 0.16509195 0.26510832 0.30597665
#> [3097] 0.34317663 0.48017427 0.43656493 0.11819820 0.22748820 0.53706423
#> [3103] 0.17352488 0.42726928 0.53205650 0.32975694 0.78060913 0.11183780
#> [3109] 0.60744955 0.25394719 0.19259711 0.41176749 0.71570597 0.10816726
#> [3115] 0.22143479 0.14457901 0.30738046 0.59548041 0.59614451 0.43714392
#> [3121] 0.26480590 0.51021622 0.18678850 0.29668635 0.20791294 0.13151894
#> [3127] 0.19722114 0.09604205 0.42747083 0.20934918 0.36770773 0.34659924
#> [3133] 0.29793164 0.18220950 0.22984441 0.65165004 0.32857572 0.22980729
#> [3139] 0.29873251 0.19073600 0.12218596 0.53478596 0.30750156 0.09208366
#> [3145] 0.35119773 0.41608086 0.08919013 0.15260188 0.14044986 0.42078653
#> [3151] 0.28836001 0.16598575 0.37040071 0.17900936 0.40026918 0.32407609
#> [3157] 0.27757340 0.29640831 0.42651186 0.22585042 0.25031069 0.15344844
#> [3163] 0.09656166 0.12088515 0.65331395 0.35831716 0.11034204 0.45758016
#> [3169] 0.34016860 0.32735523 0.10771373 0.30889561 0.55409066 0.15502808
#> [3175] 0.64089596 0.52857613 0.52418774 0.35252121 0.35596842 0.23972749
#> [3181] 0.12458022 0.32842671 0.60697731 0.33330951 0.18181796 0.33385415
#> [3187] 0.23734191 0.12335662 0.59966869 0.71523487 0.32766709 0.43480348
#> [3193] 0.35189158 0.33649081 0.22430651 0.17196674 0.78812296 0.18101879
#> [3199] 0.29900955 0.48348138 0.34161802 0.18037037 0.20202631 0.51535724
#> [3205] 0.30312252 0.26495800 0.49791650 0.19375966 0.10790469 0.15187355
#> [3211] 0.30004773 0.53594491 0.29012556 0.62016766 0.10284374 0.32025975
#> [3217] 0.37054223 0.43592204 0.26717058 0.50611716 0.26801911 0.19873324
#> [3223] 0.11400020 0.15388414 0.40267833 0.14614820 0.33415739 0.12835280
#> [3229] 0.69935945 0.48569026 0.47504497 0.12883033 0.43876821 0.24436293
#> [3235] 0.27116100 0.41163414 0.60168819 0.20693698 0.62648408 0.22869463
#> [3241] 0.17656295 0.39259272 0.40171173 0.48966390 0.27457899 0.16285167
#> [3247] 0.30524304 0.51605797 0.37897663 0.19712112 0.15398480 0.14607118
#> [3253] 0.43071465 0.18637705 0.21712632 0.57114834 0.56468083 0.17162796
#> [3259] 0.41574419 0.15264152 0.40484195 0.64070267 0.85247391 0.17495359
#> [3265] 0.57281795 0.22860338 0.45441063 0.15547758 0.24774355 0.08689213
#> [3271] 0.15298587 0.10641390 0.34144370 0.14630973 0.58195443 0.46629822
#> [3277] 0.48517133 0.19133072 0.61355183 0.23824797 0.48264500 0.64397385
#> [3283] 0.13507042 0.26769544 0.14581415 0.14194329 0.09254246 0.18881345
#> [3289] 0.13318877 0.42422232 0.54499381 0.57130232 0.59926175 0.55390741
#> [3295] 0.42446653 0.23609179 0.15417249 0.54538275 0.30983348 0.41987292
#> [3301] 0.80926020 0.85748603 0.11947927 0.33344745 0.17251561 0.15948675
#> [3307] 0.44826903 0.14961460 0.60609915 0.24242033 0.15419599 0.23821160
#> [3313] 0.30793685 0.12308940 0.14513781 0.16774796 0.29047305 0.76909788
#> [3319] 0.26539862 0.17965360 0.32089729 0.19919392 0.29076768 0.58614596
#> [3325] 0.08962508 0.12129932 0.35426756 0.41406206 0.17155886 0.14381357
#> [3331] 0.40292368 0.24347390 0.29233845