LP Depth

Computes the LP depth of a point or vectors of points with respect to a multivariate data set.

depthLP(u, X, pdim = 2, la = 1, lb = 1, threads = -1, func = NULL)

Arguments

u
Numerical vector or matrix whose depth is to be calculated. Dimension has to be the same as that of the observations.

X
The data as a matrix, data frame or list. If it is a matrix or data frame, then each row is viewed as one multivariate observation. If it is a list, all components must be numerical vectors of equal length (coordinates of observations).

pdim
dimension used in calculating depth function.

la
slope the weighing function.

lb
intercept in the weighing function.

threads
number of threads used in parallel computations. Default value -1 means that all possible cores will be used.

func
the weighing function. Currently it is not supported.

Details

Returns the depth of multivariate point u with respect to data set X.

Examples

x <- matrix(rnorm(3000), ncol = 3)

# Same results
depthLP(x, x, pdim = 2)

#> Depth method:  LP 
#>    [1] 0.2656365 0.2333785 0.2483234 0.2695947 0.2158412 0.2184458 0.2748394
#>    [8] 0.2676087 0.2363089 0.2602704 0.2753669 0.2363620 0.2732048 0.2357199
#>   [15] 0.2657356 0.2209029 0.2583631 0.2291361 0.2307944 0.2142348 0.1957403
#>   [22] 0.2596291 0.2109850 0.2612201 0.2694399 0.2722934 0.2558504 0.2622343
#>   [29] 0.2749219 0.2211371 0.2025077 0.2559913 0.2266122 0.2501357 0.2586593
#>   [36] 0.2298389 0.2689903 0.2517738 0.2761074 0.2236484 0.2692207 0.1937646
#>   [43] 0.2049432 0.2644730 0.2293096 0.2234517 0.1983188 0.1914841 0.2497572
#>   [50] 0.2377949 0.2760427 0.2364314 0.1881089 0.2544687 0.1980091 0.2308878
#>   [57] 0.2671725 0.2458655 0.2446072 0.2621539 0.2350018 0.2073066 0.1901897
#>   [64] 0.1967970 0.2633232 0.2583145 0.2093065 0.2393756 0.2428529 0.2554731
#>   [71] 0.2193116 0.2309039 0.2461559 0.2430573 0.2453877 0.2569231 0.2404934
#>   [78] 0.2010022 0.2208068 0.2358847 0.2398182 0.2347168 0.2581556 0.2431940
#>   [85] 0.2648152 0.2242152 0.1932521 0.2103191 0.2408629 0.2540924 0.2261219
#>   [92] 0.2091432 0.2277913 0.2510197 0.2571239 0.2168735 0.1987339 0.2636347
#>   [99] 0.2477432 0.2632005 0.2173886 0.2475616 0.1965299 0.2497301 0.2580690
#>  [106] 0.2133347 0.2224022 0.2324220 0.2143998 0.2003446 0.2517190 0.2294533
#>  [113] 0.2449630 0.2529757 0.2327376 0.1853130 0.2276302 0.2261517 0.1933991
#>  [120] 0.2677956 0.2398187 0.2645187 0.2551644 0.2057899 0.2359775 0.2277715
#>  [127] 0.2501314 0.2519114 0.2245952 0.2325475 0.2351775 0.2635078 0.2578968
#>  [134] 0.2779790 0.1878030 0.2332532 0.2224036 0.2632832 0.2310451 0.1879268
#>  [141] 0.2454612 0.2176221 0.2440479 0.2598990 0.2706905 0.2013545 0.2287385
#>  [148] 0.2132423 0.2086318 0.2381307 0.2566404 0.2525043 0.2306540 0.2395458
#>  [155] 0.1932492 0.2498326 0.2475905 0.2543140 0.2579938 0.2114971 0.2175167
#>  [162] 0.2398626 0.2449986 0.1898831 0.2396462 0.2203178 0.2477351 0.2633703
#>  [169] 0.2506905 0.2458179 0.2492338 0.2322758 0.2580985 0.1936653 0.2261501
#>  [176] 0.2256007 0.1926091 0.2170511 0.2685038 0.2609024 0.2369658 0.2298329
#>  [183] 0.2463434 0.2507699 0.2763080 0.2687008 0.2131663 0.1920046 0.2144707
#>  [190] 0.2700382 0.2646735 0.2142414 0.2501766 0.2538252 0.2340217 0.2232992
#>  [197] 0.2024514 0.2636620 0.2467327 0.2423631 0.2256502 0.2666223 0.2528390
#>  [204] 0.2375331 0.1887922 0.2538130 0.2533044 0.2138518 0.2361980 0.2525339
#>  [211] 0.2618995 0.2382334 0.2437900 0.2092427 0.2617558 0.2719882 0.2215951
#>  [218] 0.2664218 0.2377286 0.2665092 0.2709392 0.1944337 0.2468014 0.2446990
#>  [225] 0.2639089 0.1958534 0.2138173 0.2454510 0.2314992 0.2354919 0.2698336
#>  [232] 0.2505990 0.2407294 0.2406040 0.2579973 0.2266309 0.2379060 0.2332040
#>  [239] 0.1900852 0.2342267 0.2474287 0.2606507 0.2359168 0.2655836 0.2394696
#>  [246] 0.2686953 0.2567291 0.2369887 0.2200921 0.2750280 0.2212007 0.1781216
#>  [253] 0.2279638 0.2478215 0.2203940 0.2394546 0.2607702 0.2711873 0.2199690
#>  [260] 0.2073437 0.2229010 0.2600934 0.2540317 0.2031347 0.2406302 0.2195473
#>  [267] 0.2046664 0.2229599 0.2200139 0.2384398 0.2773969 0.2724204 0.2531798
#>  [274] 0.2245245 0.2125679 0.2449204 0.1962518 0.2629976 0.2415961 0.2639880
#>  [281] 0.2300033 0.2405135 0.2541149 0.2209019 0.2348073 0.1900590 0.2497659
#>  [288] 0.2406504 0.2319381 0.2343428 0.2477143 0.2538116 0.2439185 0.2621613
#>  [295] 0.1754911 0.2663844 0.1757110 0.2298670 0.2645219 0.2272991 0.2358535
#>  [302] 0.2363432 0.2313452 0.2504786 0.2298867 0.1974287 0.2719779 0.2705033
#>  [309] 0.2628306 0.2034306 0.2610583 0.2536006 0.2492027 0.2349992 0.2007379
#>  [316] 0.2171082 0.1952559 0.2443116 0.2030013 0.2057033 0.2527625 0.2111531
#>  [323] 0.2170376 0.1930507 0.2103610 0.2098205 0.1984913 0.2553983 0.2121417
#>  [330] 0.2329829 0.2402022 0.2327368 0.2187715 0.2770500 0.2604141 0.2753879
#>  [337] 0.2563841 0.2732028 0.2331148 0.2602442 0.2534911 0.2723996 0.2665086
#>  [344] 0.1939977 0.2118747 0.2685728 0.2278417 0.2339261 0.2403282 0.2749553
#>  [351] 0.2727273 0.2650586 0.1828418 0.2442952 0.2215639 0.2666942 0.1699645
#>  [358] 0.2665503 0.2190979 0.2577129 0.2739458 0.2737983 0.2449972 0.2371925
#>  [365] 0.1896682 0.2397679 0.2232858 0.2216739 0.2382345 0.2487280 0.2615830
#>  [372] 0.2319050 0.2261712 0.2356438 0.2193452 0.2490387 0.2513205 0.2014698
#>  [379] 0.2320372 0.2501085 0.2624375 0.2540800 0.2626532 0.2259290 0.2114068
#>  [386] 0.1978790 0.2732251 0.2735358 0.2783572 0.2184553 0.2139746 0.2494243
#>  [393] 0.2633158 0.2120255 0.2088762 0.2229848 0.2073843 0.2674544 0.2448119
#>  [400] 0.2713286 0.2354947 0.1922381 0.1761972 0.2474305 0.2490447 0.2281334
#>  [407] 0.2178435 0.2624027 0.2690241 0.1837349 0.1954889 0.2190351 0.2695643
#>  [414] 0.2327431 0.2091534 0.2718241 0.2717161 0.2438212 0.2561240 0.2244745
#>  [421] 0.1906268 0.2132612 0.2331942 0.2581578 0.2392606 0.2408851 0.2396796
#>  [428] 0.2504195 0.2137967 0.1920153 0.2072506 0.2784055 0.2760776 0.2672792
#>  [435] 0.2414499 0.2267104 0.2135284 0.2107201 0.2450534 0.2688131 0.2595890
#>  [442] 0.2700520 0.2154299 0.2572698 0.2540216 0.2731192 0.2363452 0.2211533
#>  [449] 0.2499807 0.2378720 0.2495249 0.2289830 0.2506333 0.2131067 0.2395686
#>  [456] 0.2435949 0.2391312 0.2673167 0.2432493 0.2712920 0.2113447 0.2752925
#>  [463] 0.2165613 0.2489175 0.2684753 0.2235137 0.2374033 0.2298197 0.2348803
#>  [470] 0.2471045 0.2116346 0.2248793 0.2041018 0.2597848 0.2483988 0.2203553
#>  [477] 0.2705903 0.2593923 0.2604239 0.2190074 0.2120999 0.2390590 0.2272823
#>  [484] 0.2080190 0.2225054 0.2475748 0.1990973 0.2669697 0.2158470 0.2312626
#>  [491] 0.2687073 0.2045761 0.2612341 0.2442856 0.2372274 0.2058988 0.1890680
#>  [498] 0.2752534 0.2681098 0.2496514 0.2302935 0.2638124 0.2314993 0.2263702
#>  [505] 0.2506493 0.2068836 0.2366967 0.2472795 0.2447484 0.2093608 0.2506046
#>  [512] 0.2148314 0.2187571 0.2409063 0.2647198 0.2538128 0.2398738 0.1891735
#>  [519] 0.2399081 0.2409246 0.2710481 0.2418440 0.2675663 0.2544926 0.2299364
#>  [526] 0.2329162 0.2628709 0.1730863 0.2478564 0.2160288 0.2621439 0.2407946
#>  [533] 0.2607574 0.2648504 0.2499135 0.2503174 0.2545530 0.2182222 0.1976522
#>  [540] 0.2468403 0.2655566 0.2285470 0.2305291 0.2386228 0.2604925 0.1980159
#>  [547] 0.2300866 0.2463462 0.2716985 0.2545816 0.2601133 0.2307236 0.2278026
#>  [554] 0.2461271 0.2109294 0.2470695 0.2576563 0.2383038 0.2214477 0.2153015
#>  [561] 0.2517477 0.2410514 0.2433671 0.2333838 0.2663741 0.2104174 0.2612338
#>  [568] 0.2182724 0.2278566 0.2332574 0.2412724 0.2366717 0.2409507 0.2318608
#>  [575] 0.2661196 0.2718184 0.2303475 0.2760209 0.2610680 0.2482882 0.2302735
#>  [582] 0.2171026 0.2420119 0.2490733 0.2466481 0.2688783 0.2255800 0.2515869
#>  [589] 0.2386438 0.2725744 0.2309225 0.2112396 0.2711750 0.2295871 0.2044301
#>  [596] 0.2265575 0.2671516 0.2175889 0.2511825 0.2425810 0.2509699 0.2395652
#>  [603] 0.2730082 0.2327790 0.1796731 0.2150893 0.2137725 0.2610130 0.2756891
#>  [610] 0.2406286 0.2433168 0.2405701 0.2592470 0.2630435 0.2354379 0.2538407
#>  [617] 0.2513559 0.2340858 0.2574016 0.2747405 0.2279786 0.1966728 0.2147518
#>  [624] 0.1810951 0.2125010 0.2248940 0.2342310 0.2599185 0.2051886 0.2572028
#>  [631] 0.2665201 0.2755211 0.2385590 0.2001141 0.2580100 0.2517962 0.2533118
#>  [638] 0.2680234 0.2544352 0.2194755 0.2685320 0.2429479 0.2261920 0.2727925
#>  [645] 0.2721525 0.2488260 0.2613476 0.2348619 0.2587665 0.2688501 0.2749429
#>  [652] 0.2465345 0.2251446 0.2314032 0.2699792 0.1840681 0.2172944 0.2330436
#>  [659] 0.1863262 0.2031079 0.2452443 0.2132723 0.2573409 0.1924806 0.2303902
#>  [666] 0.1992803 0.2015089 0.2011261 0.2615772 0.2405241 0.2156365 0.2232553
#>  [673] 0.2522752 0.2122606 0.2429928 0.2360499 0.2544773 0.2520285 0.2611842
#>  [680] 0.2700861 0.2036011 0.2305846 0.2219057 0.2321449 0.2102030 0.2122409
#>  [687] 0.2336948 0.2009371 0.2504241 0.2683880 0.2227101 0.2201831 0.2515717
#>  [694] 0.2499718 0.2603921 0.2149815 0.2582205 0.2732842 0.2391665 0.2243964
#>  [701] 0.2363483 0.2467998 0.2663409 0.2206740 0.2380461 0.2330762 0.2540353
#>  [708] 0.1765478 0.2231570 0.2155574 0.2218334 0.2004454 0.2662321 0.2026091
#>  [715] 0.2496116 0.2292940 0.2422794 0.2683997 0.2394231 0.2597768 0.2157838
#>  [722] 0.2631987 0.2515007 0.1886778 0.2654333 0.2637506 0.1918410 0.2560513
#>  [729] 0.2317970 0.2065873 0.2542248 0.2347834 0.2087295 0.2302434 0.2378443
#>  [736] 0.2267734 0.2225575 0.2692240 0.1783818 0.2690669 0.2550186 0.2581400
#>  [743] 0.2413657 0.2484065 0.2050099 0.2158605 0.2238422 0.2121499 0.1915203
#>  [750] 0.2367106 0.2591402 0.2722873 0.2480795 0.2427922 0.2226337 0.2270092
#>  [757] 0.2390810 0.2326460 0.2514240 0.2244850 0.2569367 0.2564431 0.2461298
#>  [764] 0.2397539 0.2222900 0.2054494 0.2666953 0.2667583 0.2599952 0.2567324
#>  [771] 0.2665631 0.1897787 0.2257106 0.2668326 0.2697518 0.2629075 0.2283355
#>  [778] 0.2140754 0.2704917 0.2158765 0.2650355 0.2580063 0.2477954 0.2400389
#>  [785] 0.2202717 0.2709300 0.2608721 0.2730306 0.2669799 0.2482475 0.2563268
#>  [792] 0.1974510 0.2539215 0.2200350 0.2355277 0.2600709 0.2504827 0.2427784
#>  [799] 0.2210638 0.2493796 0.2549457 0.2641988 0.2464195 0.2592435 0.2160335
#>  [806] 0.2335155 0.2408624 0.2319130 0.2385086 0.2660430 0.2192760 0.1999973
#>  [813] 0.2277538 0.2375203 0.2510475 0.2502068 0.2132280 0.2280028 0.2213269
#>  [820] 0.2375703 0.2550569 0.2492356 0.2581836 0.2584749 0.2625232 0.2584345
#>  [827] 0.2526289 0.2119847 0.2177525 0.2533134 0.2342378 0.2235325 0.2502645
#>  [834] 0.2630480 0.2276876 0.2555032 0.2211483 0.2666208 0.2340377 0.2172956
#>  [841] 0.2473197 0.2205668 0.1981228 0.2375736 0.2328460 0.2564571 0.2535366
#>  [848] 0.2576328 0.2591485 0.2536856 0.1889084 0.2043281 0.2365895 0.2641022
#>  [855] 0.2267621 0.2319788 0.2395577 0.2192016 0.2153945 0.2112548 0.2277864
#>  [862] 0.2639328 0.2703509 0.2338750 0.2765522 0.2707759 0.2616349 0.2785447
#>  [869] 0.2460487 0.2552532 0.2426550 0.2660615 0.2632837 0.2376789 0.2744700
#>  [876] 0.2557647 0.2418268 0.2401926 0.2232713 0.2190895 0.2440002 0.2361583
#>  [883] 0.2603278 0.2294203 0.2563217 0.2674920 0.2545204 0.2255317 0.2101837
#>  [890] 0.2412140 0.1951017 0.1830535 0.2595970 0.2201024 0.1867365 0.2305266
#>  [897] 0.2391149 0.2616419 0.2080358 0.2400160 0.2148928 0.2430305 0.2521594
#>  [904] 0.2405167 0.2652036 0.2604587 0.2709045 0.2291131 0.2537127 0.2189071
#>  [911] 0.2443493 0.2536010 0.2220306 0.2067846 0.2000625 0.2410507 0.2596589
#>  [918] 0.2327913 0.2225203 0.2235860 0.2205565 0.2307188 0.2201065 0.2237597
#>  [925] 0.2666254 0.2353669 0.1834686 0.2422462 0.2016399 0.2539341 0.2363926
#>  [932] 0.2425799 0.2595582 0.2322260 0.2767808 0.2120716 0.2556749 0.2191170
#>  [939] 0.2575991 0.2224725 0.2161850 0.2583441 0.2572564 0.2305069 0.2216261
#>  [946] 0.2341835 0.2604217 0.2378934 0.2664104 0.2641772 0.2013576 0.2722254
#>  [953] 0.2215492 0.2288837 0.2596156 0.2432767 0.2453813 0.2686086 0.2596775
#>  [960] 0.2513111 0.2372397 0.2484061 0.2346392 0.2637913 0.2308446 0.2296615
#>  [967] 0.2254646 0.2360802 0.2444263 0.1886242 0.2451737 0.2716763 0.2653050
#>  [974] 0.2571026 0.2312676 0.2269507 0.2514367 0.2752055 0.2263090 0.2269578
#>  [981] 0.2615793 0.2300775 0.2068275 0.2536329 0.1683373 0.2127179 0.1876650
#>  [988] 0.2541244 0.2406226 0.2609926 0.2359248 0.2442603 0.2610715 0.2566403
#>  [995] 0.2323452 0.2532542 0.2489761 0.2041796 0.2738534 0.2471878