Mahalanobis Depth

Computes the mahalanobis depth of a point or vectors of points with respect to a multivariate data set.

depthMah(u, X, cov = NULL, mean = NULL, threads = -1)

Arguments

u
Numerical vector or matrix whose depth is to be calculated. Dimension has to be the same as that of the observations.

X
The data as a matrix, data frame or list. If it is a matrix or data frame, then each row is viewed as one multivariate observation. If it is a list, all components must be numerical vectors of equal length (coordinates of observations).

cov
custom covariance matrix passed. If NULL standard calculations will be based on standard covariance estimator.

mean
custom mean vector. If null --- mean average will be used.

threads
number of threads used in parallel computations. Default value -1 means that all possible cores will be used.

…
currently not supported.

Details

Calculation of Mahalanobis depth is exact.

Returns the depth of multivariate point u with respect to data set X.

Examples

x <- matrix(rnorm(9999), nc = 3)
depthMah(x, x)

#> Depth method:  Mahalanobis 
#>    [1] 0.18084195 0.79852954 0.24711008 0.19639983 0.35312025 0.42367538
#>    [7] 0.36748615 0.13621329 0.10803849 0.41565862 0.38060035 0.15927702
#>   [13] 0.23268478 0.41749100 0.43339679 0.21233509 0.10179600 0.36265980
#>   [19] 0.27454256 0.23286013 0.54120752 0.82049527 0.30184103 0.15034464
#>   [25] 0.33315895 0.21198449 0.37807199 0.20057012 0.24276073 0.52984727
#>   [31] 0.46306086 0.67697202 0.33761931 0.20038476 0.68761148 0.32465342
#>   [37] 0.43334816 0.39030253 0.40899031 0.36443944 0.28804153 0.25438195
#>   [43] 0.14571080 0.40919866 0.54284673 0.56226406 0.49841970 0.26390354
#>   [49] 0.68906633 0.25984009 0.30455305 0.26852536 0.18209935 0.05912349
#>   [55] 0.38916212 0.27541453 0.69164293 0.31586340 0.54158134 0.49207791
#>   [61] 0.13757646 0.44670414 0.30914607 0.61402894 0.13253641 0.18781594
#>   [67] 0.52016331 0.54058486 0.45383745 0.21281160 0.35943038 0.40807510
#>   [73] 0.12962147 0.91531049 0.26354333 0.59806563 0.28043709 0.81065184
#>   [79] 0.46878748 0.57738440 0.27045864 0.19358254 0.60935509 0.27476786
#>   [85] 0.48817997 0.13352467 0.25486161 0.37660293 0.81346901 0.57085453
#>   [91] 0.63237731 0.13013199 0.26303417 0.66646429 0.50870768 0.18004079
#>   [97] 0.36416823 0.11863489 0.24830689 0.17937393 0.66787643 0.39864196
#>  [103] 0.40476261 0.42625141 0.44001588 0.48200049 0.08829280 0.41132542
#>  [109] 0.19336434 0.44251468 0.29789434 0.67953535 0.15039173 0.53393572
#>  [115] 0.15225113 0.40621822 0.38319118 0.13010478 0.11386777 0.32476524
#>  [121] 0.38036153 0.25657607 0.27595018 0.22710210 0.18763594 0.31463385
#>  [127] 0.25674238 0.24716969 0.43141537 0.74725126 0.59816452 0.58871040
#>  [133] 0.15494507 0.28279300 0.11281130 0.54936656 0.22040148 0.32795404
#>  [139] 0.25461028 0.62427945 0.63753411 0.54371290 0.36124154 0.47651827
#>  [145] 0.45790828 0.17874099 0.24195731 0.16810837 0.07703809 0.13658658
#>  [151] 0.16388509 0.25057126 0.39184682 0.84345279 0.18256126 0.18725768
#>  [157] 0.13875573 0.80187728 0.55704530 0.34315411 0.24433223 0.41035731
#>  [163] 0.33123555 0.35190604 0.12039937 0.13347659 0.14132396 0.41962630
#>  [169] 0.34468725 0.47268629 0.51987868 0.16937169 0.72114758 0.22159067
#>  [175] 0.16885698 0.25260026 0.14707919 0.22039158 0.24558945 0.38392358
#>  [181] 0.39517262 0.22242537 0.43667047 0.92471192 0.70263911 0.19197751
#>  [187] 0.19024929 0.29421774 0.11743417 0.31897217 0.13861631 0.15052128
#>  [193] 0.15127720 0.33085672 0.50654110 0.41803553 0.73570674 0.10910274
#>  [199] 0.26383616 0.12376145 0.30687769 0.66521594 0.49471801 0.30821638
#>  [205] 0.77140029 0.17646975 0.25932364 0.20352909 0.81704342 0.26651240
#>  [211] 0.07893993 0.26566066 0.30672881 0.59515714 0.23880330 0.20476719
#>  [217] 0.28993737 0.41022159 0.20218149 0.52349120 0.49057889 0.27338816
#>  [223] 0.52597663 0.28660785 0.62799143 0.24729603 0.44345196 0.38087160
#>  [229] 0.20082181 0.55684488 0.18304175 0.35548822 0.41335205 0.16145169
#>  [235] 0.62850374 0.25288392 0.15786469 0.21265848 0.31916646 0.17317103
#>  [241] 0.13736326 0.25857636 0.26219194 0.12593525 0.23936708 0.31350533
#>  [247] 0.77034967 0.18132841 0.37954658 0.17605662 0.29038519 0.11160385
#>  [253] 0.45689925 0.38345948 0.45527675 0.28858826 0.23858525 0.14760104
#>  [259] 0.15079341 0.24995772 0.46257971 0.16565088 0.25626858 0.19614225
#>  [265] 0.35064730 0.21473575 0.33692129 0.65291732 0.83735141 0.19684052
#>  [271] 0.50200332 0.22415072 0.62622176 0.12145956 0.24860193 0.51187872
#>  [277] 0.13481726 0.48996303 0.31155768 0.27534602 0.29375026 0.23181266
#>  [283] 0.34024049 0.31961099 0.22213768 0.72784199 0.46717397 0.24833470
#>  [289] 0.16940897 0.74150638 0.13187809 0.72974577 0.18852526 0.75909901
#>  [295] 0.41891699 0.52462124 0.23726605 0.25378904 0.38299614 0.19333436
#>  [301] 0.65344610 0.25033074 0.23171132 0.50665825 0.33282098 0.36492545
#>  [307] 0.38093719 0.08906179 0.84265808 0.16549787 0.20379697 0.82980565
#>  [313] 0.48326480 0.28752783 0.28907607 0.12261824 0.26922723 0.75729314
#>  [319] 0.36089038 0.25829642 0.33770021 0.56297194 0.33078195 0.16257086
#>  [325] 0.47379123 0.26724562 0.35747823 0.36347004 0.35584888 0.24401619
#>  [331] 0.16240121 0.23674069 0.22530897 0.11878226 0.44588357 0.18637451
#>  [337] 0.85384674 0.22516186 0.25598733 0.29928026 0.16725189 0.18460603
#>  [343] 0.32765125 0.36280478 0.12093849 0.52145665 0.35315152 0.24105032
#>  [349] 0.22056283 0.54456023 0.55971339 0.62874527 0.15265680 0.25444063
#>  [355] 0.37889159 0.50203843 0.16853808 0.21148888 0.79510153 0.21138434
#>  [361] 0.57348248 0.15322184 0.63629159 0.10259486 0.35439940 0.53439768
#>  [367] 0.31328051 0.21699821 0.21963492 0.07128502 0.20292920 0.20983632
#>  [373] 0.24984001 0.67531012 0.10750485 0.12993651 0.15110032 0.32311649
#>  [379] 0.43726936 0.28380941 0.12597834 0.29887917 0.28666585 0.74358071
#>  [385] 0.24911612 0.27778481 0.75203111 0.20851922 0.28396075 0.40712983
#>  [391] 0.17051070 0.24559591 0.18622507 0.15374919 0.47240490 0.18067378
#>  [397] 0.34784899 0.18337785 0.74836861 0.38909207 0.63762351 0.49633896
#>  [403] 0.32992671 0.74806519 0.81184475 0.10735840 0.19295068 0.23521024
#>  [409] 0.16568698 0.25718930 0.62410138 0.32806793 0.21617886 0.25032698
#>  [415] 0.13695446 0.10402489 0.45225876 0.24308234 0.52486577 0.17932692
#>  [421] 0.39023704 0.49167937 0.39600039 0.28545527 0.47786486 0.47930593
#>  [427] 0.50358509 0.47642871 0.38914578 0.72906467 0.14961432 0.50336551
#>  [433] 0.41809711 0.35869757 0.16996154 0.15330947 0.57906120 0.19982515
#>  [439] 0.23558123 0.75351964 0.21516336 0.96704449 0.60030154 0.11664525
#>  [445] 0.32371122 0.12819880 0.24166722 0.26134267 0.15718445 0.18326855
#>  [451] 0.46154537 0.35734002 0.27595829 0.43498717 0.45078022 0.23422587
#>  [457] 0.29308522 0.28228176 0.22888814 0.71848659 0.22425534 0.16362683
#>  [463] 0.17974971 0.11292010 0.88347959 0.42899845 0.14039179 0.18766907
#>  [469] 0.12914786 0.34403972 0.65846305 0.55982040 0.41017956 0.80907027
#>  [475] 0.13692814 0.58412477 0.87560597 0.69869821 0.08653447 0.90790601
#>  [481] 0.28127540 0.29591375 0.24781196 0.29744262 0.13710159 0.16844688
#>  [487] 0.56619847 0.44206351 0.16570142 0.33695926 0.26408856 0.50541717
#>  [493] 0.14637521 0.11643140 0.16413713 0.75723000 0.48462390 0.12019856
#>  [499] 0.30970192 0.39639643 0.22942046 0.18362936 0.19425443 0.34026501
#>  [505] 0.39300290 0.66503654 0.18056537 0.31934740 0.26280204 0.33726694
#>  [511] 0.68544340 0.25694191 0.13312213 0.36856904 0.12588030 0.48351399
#>  [517] 0.66428963 0.29154706 0.19969949 0.56508337 0.20062034 0.28174229
#>  [523] 0.32209192 0.26406362 0.10582753 0.50981206 0.08075718 0.10669798
#>  [529] 0.07375351 0.16289488 0.27114967 0.53353691 0.46429584 0.11244125
#>  [535] 0.51519337 0.17319022 0.38334582 0.40075973 0.25742427 0.15895097
#>  [541] 0.12849406 0.55861046 0.21892368 0.24352873 0.26279467 0.12866813
#>  [547] 0.23548919 0.26644110 0.40910198 0.20753007 0.28180680 0.30288558
#>  [553] 0.16106130 0.08858445 0.86744097 0.53800953 0.25888662 0.15557555
#>  [559] 0.24470129 0.16608945 0.45337103 0.51150863 0.44381756 0.30650337
#>  [565] 0.26230442 0.27008829 0.13721077 0.33046080 0.28515143 0.19157926
#>  [571] 0.20068274 0.37352897 0.23753569 0.13486195 0.42839651 0.23722061
#>  [577] 0.37176171 0.32863084 0.31151841 0.28367819 0.18765244 0.63448974
#>  [583] 0.18037994 0.21276281 0.11588392 0.40704098 0.24843496 0.16446188
#>  [589] 0.15179131 0.16672834 0.27258958 0.63200278 0.56130064 0.35029227
#>  [595] 0.11330630 0.21673732 0.41040559 0.23013462 0.77648841 0.42439250
#>  [601] 0.16573510 0.40190515 0.35641917 0.32279552 0.37636860 0.19474956
#>  [607] 0.26355906 0.24701252 0.47119330 0.24341507 0.30773960 0.58892480
#>  [613] 0.48913887 0.25553522 0.75174524 0.45928191 0.35296387 0.18404115
#>  [619] 0.55365454 0.28290302 0.11221993 0.19265710 0.88497106 0.17908858
#>  [625] 0.54430472 0.33933534 0.58986136 0.43966006 0.37452347 0.39091442
#>  [631] 0.33840601 0.51793409 0.12308438 0.16519382 0.12319687 0.30787824
#>  [637] 0.32769148 0.35650921 0.12958279 0.42775423 0.15757144 0.19178523
#>  [643] 0.06861889 0.09839702 0.16576754 0.15766828 0.65468927 0.18861226
#>  [649] 0.44283060 0.26860989 0.48381083 0.67130102 0.11532580 0.12182654
#>  [655] 0.18204103 0.20357404 0.37134692 0.70946790 0.10494059 0.42453607
#>  [661] 0.15199816 0.19104088 0.34282665 0.28724179 0.31186188 0.44464590
#>  [667] 0.56206608 0.31735723 0.18346853 0.28042761 0.24159781 0.59734715
#>  [673] 0.31427928 0.32258338 0.46311529 0.35820992 0.40350350 0.63129068
#>  [679] 0.13886801 0.47648395 0.18631803 0.18255299 0.38228705 0.34398922
#>  [685] 0.42084543 0.18301106 0.18041465 0.69361135 0.14697094 0.42051560
#>  [691] 0.56365008 0.23112719 0.27088198 0.28105359 0.37222338 0.17876680
#>  [697] 0.34038140 0.32623106 0.14743668 0.59649952 0.16332280 0.39797160
#>  [703] 0.37362070 0.25337887 0.17197608 0.23526130 0.26974373 0.82511596
#>  [709] 0.27616857 0.40957148 0.28630055 0.36962368 0.20444931 0.23833283
#>  [715] 0.54391964 0.37719337 0.15803299 0.23813814 0.36295795 0.46016738
#>  [721] 0.44876171 0.38102092 0.56040059 0.10498067 0.16395691 0.28554963
#>  [727] 0.42817066 0.30326366 0.40902031 0.32579674 0.46567068 0.07179346
#>  [733] 0.53982106 0.22950465 0.55979110 0.11665740 0.26790330 0.62446726
#>  [739] 0.33959133 0.21919057 0.51882571 0.16348393 0.38424396 0.14196197
#>  [745] 0.76403608 0.39417470 0.55632586 0.21756776 0.23614161 0.51250660
#>  [751] 0.74459186 0.22002173 0.23091633 0.16032457 0.22739191 0.27883850
#>  [757] 0.09398597 0.50695523 0.13863367 0.49414468 0.43016177 0.28760628
#>  [763] 0.31739301 0.33053370 0.25602768 0.43496521 0.43490863 0.20044561
#>  [769] 0.20852216 0.26101984 0.47814497 0.35202164 0.43114861 0.65397797
#>  [775] 0.50605970 0.16938306 0.46081187 0.21975741 0.18757359 0.67489874
#>  [781] 0.34367896 0.93062587 0.38426823 0.33697205 0.67407328 0.13938147
#>  [787] 0.12620037 0.43260918 0.67466495 0.25662566 0.72738333 0.83749447
#>  [793] 0.64139202 0.54132141 0.16821338 0.16018685 0.47851641 0.25117833
#>  [799] 0.43222880 0.58276776 0.12649175 0.71920624 0.53457504 0.15511566
#>  [805] 0.31823372 0.19363003 0.22876294 0.60277778 0.55515591 0.67633570
#>  [811] 0.28911370 0.30052544 0.35103827 0.36652674 0.16954378 0.44343353
#>  [817] 0.44010383 0.15481456 0.21957754 0.32298080 0.14086367 0.14592999
#>  [823] 0.59004677 0.57803426 0.28435567 0.26068486 0.15545825 0.29282897
#>  [829] 0.31168887 0.12345200 0.70812609 0.94261744 0.17449864 0.18156672
#>  [835] 0.27164480 0.56080648 0.31120811 0.22943915 0.24015666 0.53948770
#>  [841] 0.28462636 0.39672447 0.64527255 0.19378048 0.23961513 0.44786845
#>  [847] 0.16220140 0.06691236 0.91420578 0.44875885 0.51315795 0.29733282
#>  [853] 0.44287866 0.31651184 0.14116607 0.35989274 0.36883612 0.22221892
#>  [859] 0.20172735 0.50575592 0.17684004 0.37811332 0.62371074 0.75363379
#>  [865] 0.26600609 0.43672041 0.17765985 0.91856812 0.33050332 0.24108757
#>  [871] 0.09310812 0.16386395 0.13829317 0.17326158 0.19035429 0.53744585
#>  [877] 0.17789986 0.40351778 0.12592346 0.31245022 0.32732174 0.22945245
#>  [883] 0.35381871 0.51334182 0.73509802 0.32441737 0.16970624 0.16434905
#>  [889] 0.24349163 0.52553876 0.30610282 0.37391557 0.28677350 0.51799446
#>  [895] 0.08390979 0.72482763 0.76766887 0.51403874 0.32981327 0.11937875
#>  [901] 0.11930908 0.51424759 0.38832184 0.21720690 0.34263938 0.18554002
#>  [907] 0.25880897 0.82709689 0.25014912 0.46182438 0.49879278 0.09444607
#>  [913] 0.79741489 0.19861480 0.17838773 0.22044235 0.16483143 0.33868739
#>  [919] 0.27411024 0.36220830 0.70295708 0.41347949 0.15882824 0.54679269
#>  [925] 0.08743656 0.26090338 0.51300543 0.27625823 0.27595248 0.23643893
#>  [931] 0.32852976 0.36774439 0.30369852 0.17644793 0.70665199 0.72566913
#>  [937] 0.11091621 0.31103875 0.15275278 0.58721029 0.25606696 0.20985224
#>  [943] 0.35614564 0.25230638 0.21206005 0.45819282 0.42516672 0.39491178
#>  [949] 0.58468606 0.53909222 0.43748993 0.54870337 0.13373085 0.38811448
#>  [955] 0.23809083 0.27964489 0.40351797 0.48280628 0.23293477 0.24476502
#>  [961] 0.25747101 0.25490514 0.18707046 0.24404890 0.60175417 0.24984994
#>  [967] 0.34799059 0.59244053 0.30004540 0.06619681 0.52918964 0.19153244
#>  [973] 0.40190973 0.34094241 0.17533926 0.42666459 0.47221977 0.81864287
#>  [979] 0.65729082 0.25321237 0.27899734 0.09825110 0.50501397 0.67711326
#>  [985] 0.44499893 0.07171115 0.91176473 0.22581756 0.08439314 0.12615183
#>  [991] 0.63896371 0.55835214 0.30132960 0.60487879 0.13950619 0.44740120
#>  [997] 0.21304350 0.50352482 0.14294413 0.41385031 0.10893644 0.21777269
#> [1003] 0.30111292 0.13663352 0.50825559 0.20368227 0.43070212 0.19642522
#> [1009] 0.56228987 0.25670740 0.40220073 0.09379879 0.45116777 0.13299932
#> [1015] 0.69413548 0.26105952 0.18605438 0.39791500 0.08066371 0.20810037
#> [1021] 0.33310125 0.18306696 0.33952890 0.28781501 0.11190564 0.14479565
#> [1027] 0.19956262 0.46257816 0.16185322 0.30921315 0.12421654 0.19878386
#> [1033] 0.57269063 0.40417206 0.33289765 0.19676993 0.20255636 0.07077564
#> [1039] 0.48989909 0.37793365 0.33114361 0.53994130 0.19992835 0.24452298
#> [1045] 0.29426745 0.55024587 0.56669056 0.54619895 0.21182045 0.16711570
#> [1051] 0.25131253 0.16170338 0.19361096 0.23973862 0.43752254 0.60319705
#> [1057] 0.18742741 0.32429986 0.31847697 0.23801037 0.11683073 0.50702281
#> [1063] 0.16865465 0.96505417 0.79274095 0.21723863 0.70510538 0.61436736
#> [1069] 0.18563654 0.26101361 0.35833695 0.10855304 0.40095561 0.11201614
#> [1075] 0.14789979 0.16496606 0.89312065 0.28275182 0.21050912 0.32392614
#> [1081] 0.29329651 0.17178898 0.28383132 0.33248618 0.19423777 0.22801924
#> [1087] 0.48577485 0.17265356 0.15119271 0.10569187 0.65304439 0.44533644
#> [1093] 0.24565504 0.40063480 0.26319181 0.36747929 0.17709239 0.50822665
#> [1099] 0.21897685 0.20052043 0.75845684 0.47003091 0.53053325 0.39164148
#> [1105] 0.18849372 0.28449545 0.39793837 0.59452328 0.26819374 0.43941029
#> [1111] 0.14571757 0.65394745 0.25650967 0.28814168 0.33715341 0.91907035
#> [1117] 0.30134225 0.61459216 0.57733332 0.24400112 0.26477508 0.23584098
#> [1123] 0.23739840 0.17252149 0.34519944 0.38561959 0.18916611 0.27436609
#> [1129] 0.58710889 0.28522999 0.15342350 0.37417336 0.55125804 0.16123544
#> [1135] 0.15431409 0.57616518 0.15570902 0.72530346 0.28557720 0.18240118
#> [1141] 0.17392740 0.11589858 0.45289994 0.32389788 0.42095117 0.37795247
#> [1147] 0.40268433 0.30013413 0.30709650 0.11526103 0.28073704 0.25458023
#> [1153] 0.33112726 0.20744846 0.22621742 0.16205284 0.41939191 0.22117601
#> [1159] 0.36817182 0.19638765 0.27128726 0.24339264 0.42176427 0.16504097
#> [1165] 0.14480749 0.23371999 0.26401630 0.11996620 0.20421479 0.38019994
#> [1171] 0.11492010 0.13566664 0.36855357 0.30207748 0.78916651 0.56477401
#> [1177] 0.54251134 0.62044214 0.14652451 0.22332488 0.35626295 0.70857934
#> [1183] 0.18478472 0.20293096 0.32664789 0.40965279 0.21783309 0.37193314
#> [1189] 0.74293478 0.46537389 0.26465239 0.18306139 0.31027890 0.07984715
#> [1195] 0.53839404 0.15256953 0.19470203 0.63183406 0.21285980 0.24810660
#> [1201] 0.41224297 0.29700255 0.10178285 0.48875780 0.78968450 0.90280927
#> [1207] 0.29840352 0.71712581 0.24313802 0.71149939 0.84214164 0.17900728
#> [1213] 0.21070456 0.17391966 0.49469753 0.16862461 0.78424419 0.26343703
#> [1219] 0.53595011 0.12641349 0.11297337 0.40406485 0.27436395 0.70385813
#> [1225] 0.83380597 0.24416639 0.38312654 0.50086050 0.18199505 0.49254542
#> [1231] 0.54847336 0.29526312 0.42818319 0.18869946 0.64688537 0.38229826
#> [1237] 0.24801706 0.18393353 0.38109139 0.35624935 0.15137181 0.26425281
#> [1243] 0.27666995 0.15118341 0.15452689 0.45724220 0.20455184 0.27029572
#> [1249] 0.93116110 0.31951663 0.63366186 0.54777769 0.49157947 0.25146474
#> [1255] 0.50527232 0.21277422 0.46127330 0.72591214 0.88369372 0.80542150
#> [1261] 0.06690012 0.34593658 0.59854656 0.16199189 0.78440310 0.14186634
#> [1267] 0.32446819 0.47317370 0.37199559 0.49540676 0.36733590 0.19263503
#> [1273] 0.32197974 0.21640168 0.20901002 0.38330482 0.08157534 0.32412575
#> [1279] 0.53405483 0.19571658 0.25692529 0.36693336 0.55664232 0.73952833
#> [1285] 0.20975679 0.26773245 0.88215354 0.28663396 0.45472724 0.36579240
#> [1291] 0.74266554 0.48699265 0.41089651 0.29569550 0.25813121 0.39586828
#> [1297] 0.87711662 0.26189221 0.11175953 0.59374517 0.35142931 0.15841178
#> [1303] 0.13824004 0.32215855 0.35113479 0.08864401 0.18071740 0.35491429
#> [1309] 0.25848513 0.34214401 0.14912684 0.77276854 0.92970342 0.20146621
#> [1315] 0.12929490 0.47269785 0.12081271 0.35046206 0.20636648 0.22916460
#> [1321] 0.57822067 0.38010640 0.48013856 0.27955526 0.46422972 0.26475301
#> [1327] 0.30023362 0.14640478 0.51232597 0.14941440 0.21382864 0.22987571
#> [1333] 0.18242510 0.38085968 0.24442892 0.39526806 0.43801008 0.56708540
#> [1339] 0.19725159 0.28147287 0.20704682 0.18955463 0.39362038 0.08205648
#> [1345] 0.27542804 0.43029323 0.15404739 0.36294453 0.21762682 0.19568306
#> [1351] 0.22683187 0.94216252 0.78949446 0.69847226 0.12709499 0.33320341
#> [1357] 0.66104198 0.35991430 0.28365281 0.30178118 0.16785228 0.27639708
#> [1363] 0.14302381 0.36548526 0.63618485 0.25731607 0.24960522 0.08780466
#> [1369] 0.31744364 0.08754982 0.44804584 0.09729138 0.27200378 0.47594422
#> [1375] 0.55738928 0.11779191 0.16971387 0.43868729 0.24310960 0.17459139
#> [1381] 0.40261072 0.11633303 0.41675238 0.16243426 0.69892189 0.49922600
#> [1387] 0.11897717 0.26354174 0.49575177 0.13806088 0.67906157 0.35917908
#> [1393] 0.23217031 0.07767205 0.31561998 0.69089111 0.14756503 0.26726039
#> [1399] 0.11974921 0.77371553 0.64413195 0.26406870 0.45563449 0.28656692
#> [1405] 0.46003752 0.42646316 0.89094241 0.42880487 0.12579373 0.39046220
#> [1411] 0.21679584 0.12569794 0.61866662 0.65690201 0.08541098 0.50222512
#> [1417] 0.35257457 0.14832100 0.36517175 0.30298461 0.43060168 0.26051099
#> [1423] 0.12872861 0.28882686 0.22721867 0.32794784 0.14340420 0.22464867
#> [1429] 0.18291210 0.23718758 0.49597723 0.27059322 0.43917704 0.76170754
#> [1435] 0.14185274 0.08967971 0.35057787 0.25921399 0.33790856 0.45565162
#> [1441] 0.22371594 0.21978276 0.17233765 0.56494389 0.14201874 0.39022416
#> [1447] 0.43624682 0.22487531 0.34954272 0.50159825 0.18730630 0.57859886
#> [1453] 0.71837795 0.10344262 0.52149370 0.36743991 0.58520164 0.15904339
#> [1459] 0.18836976 0.09456708 0.54021751 0.15688754 0.58737477 0.19132955
#> [1465] 0.24173620 0.36432697 0.24337954 0.18282345 0.44737765 0.28396958
#> [1471] 0.25169476 0.22679489 0.43430945 0.46451202 0.24890732 0.22816672
#> [1477] 0.43080348 0.35646505 0.82161438 0.16157220 0.28348241 0.27933413
#> [1483] 0.22454033 0.43745143 0.64123331 0.18751645 0.20325532 0.39284393
#> [1489] 0.47400548 0.10134143 0.77138401 0.69636750 0.53041268 0.14808622
#> [1495] 0.19825597 0.54770878 0.18248748 0.08996142 0.24139454 0.97687501
#> [1501] 0.22252904 0.30528573 0.24174993 0.21257372 0.30873726 0.36788003
#> [1507] 0.21821415 0.51622135 0.32133114 0.21680869 0.17832241 0.52748772
#> [1513] 0.49434107 0.15383110 0.49751604 0.88273583 0.19371065 0.31072081
#> [1519] 0.16658114 0.67478017 0.60695617 0.20318477 0.18232887 0.20805929
#> [1525] 0.39805122 0.15550732 0.27126227 0.52931021 0.80068762 0.58032906
#> [1531] 0.42762946 0.26983489 0.24941334 0.49810303 0.23167160 0.28161772
#> [1537] 0.19648112 0.27083093 0.36013519 0.12277616 0.14319628 0.30780125
#> [1543] 0.33131331 0.26279367 0.15433484 0.13559604 0.08562786 0.35745795
#> [1549] 0.33892172 0.18909764 0.64081349 0.24884283 0.43837476 0.21055668
#> [1555] 0.25967109 0.34106246 0.34526257 0.25800402 0.34426802 0.43154303
#> [1561] 0.56991202 0.19035223 0.29940652 0.49707807 0.62582680 0.33786303
#> [1567] 0.56970443 0.15180540 0.42472769 0.52231350 0.18160260 0.42677213
#> [1573] 0.36290310 0.31391443 0.19072816 0.24792296 0.12977002 0.19284801
#> [1579] 0.11597289 0.22125034 0.46941212 0.35658037 0.20288930 0.38065215
#> [1585] 0.40128719 0.25245961 0.94915862 0.50463929 0.44663220 0.19806431
#> [1591] 0.40328306 0.32387386 0.13851339 0.53946307 0.26362997 0.50827379
#> [1597] 0.75148486 0.23645502 0.16835477 0.18351926 0.14330298 0.08742371
#> [1603] 0.34187093 0.26773685 0.74467950 0.23996170 0.20511931 0.19308460
#> [1609] 0.11190658 0.16664631 0.62620691 0.24093018 0.72988372 0.16082017
#> [1615] 0.18719520 0.81495492 0.16288827 0.19041654 0.32028587 0.26051758
#> [1621] 0.10904065 0.57820789 0.17641082 0.26403949 0.14086293 0.67058068
#> [1627] 0.15226726 0.39602936 0.18134606 0.34873961 0.49277995 0.25776761
#> [1633] 0.24928929 0.25474879 0.20098681 0.19982013 0.09689179 0.11766501
#> [1639] 0.13176336 0.42226184 0.38397504 0.65802638 0.29066559 0.41963132
#> [1645] 0.22066088 0.10839897 0.25264108 0.37470682 0.25800832 0.76887208
#> [1651] 0.08679790 0.54910045 0.13551434 0.13279777 0.11913500 0.29670141
#> [1657] 0.55149734 0.27809442 0.10682501 0.49976102 0.30388889 0.43853163
#> [1663] 0.31278749 0.10408707 0.38653742 0.18511333 0.29187087 0.19069604
#> [1669] 0.61293603 0.14065963 0.30770260 0.24125851 0.25713412 0.23573507
#> [1675] 0.32626558 0.22942453 0.95254845 0.29201127 0.51439506 0.40910814
#> [1681] 0.21587741 0.50603231 0.19713349 0.35319369 0.27305305 0.29101850
#> [1687] 0.14229067 0.22183198 0.14587349 0.45639650 0.15804918 0.95861924
#> [1693] 0.15729375 0.48966874 0.30905857 0.18352157 0.16349231 0.16057710
#> [1699] 0.28876457 0.60715465 0.20120080 0.66432046 0.50763446 0.31373313
#> [1705] 0.36156579 0.08657799 0.21998447 0.70706982 0.56557106 0.15727357
#> [1711] 0.57173212 0.25197313 0.36774441 0.65447740 0.24608925 0.45811848
#> [1717] 0.13797999 0.29045192 0.20635864 0.64544497 0.18375046 0.85924490
#> [1723] 0.32627885 0.51578872 0.43746945 0.22884428 0.31938718 0.15781085
#> [1729] 0.50128550 0.40698098 0.17805298 0.28275886 0.30865756 0.12210165
#> [1735] 0.30990423 0.09996127 0.49315436 0.91207994 0.55774128 0.09652855
#> [1741] 0.29750933 0.62275437 0.20180086 0.30027233 0.25334394 0.70070323
#> [1747] 0.43102788 0.20653709 0.24505319 0.10249004 0.21538984 0.29702483
#> [1753] 0.25743526 0.21159822 0.12425699 0.28533825 0.57806121 0.43431372
#> [1759] 0.57646879 0.55969776 0.27412012 0.09386951 0.27900632 0.66447308
#> [1765] 0.55512350 0.76895947 0.13541064 0.37898562 0.40332030 0.29277517
#> [1771] 0.28328965 0.74832590 0.13811675 0.44191213 0.47183929 0.24840107
#> [1777] 0.08593128 0.37108708 0.33474245 0.15057150 0.20613222 0.75971568
#> [1783] 0.30482569 0.55565852 0.21522429 0.20877780 0.31690186 0.22535989
#> [1789] 0.29560066 0.73193297 0.48352398 0.15605867 0.39623513 0.36912032
#> [1795] 0.13926066 0.11340089 0.43809789 0.62973789 0.22643829 0.30007864
#> [1801] 0.11008319 0.34103008 0.15759628 0.28391994 0.32511480 0.72909502
#> [1807] 0.22483750 0.31860824 0.44076971 0.14689499 0.86671161 0.13091494
#> [1813] 0.17034975 0.10776350 0.46438795 0.17884285 0.27367437 0.62283359
#> [1819] 0.15538542 0.17210613 0.24292104 0.37770145 0.39832722 0.25489399
#> [1825] 0.30212574 0.41737609 0.26236239 0.59692097 0.27982632 0.66221354
#> [1831] 0.11416033 0.14353287 0.15636084 0.31015529 0.24345910 0.21848323
#> [1837] 0.25461229 0.35617628 0.21541622 0.33601291 0.10509971 0.41679456
#> [1843] 0.16371709 0.19460922 0.16269401 0.26516720 0.35324569 0.17710802
#> [1849] 0.22741999 0.11161241 0.60493814 0.40809214 0.31449353 0.23023363
#> [1855] 0.40590262 0.35865069 0.50989006 0.29493521 0.45633286 0.58843135
#> [1861] 0.41356974 0.15321190 0.17369326 0.51928468 0.17844172 0.31512871
#> [1867] 0.81567522 0.21276338 0.31767110 0.22369509 0.25462519 0.78064364
#> [1873] 0.20960485 0.35979076 0.23277201 0.11057075 0.26482176 0.29367379
#> [1879] 0.80951889 0.54656094 0.43490874 0.33616674 0.77970134 0.15213969
#> [1885] 0.11070453 0.06510178 0.27245857 0.22375330 0.89608381 0.35092885
#> [1891] 0.50996836 0.38907456 0.45802457 0.36844920 0.15404731 0.30770659
#> [1897] 0.11538641 0.10124952 0.32889462 0.50584570 0.41341771 0.45617905
#> [1903] 0.12881039 0.52333583 0.92822350 0.22650642 0.13391568 0.22697078
#> [1909] 0.26332061 0.48194872 0.23957489 0.14303148 0.73859006 0.26294377
#> [1915] 0.25975602 0.65121028 0.19669218 0.09596177 0.21561255 0.58799705
#> [1921] 0.10833811 0.31677881 0.54978180 0.37651020 0.97267467 0.24432521
#> [1927] 0.28455787 0.21881539 0.08495916 0.29351760 0.36605771 0.12815204
#> [1933] 0.71868011 0.76634335 0.15357211 0.25878475 0.22888166 0.48632229
#> [1939] 0.49520220 0.43549714 0.30813826 0.56236435 0.25371099 0.60999147
#> [1945] 0.09565868 0.38812764 0.79685191 0.35941038 0.25607335 0.40907919
#> [1951] 0.27536859 0.57546047 0.53853332 0.17913237 0.20597091 0.27784784
#> [1957] 0.17935219 0.28614793 0.30981006 0.16542165 0.11799586 0.35548883
#> [1963] 0.67894101 0.38231400 0.24877186 0.65457578 0.19592990 0.60630194
#> [1969] 0.27221237 0.48443300 0.15172314 0.37400639 0.27796070 0.27996943
#> [1975] 0.17922542 0.20088703 0.46404751 0.19875780 0.27038619 0.49525508
#> [1981] 0.43347872 0.26267526 0.45317847 0.05703926 0.53579923 0.34414953
#> [1987] 0.27918839 0.31198303 0.38081447 0.32437874 0.34956126 0.29948579
#> [1993] 0.54945087 0.61003718 0.16912254 0.22799992 0.30237108 0.25893914
#> [1999] 0.23057053 0.22161104 0.11364030 0.37172939 0.41915950 0.17499233
#> [2005] 0.30127277 0.26726190 0.60773140 0.22849789 0.72312758 0.22808273
#> [2011] 0.70348330 0.71888547 0.65646041 0.21929802 0.22279205 0.24498196
#> [2017] 0.96169052 0.14638383 0.29299536 0.15528709 0.53238922 0.39665025
#> [2023] 0.25015146 0.63794788 0.34966060 0.13922723 0.10748697 0.13947858
#> [2029] 0.26512926 0.45761691 0.12718773 0.33509113 0.25702741 0.21049278
#> [2035] 0.62432675 0.36830737 0.32552047 0.19663084 0.53343631 0.41388142
#> [2041] 0.30169566 0.26795795 0.30650224 0.43667319 0.39811112 0.25060505
#> [2047] 0.19646140 0.52153459 0.56045771 0.68148652 0.24757497 0.16978562
#> [2053] 0.17814589 0.20920098 0.32314309 0.18899478 0.16769029 0.51596691
#> [2059] 0.38625975 0.16372691 0.50556905 0.14043467 0.91945022 0.20679268
#> [2065] 0.28562530 0.70318613 0.33085094 0.54701743 0.08612874 0.57879777
#> [2071] 0.36415228 0.69724660 0.12143307 0.20494100 0.20814148 0.80073509
#> [2077] 0.25628091 0.16300971 0.23651965 0.25211207 0.84868277 0.30899738
#> [2083] 0.32476952 0.26908703 0.34684395 0.25094212 0.36730549 0.17755541
#> [2089] 0.42141149 0.30232700 0.51049158 0.77392473 0.30125633 0.38379465
#> [2095] 0.36989385 0.18084440 0.24534231 0.58020111 0.48588051 0.32093104
#> [2101] 0.20912391 0.48536301 0.18548066 0.83133458 0.36699306 0.37220212
#> [2107] 0.22388081 0.46737239 0.55496293 0.23549603 0.34879358 0.13593991
#> [2113] 0.21092298 0.38134925 0.52821779 0.52414706 0.15491356 0.21920158
#> [2119] 0.54223501 0.12731644 0.86350260 0.45394617 0.20263238 0.16578609
#> [2125] 0.09250419 0.43484350 0.29496602 0.83095237 0.38903942 0.30418213
#> [2131] 0.43006432 0.62014896 0.22705481 0.51215242 0.14900398 0.51135050
#> [2137] 0.14757540 0.09748424 0.53255711 0.43612294 0.18652139 0.36534662
#> [2143] 0.23052303 0.56478370 0.18723473 0.41886085 0.30346250 0.18323530
#> [2149] 0.91305099 0.24864948 0.08362483 0.15302460 0.20400637 0.21481992
#> [2155] 0.24945841 0.90194721 0.64516732 0.10238758 0.40234136 0.12233968
#> [2161] 0.15939364 0.60062756 0.52998987 0.14578975 0.50081395 0.28800454
#> [2167] 0.74984377 0.25222420 0.50377471 0.18098735 0.28178811 0.41816833
#> [2173] 0.27897310 0.35214852 0.12509785 0.55355909 0.09587431 0.08813164
#> [2179] 0.32658945 0.37396006 0.12146857 0.19950087 0.21977255 0.62570495
#> [2185] 0.36412544 0.34426124 0.75358228 0.30141893 0.42641732 0.43801437
#> [2191] 0.26930484 0.20984350 0.43220026 0.43743028 0.09884306 0.10920089
#> [2197] 0.73458407 0.52064356 0.52914100 0.20414131 0.47223424 0.10408336
#> [2203] 0.40872002 0.22463484 0.13001441 0.34669883 0.15429115 0.51396728
#> [2209] 0.16841004 0.42848407 0.26555110 0.38652211 0.35319652 0.16793266
#> [2215] 0.57491912 0.93646440 0.67433014 0.25797768 0.18462554 0.24945484
#> [2221] 0.23540117 0.68707569 0.30089384 0.30268887 0.21578204 0.32214845
#> [2227] 0.13770112 0.77218925 0.09943232 0.32682309 0.41059696 0.17488678
#> [2233] 0.23102167 0.49141692 0.08559123 0.32234868 0.17764867 0.16066204
#> [2239] 0.27282279 0.29024504 0.17513626 0.71254760 0.22953343 0.33405297
#> [2245] 0.09173056 0.32255847 0.28615400 0.26270274 0.55228988 0.21749738
#> [2251] 0.52103541 0.74152223 0.22847532 0.15543829 0.20480664 0.72561035
#> [2257] 0.33242290 0.30600372 0.28985531 0.18618367 0.33245246 0.12334722
#> [2263] 0.43417938 0.16662522 0.15444788 0.27744085 0.45444804 0.60011832
#> [2269] 0.44131660 0.21410165 0.24817819 0.32983516 0.22101828 0.22895516
#> [2275] 0.13514886 0.08614838 0.35735141 0.16173205 0.50101679 0.18256566
#> [2281] 0.22155425 0.52818562 0.17925882 0.23203950 0.40334877 0.50097271
#> [2287] 0.33016167 0.32480724 0.17683943 0.24874641 0.18763487 0.18052477
#> [2293] 0.34268132 0.31889609 0.45076240 0.16208421 0.26912842 0.22728728
#> [2299] 0.11059385 0.12973770 0.23863605 0.07178308 0.54429651 0.17747075
#> [2305] 0.14216551 0.56008262 0.28899552 0.30067759 0.37039704 0.53257597
#> [2311] 0.24937267 0.08672160 0.16786894 0.38058396 0.84821200 0.31137908
#> [2317] 0.34744487 0.26117508 0.45921793 0.20567682 0.62410275 0.14680463
#> [2323] 0.26072280 0.32095857 0.30862551 0.28954279 0.09073110 0.69202748
#> [2329] 0.20739228 0.49653832 0.34074611 0.64175786 0.20565866 0.27871678
#> [2335] 0.73946989 0.38528197 0.20169833 0.69853146 0.23509630 0.41557021
#> [2341] 0.31521700 0.16458052 0.26123173 0.57786952 0.48620272 0.13750841
#> [2347] 0.19663144 0.71392604 0.17151100 0.43669220 0.23018382 0.20138816
#> [2353] 0.41798874 0.30791307 0.78627643 0.16210510 0.24541422 0.71896489
#> [2359] 0.47437656 0.85356333 0.16965728 0.32818523 0.50961292 0.46187776
#> [2365] 0.11231622 0.29571289 0.35563780 0.10460408 0.17197645 0.24684593
#> [2371] 0.45543997 0.51397088 0.09319416 0.14276221 0.13322375 0.24983733
#> [2377] 0.15423138 0.17952966 0.22701121 0.12707455 0.46340945 0.21898150
#> [2383] 0.28703606 0.38482514 0.26563836 0.34176851 0.20107213 0.66553748
#> [2389] 0.57877007 0.36762195 0.25666350 0.15326265 0.16213655 0.38307722
#> [2395] 0.39090440 0.36160751 0.34319528 0.20157890 0.13831799 0.19536373
#> [2401] 0.31911479 0.24450827 0.27634235 0.58497081 0.20984293 0.36701049
#> [2407] 0.26814115 0.32154894 0.22469784 0.10154309 0.20130548 0.48484325
#> [2413] 0.54462210 0.69107480 0.49256114 0.23692238 0.18163894 0.86844606
#> [2419] 0.29297929 0.21954893 0.23398205 0.20086808 0.23171876 0.85056578
#> [2425] 0.37395399 0.55234321 0.53897954 0.17681271 0.08235244 0.14769522
#> [2431] 0.13986605 0.27149741 0.22203917 0.52076851 0.06061442 0.43935468
#> [2437] 0.13828645 0.24656771 0.50348532 0.50286074 0.31000491 0.78316755
#> [2443] 0.25091127 0.11172549 0.18234366 0.30856848 0.24629022 0.63082376
#> [2449] 0.13884225 0.35982564 0.26380514 0.21939957 0.31091161 0.06134697
#> [2455] 0.69500533 0.60583645 0.30533980 0.76961983 0.09684964 0.08475429
#> [2461] 0.15659691 0.26363009 0.13202809 0.21917705 0.19037768 0.29099940
#> [2467] 0.49379644 0.34144117 0.39637899 0.31129550 0.21920089 0.52146356
#> [2473] 0.54427372 0.22919249 0.54554086 0.14893543 0.33074260 0.08136315
#> [2479] 0.33130412 0.20885296 0.11469215 0.27786639 0.16464574 0.37669682
#> [2485] 0.27135714 0.30247551 0.42340922 0.41690891 0.24318479 0.56331024
#> [2491] 0.38976673 0.12804621 0.11959539 0.21290160 0.36236119 0.47222626
#> [2497] 0.17594854 0.26550353 0.19838216 0.16391600 0.58430032 0.47327645
#> [2503] 0.16050149 0.28444301 0.60361102 0.09830743 0.66426307 0.13633578
#> [2509] 0.25805684 0.17520974 0.50403599 0.17389041 0.28037870 0.18004288
#> [2515] 0.32754255 0.22261125 0.16505525 0.22033173 0.27165599 0.28192950
#> [2521] 0.19496563 0.34212797 0.71022014 0.32746212 0.20443774 0.15270282
#> [2527] 0.14962771 0.46178014 0.54664377 0.40132984 0.26029112 0.40070291
#> [2533] 0.23365884 0.29482917 0.17162930 0.20072068 0.12335297 0.16784416
#> [2539] 0.29950731 0.34591921 0.39042275 0.32877803 0.23283954 0.21863595
#> [2545] 0.25033662 0.22745582 0.48416557 0.61822589 0.20351976 0.37302218
#> [2551] 0.67835526 0.28652983 0.25537038 0.21193165 0.52575828 0.20370358
#> [2557] 0.87920398 0.17054283 0.44755922 0.85935721 0.55467813 0.21141327
#> [2563] 0.52371518 0.31996837 0.15059876 0.11189777 0.13069851 0.24659058
#> [2569] 0.32663180 0.32640644 0.69632601 0.19821960 0.08460306 0.17522632
#> [2575] 0.22365922 0.20137501 0.22236200 0.21405359 0.07908296 0.23045269
#> [2581] 0.38877309 0.43653226 0.33439506 0.34431818 0.30498116 0.24726524
#> [2587] 0.43312073 0.90353882 0.32687562 0.10950708 0.19253093 0.38965049
#> [2593] 0.31307821 0.07709593 0.19161981 0.69692217 0.21123007 0.32356551
#> [2599] 0.62193478 0.36703196 0.67758874 0.25336229 0.18431335 0.19754758
#> [2605] 0.56353493 0.35270256 0.49323100 0.30432263 0.16177161 0.13095716
#> [2611] 0.14924416 0.24462506 0.49675657 0.27604823 0.47375298 0.18146787
#> [2617] 0.34706593 0.77187275 0.29165883 0.33003271 0.67104487 0.28895970
#> [2623] 0.11738859 0.23993986 0.32640204 0.82758121 0.52126870 0.34358190
#> [2629] 0.50199294 0.08963237 0.32877722 0.22774108 0.75544620 0.25212262
#> [2635] 0.58626724 0.11056140 0.34447428 0.67547957 0.11403437 0.22184622
#> [2641] 0.43475409 0.54106271 0.43221300 0.34007250 0.79224621 0.18226252
#> [2647] 0.13762985 0.39773930 0.36462220 0.41976073 0.22111810 0.19913905
#> [2653] 0.23526733 0.21910446 0.24127146 0.24623407 0.43510297 0.10165050
#> [2659] 0.50586418 0.16930233 0.47625842 0.20235906 0.32343660 0.15182855
#> [2665] 0.49463052 0.23488991 0.12048166 0.57854250 0.28424293 0.10965859
#> [2671] 0.26879097 0.82103798 0.90284685 0.42396197 0.17155025 0.18402136
#> [2677] 0.28643202 0.24340329 0.61724504 0.32005415 0.40594602 0.12364099
#> [2683] 0.22291956 0.36418210 0.41253991 0.24235642 0.61421202 0.23099725
#> [2689] 0.74122941 0.12038116 0.38402988 0.62948627 0.19530004 0.37275469
#> [2695] 0.24187269 0.25402260 0.28219382 0.15298227 0.97296546 0.32437001
#> [2701] 0.18074052 0.27017901 0.21634873 0.16263381 0.23443834 0.17182203
#> [2707] 0.59010465 0.20319280 0.21640573 0.36114815 0.70974228 0.24979430
#> [2713] 0.38420907 0.30241062 0.20619628 0.09598380 0.34610490 0.12298604
#> [2719] 0.44479066 0.16627470 0.15219872 0.22866461 0.35807480 0.23336696
#> [2725] 0.20721940 0.24597951 0.21251612 0.40752590 0.28236062 0.10240796
#> [2731] 0.18734765 0.49234546 0.45341166 0.21008184 0.23830955 0.08315416
#> [2737] 0.18536238 0.45656192 0.18341819 0.24935768 0.20851601 0.48601615
#> [2743] 0.53070571 0.46247516 0.23156935 0.65144427 0.59843376 0.18630386
#> [2749] 0.23984053 0.25231907 0.30618808 0.47932983 0.53698776 0.21118134
#> [2755] 0.27294217 0.37297107 0.24577870 0.20689967 0.15861118 0.52803806
#> [2761] 0.35345415 0.46058611 0.21661091 0.13601815 0.26571583 0.22868068
#> [2767] 0.90341591 0.19029876 0.25475576 0.47815981 0.23702779 0.19473081
#> [2773] 0.20017946 0.43095862 0.17023106 0.19585468 0.27460149 0.23724307
#> [2779] 0.25560142 0.18446690 0.68888396 0.37007849 0.48143863 0.24502301
#> [2785] 0.47874583 0.23462171 0.13951057 0.42557646 0.10065894 0.55698673
#> [2791] 0.32915153 0.07400039 0.60280252 0.27534118 0.69216451 0.25109676
#> [2797] 0.17822054 0.78881643 0.65177631 0.37135676 0.50945240 0.21034372
#> [2803] 0.57927832 0.42561825 0.22408048 0.23212262 0.36315505 0.66167476
#> [2809] 0.42429233 0.59964424 0.08033044 0.20426608 0.48784928 0.22571114
#> [2815] 0.08838234 0.14328248 0.42196540 0.33297097 0.33806666 0.23953490
#> [2821] 0.60497684 0.29298921 0.31394934 0.50194099 0.32159587 0.20345747
#> [2827] 0.58194790 0.75555198 0.18604855 0.54864236 0.29561035 0.32894384
#> [2833] 0.09448840 0.23262459 0.62851581 0.17394535 0.33100298 0.44839609
#> [2839] 0.11143102 0.71316243 0.78946016 0.33595649 0.20519353 0.14408597
#> [2845] 0.22684866 0.14621324 0.23924142 0.34245953 0.10333921 0.41000638
#> [2851] 0.49125469 0.13666370 0.46051033 0.18180162 0.12361755 0.69099267
#> [2857] 0.36915032 0.46860814 0.28783331 0.73005762 0.41985259 0.34976977
#> [2863] 0.11705127 0.19759845 0.22878430 0.29505397 0.25118296 0.93011984
#> [2869] 0.49878543 0.71855438 0.34271705 0.17392903 0.28240130 0.27456674
#> [2875] 0.18389828 0.15267265 0.48519507 0.36162082 0.16890014 0.21569223
#> [2881] 0.31882955 0.43053760 0.08478782 0.39649225 0.17086103 0.30215119
#> [2887] 0.69346983 0.14636846 0.14472699 0.23976465 0.22521725 0.19784315
#> [2893] 0.13769663 0.18138944 0.55234560 0.18422136 0.18339237 0.39870582
#> [2899] 0.35811532 0.32054149 0.71703458 0.38944924 0.17828406 0.29663641
#> [2905] 0.49605166 0.20137124 0.10745821 0.21052646 0.45481681 0.21895064
#> [2911] 0.60432818 0.19136418 0.11081114 0.25998640 0.43556615 0.25486493
#> [2917] 0.35181473 0.31109778 0.78912278 0.26460238 0.76487154 0.37619467
#> [2923] 0.25274140 0.72077354 0.45948538 0.61697912 0.19011633 0.17217515
#> [2929] 0.19370923 0.22045104 0.22557303 0.17020025 0.26230233 0.30381779
#> [2935] 0.67538796 0.29915551 0.63256828 0.13917861 0.69321806 0.44001646
#> [2941] 0.15506011 0.10461565 0.10156106 0.98024035 0.57379000 0.20738012
#> [2947] 0.18827283 0.40495528 0.44635404 0.28918262 0.15867568 0.27969444
#> [2953] 0.32160900 0.36147166 0.19871273 0.07904480 0.39268407 0.33519576
#> [2959] 0.20643376 0.79314063 0.22772891 0.08822388 0.27432539 0.45262094
#> [2965] 0.13151050 0.32986749 0.41258737 0.21311572 0.38633001 0.07679067
#> [2971] 0.21734532 0.08080568 0.27539376 0.35245755 0.26092224 0.69962824
#> [2977] 0.15742302 0.33922374 0.19278844 0.54176627 0.20612731 0.55441094
#> [2983] 0.36014391 0.17757779 0.33029497 0.18307665 0.28174304 0.21792554
#> [2989] 0.52820011 0.29579379 0.26507133 0.63314456 0.48541940 0.55260488
#> [2995] 0.10786460 0.41621298 0.05379585 0.25510635 0.39627766 0.25809818
#> [3001] 0.44902294 0.24262987 0.80479814 0.28933388 0.08839303 0.27809039
#> [3007] 0.52735181 0.35589188 0.22758307 0.15440808 0.25242322 0.70256521
#> [3013] 0.25980583 0.26038749 0.82393137 0.20028374 0.56047929 0.20356564
#> [3019] 0.39644379 0.30916705 0.52132808 0.12344917 0.30844042 0.37844881
#> [3025] 0.27482434 0.36442107 0.35593972 0.38155805 0.30520266 0.10279414
#> [3031] 0.16887108 0.59567277 0.28188319 0.31907881 0.13329839 0.49899226
#> [3037] 0.16273074 0.14904089 0.26044463 0.25548600 0.33905520 0.59365116
#> [3043] 0.21700049 0.73306772 0.29821963 0.22241884 0.47753290 0.30001302
#> [3049] 0.14662527 0.28357810 0.94989724 0.19006709 0.85387821 0.65652240
#> [3055] 0.18303770 0.72743524 0.29377314 0.23444762 0.50939863 0.26910265
#> [3061] 0.13755164 0.24323684 0.84886754 0.64526154 0.12098410 0.72631045
#> [3067] 0.34725191 0.40556631 0.27410633 0.16871017 0.08148250 0.57938166
#> [3073] 0.08952657 0.92408367 0.41117375 0.37404477 0.29255316 0.12731084
#> [3079] 0.49676938 0.41122127 0.30648996 0.16999160 0.66994865 0.64155710
#> [3085] 0.29907353 0.20563108 0.26476899 0.41002914 0.49051494 0.48966022
#> [3091] 0.53754133 0.21575797 0.20870115 0.26401749 0.50034080 0.19907715
#> [3097] 0.25749841 0.16627882 0.15822056 0.30217203 0.38600637 0.24999176
#> [3103] 0.20577605 0.36277622 0.11409693 0.54636358 0.41035562 0.24768672
#> [3109] 0.12158900 0.11474481 0.33516202 0.23260404 0.20569757 0.40480276
#> [3115] 0.42556640 0.10885113 0.31859636 0.08432728 0.79352480 0.63532739
#> [3121] 0.26097739 0.19694454 0.17041029 0.21449005 0.29605758 0.64391979
#> [3127] 0.13083064 0.26083229 0.16859459 0.16836823 0.18224230 0.37284258
#> [3133] 0.15652969 0.22540876 0.31366087 0.14355420 0.15940236 0.41682240
#> [3139] 0.22509276 0.27887163 0.37469759 0.17469731 0.40756888 0.32254689
#> [3145] 0.15331367 0.43893634 0.12132492 0.72753296 0.52131842 0.45198981
#> [3151] 0.21574413 0.11061249 0.31991328 0.57531224 0.26660661 0.19459730
#> [3157] 0.24789323 0.21258504 0.13789888 0.39043442 0.24016586 0.40086043
#> [3163] 0.28733774 0.20761259 0.88310778 0.20939489 0.15668884 0.29438030
#> [3169] 0.51008857 0.31548334 0.09266797 0.57857958 0.55498293 0.41079121
#> [3175] 0.24015131 0.24554596 0.66388428 0.36946212 0.34847630 0.27702862
#> [3181] 0.44149317 0.42204674 0.27459326 0.56961789 0.14773081 0.37716400
#> [3187] 0.13302810 0.58218305 0.38965642 0.30692319 0.17449233 0.24754357
#> [3193] 0.34496435 0.31185275 0.34298680 0.19258278 0.91866932 0.49738468
#> [3199] 0.76546016 0.31450197 0.55106816 0.46187119 0.65203022 0.35401615
#> [3205] 0.15219943 0.07873916 0.35982203 0.20931689 0.79933947 0.44760936
#> [3211] 0.30427292 0.85287826 0.29580573 0.66749403 0.32830094 0.49612054
#> [3217] 0.78812890 0.42809120 0.35180451 0.36199700 0.62342899 0.28619104
#> [3223] 0.20930134 0.38857245 0.63666322 0.16931361 0.21456541 0.22340662
#> [3229] 0.36806867 0.27816617 0.13616847 0.27063748 0.23913390 0.18834507
#> [3235] 0.32815618 0.31533017 0.46457674 0.37325320 0.13769090 0.12389251
#> [3241] 0.43282194 0.22158339 0.61802453 0.40356781 0.19815562 0.17799659
#> [3247] 0.24461487 0.26978761 0.12760953 0.53524895 0.30675448 0.19778737
#> [3253] 0.28688358 0.25422264 0.13770125 0.31089142 0.29323731 0.25561158
#> [3259] 0.10335633 0.25223544 0.74810647 0.59374571 0.23445097 0.68694097
#> [3265] 0.29047415 0.81146084 0.29007786 0.23233527 0.15760312 0.09550698
#> [3271] 0.26706188 0.24412754 0.11521640 0.58608975 0.27737938 0.11414464
#> [3277] 0.19796347 0.28699423 0.41249095 0.12021006 0.29447472 0.33993988
#> [3283] 0.63878298 0.50945973 0.52408404 0.23052778 0.22416522 0.28904881
#> [3289] 0.82050122 0.25406261 0.34098489 0.25622116 0.26678719 0.30116277
#> [3295] 0.15534276 0.18600306 0.22836953 0.26322994 0.49946944 0.20032378
#> [3301] 0.37498589 0.25746565 0.13004571 0.35127491 0.08051312 0.70349497
#> [3307] 0.38549808 0.69112318 0.11916885 0.32879892 0.14244222 0.19389455
#> [3313] 0.91417684 0.44009414 0.17884371 0.43234796 0.10611944 0.38057199
#> [3319] 0.40146180 0.38729629 0.48908851 0.28755014 0.38881882 0.48440570
#> [3325] 0.34133617 0.26026680 0.55152176 0.36031727 0.13219019 0.29896851
#> [3331] 0.20885471 0.41930691 0.16195625