Calculates the maximum sample location-scale depth

Calculates the maximum Student depth estimator of location and scale for one dimensional data (an alternative for MED and MAD or for the mean and standard deviation).

lsdSampleMaxDepth(x, iter = 100, eps = 1e-04, p_length = 10)

Arguments

x
one dimensional vector with sample

iter
maximum number of iterations in algorith for calculation Location-Scale Depth

eps
tolerance level

p_length
is the maximum length of the precision step at the end

Details

Calculations are based on lsdepth algorithm written by Ch. Muller.

References

Mizera, I., Muller, C. H., 2004. Location-scale depth (with discussion). Journal of the American Statistical Association 99, 949--966.

Examples

x <- rnorm(100)
lsdSampleMaxDepth(x)

#> An object of class "LSDepth"
#> Slot "max_depth":
#> [1] 42
#> 
#> Slot "mu":
#> [1] -0.1461054
#> 
#> Slot "sigma":
#> [1] 0.6705236
#> 

y <- rf(100, 4, 10)
lsdSampleMaxDepth(y)

#> An object of class "LSDepth"
#> Slot "max_depth":
#> [1] 42
#> 
#> Slot "mu":
#> [1] 0.6283853
#> 
#> Slot "sigma":
#> [1] 0.3431733
#>