Depth calculation

Calculate depth functions.

depth(u, X, method = "Projection", threads = -1, ...)

Arguments

u
Numerical vector or matrix whose depth is to be calculated. Dimension has to be the same as that of the observations.

X
The data as a matrix, data frame or list. If it is a matrix or data frame, then each row is viewed as one multivariate observation. If it is a list, all components must be numerical vectors of equal length (coordinates of observations).

method
Character string which determines the depth function. method can be "Projection" (the default), "Mahalanobis", "Euclidean" or "Tukey". For details see depth.

threads
number of threads used in parallel computations. Default value -1 means that all possible cores will be used.

...
parameters specific to method --- see depthEuclid

Details

The Mahalanobis depth $$ {D}_{MAH}(y, {X} ^ {n}) = \frac{ 1 }{ 1 + {{(y - \bar{x})} ^ {T}}{{S} ^ {-1}}(y - \bar{x}) }, $$ where \( S \) denotes the sample covariance matrix \( {X} ^ {n} \).

A symmetric projection depth \( D\left( x, X\right) \) of a point \( x \in {{{R}} ^ {d}} \), \( d \ge 1 \) is defined as $$ D\left( x, X\right)_{PRO} = {{\left[ 1 + su{{p}_{\left\| u \right\| = 1}}\frac{ \left| {{u} ^ {T}}x - Med\left( {{u} ^ {T}}X\right)\right| }{ MAD\left( {{u} ^ {T}}X\right) }\right]} ^ {-1}}, $$ where Med denotes the univariate median, \( MAD\left( Z \right) \) = \( Med\left(\left| Z - Med\left( Z \right)\right|\right) \). Its sample version denoted by \( D\left( x, {X} ^ {n} \right) \) or \( D\left( x, {X} ^ {n} \right) \) is obtained by replacing \( F \) by its empirical counterpart \( {{F}_{n}} \) calculated from the sample \( {X} ^ {n} \) .

Next interesting depth is the weighted \( {L} ^ {p} \) depth. The weighted \( {L} ^ {p} \) depth \( D({x}, F) \) of a point \( {x} \in {R} ^ {d} \), \( d \ge 1 \) generated by \( d \) dimensional random vector \( {X} \) with distribution \( F \), is defined as \( D({x}, F) = \frac{1 }{ 1 + Ew({{\left\| x - X \right\| }_{p}}) }, \) where \( w \) is a suitable weight function on \( [0, \infty) \), and \( {{\left\| \cdot \right\| }_{p}} \) stands for the \( {L} ^ {p} \) norm (when p = 2 we have usual Euclidean norm). We assume that \( w \) is non-decreasing and continuous on \( [0, \infty) \) with \( w(\infty-) = \infty \), and for \( a, b \in {{{R}} ^ {d}} \) satisfying \( w(\left\| a + b \right\|) \le w(\left\| a \right\|) + w(\left\| b \right\|) \). Examples of the weight functions are: \( w(x) = a + bx \), \( a, b > 0 \) or \( w(x) = {x} ^ {\alpha} \). The empirical version of the weighted \( {L} ^ {p} \) depth is obtained by replacing distribution \( F \) of \( {X} \) in \( Ew({{\left\| {x} - {X} \right\| }_{p}}) = \int {w({{\left\| x - t \right\| }_{p}})}dF(t) \) by its empirical counterpart calculated from the sample \( {{{X}} ^ {n}} \)...

The Projection and Tukey's depths are calculated using an approximate algorithm. Calculations of Mahalanobis, Euclidean and \( L ^ p \) depths are exact. Returns the depth of multivariate point u with respect to data set X.

References

Liu, R.Y., Parelius, J.M. and Singh, K. (1999), Multivariate analysis by data depth: Descriptive statistics, graphics and inference (with discussion), Ann. Statist., 27, 783--858.

Mosler K (2013). Depth statistics. In C Becker, R Fried, K S (eds.), Robustness and Complex Data Structures, Festschrift in Honour of Ursula Gather, pp. 17--34. Springer.

Rousseeuw, P.J. and Struyf, A. (1998), Computing location depth and regression depth in higher dimensions, Stat. Comput., 8, 193--203.

Zuo, Y. and Serfling, R. (2000), General Notions of Statistical Depth Functions, Ann. Statist., 28, no. 2, 461--482.

See also

Examples

library(robustbase)

# Calculation of Projection depth
data(starsCYG, package = "robustbase")
depth(t(colMeans(starsCYG)), starsCYG)

#> Depth method:  Projection 
#> [1] 0.3886082

# Also for matrices
depth(starsCYG, starsCYG)

#> Depth method:  Projection 
#>  [1] 0.45003472 0.32343780 0.31828183 0.32343780 0.31410542 0.45410905
#>  [7] 0.10820392 0.31631585 0.20415199 0.51755304 0.05766254 0.42528161
#> [13] 0.46077685 0.18085653 0.30104171 0.35406191 0.23511218 0.23670898
#> [19] 0.27372596 0.05648122 0.33092872 0.29224395 0.29579061 0.35694299
#> [25] 0.65517241 0.39274795 0.43070055 0.64850363 0.31743977 0.05485701
#> [31] 0.33080059 0.29711828 0.57675096 0.05372935 0.30738125 0.29868311
#> [37] 0.34637072 0.57675096 0.36820436 0.36705813 0.42965240 0.58946368
#> [43] 0.45561061 0.52436208 0.36532716 0.53543097 0.32231371

# Projection depth applied to a large bivariate data set
x <- matrix(rnorm(9999), nc = 3)
depth(x, x)

#> Depth method:  Projection 
#>    [1] 0.3070518 0.2490314 0.2734334 0.1986670 0.4699154 0.3227532 0.3157423
#>    [8] 0.3283721 0.2052264 0.1765570 0.2389734 0.1748218 0.3381469 0.1887558
#>   [15] 0.2283258 0.3088575 0.2708460 0.3206769 0.3857686 0.3081382 0.4656759
#>   [22] 0.3065300 0.3401081 0.2781755 0.3822524 0.2642899 0.4075856 0.1535460
#>   [29] 0.3150429 0.5183088 0.2215239 0.2408212 0.2625006 0.2854482 0.3822303
#>   [36] 0.1957103 0.1916515 0.2915228 0.2998886 0.2731612 0.4087215 0.2143645
#>   [43] 0.2845950 0.4755472 0.3240517 0.3416377 0.3381159 0.2895172 0.1933480
#>   [50] 0.5361023 0.2241425 0.4626990 0.2098977 0.3948602 0.2122756 0.2544647
#>   [57] 0.3611733 0.3096246 0.2290822 0.4151404 0.2399463 0.4236371 0.2881973
#>   [64] 0.2723428 0.5683663 0.4590164 0.2093216 0.3622729 0.2886620 0.5151168
#>   [71] 0.2597606 0.2162119 0.2164489 0.4789800 0.4167448 0.1839422 0.2611289
#>   [78] 0.4176858 0.3172762 0.3218698 0.4158773 0.4167604 0.2571741 0.2744758
#>   [85] 0.4997094 0.2759416 0.2184083 0.2745829 0.3456681 0.3407700 0.2576053
#>   [92] 0.2662004 0.3362765 0.7353151 0.2711032 0.1785856 0.2273744 0.4469855
#>   [99] 0.3739890 0.3368456 0.3339971 0.2534243 0.4224530 0.4315258 0.3572663
#>  [106] 0.3153856 0.2292257 0.3775530 0.5495940 0.1731838 0.3769145 0.4093001
#>  [113] 0.2048517 0.3431570 0.2939610 0.2045215 0.3432215 0.3001001 0.4323532
#>  [120] 0.3114274 0.2559563 0.2676785 0.3616123 0.2203665 0.2437217 0.3367771
#>  [127] 0.3709670 0.2594822 0.2987666 0.4630251 0.4414990 0.2843633 0.2032991
#>  [134] 0.3241405 0.2063021 0.3757519 0.3113587 0.4912460 0.2094625 0.2846753
#>  [141] 0.2274213 0.5155985 0.2745405 0.2633156 0.7357647 0.2268325 0.5061628
#>  [148] 0.5352033 0.2684876 0.2873019 0.4379620 0.2570656 0.2593247 0.4208204
#>  [155] 0.1754863 0.4603787 0.2128743 0.2877590 0.5093338 0.2596060 0.2870185
#>  [162] 0.2882836 0.4232047 0.1773809 0.3028899 0.4789176 0.3149678 0.1527523
#>  [169] 0.2277277 0.1714917 0.2580083 0.3819230 0.2857042 0.2148970 0.2472598
#>  [176] 0.6351412 0.2648267 0.2905798 0.4474506 0.2758789 0.3240452 0.2197375
#>  [183] 0.2528467 0.2645877 0.4173570 0.3680720 0.4210359 0.2570352 0.2946061
#>  [190] 0.2696308 0.3683686 0.2097788 0.3122985 0.4729965 0.2138428 0.2957088
#>  [197] 0.2718495 0.4807170 0.2724757 0.3796250 0.2505673 0.2339171 0.2507948
#>  [204] 0.3032227 0.1499812 0.2040734 0.4800212 0.4716393 0.1522027 0.2590395
#>  [211] 0.4646735 0.2801821 0.3774891 0.6598673 0.2337382 0.2160835 0.3010636
#>  [218] 0.4638388 0.1821320 0.5135497 0.2993717 0.2930070 0.3288407 0.3071956
#>  [225] 0.4773618 0.4025005 0.4818515 0.2069605 0.1932024 0.2701112 0.3040798
#>  [232] 0.2520949 0.2488558 0.2061670 0.4147064 0.2346529 0.4710847 0.4792672
#>  [239] 0.4186994 0.3731656 0.3649938 0.4688885 0.2426300 0.3925669 0.3173797
#>  [246] 0.4819357 0.3060046 0.3337579 0.2539702 0.2150893 0.2120263 0.2644137
#>  [253] 0.2653251 0.2459941 0.2747497 0.4886482 0.3847773 0.3735235 0.3324196
#>  [260] 0.4179926 0.1960083 0.3637967 0.2834823 0.2175019 0.2562328 0.2690778
#>  [267] 0.5185551 0.1656283 0.3474105 0.5654876 0.2344814 0.4765745 0.2862520
#>  [274] 0.3360029 0.2196019 0.1832185 0.3530067 0.4387796 0.2624899 0.2806596
#>  [281] 0.2862940 0.4128663 0.2694131 0.3278916 0.3401560 0.2724143 0.2392897
#>  [288] 0.3004371 0.3639926 0.4336232 0.3474391 0.2930857 0.1918134 0.3586494
#>  [295] 0.4293394 0.1624990 0.2825768 0.2033096 0.3497972 0.3462687 0.3274021
#>  [302] 0.3829639 0.3266121 0.2952967 0.3766217 0.2807236 0.2135655 0.5239025
#>  [309] 0.2641280 0.1858757 0.2652651 0.4947936 0.3462157 0.3641362 0.3107979
#>  [316] 0.3342780 0.2282770 0.1974798 0.3143646 0.3181223 0.2470442 0.2602283
#>  [323] 0.4579782 0.3556247 0.2454393 0.2136463 0.3738244 0.2235365 0.3000548
#>  [330] 0.1794416 0.2912024 0.3320895 0.4865985 0.4919572 0.3860318 0.5144553
#>  [337] 0.3151452 0.4953453 0.2690840 0.2952908 0.2864209 0.3356108 0.2478821
#>  [344] 0.3857265 0.3166034 0.2771173 0.3440143 0.5454960 0.3369515 0.2833648
#>  [351] 0.3208195 0.5045469 0.2848772 0.4104949 0.4184584 0.3902708 0.5417812
#>  [358] 0.2229984 0.3516085 0.2200850 0.4044289 0.4930894 0.3218094 0.3518993
#>  [365] 0.3659842 0.2504661 0.2878242 0.3913177 0.1472661 0.5424295 0.4084360
#>  [372] 0.2894909 0.2542095 0.2101133 0.6171412 0.2609715 0.3740573 0.5456118
#>  [379] 0.4099615 0.7262064 0.2856922 0.3529671 0.3578127 0.3018256 0.2184175
#>  [386] 0.5012349 0.3575548 0.6507134 0.4699736 0.2915855 0.2523621 0.3041558
#>  [393] 0.2465540 0.4740903 0.5479451 0.3956088 0.4129123 0.1965841 0.1797469
#>  [400] 0.5000000 0.2742154 0.2823072 0.2631395 0.2171769 0.2650391 0.4841315
#>  [407] 0.4390009 0.2921301 0.3372436 0.2514927 0.2586405 0.4106515 0.3348862
#>  [414] 0.2660133 0.2465943 0.2165820 0.3823031 0.3500634 0.4200599 0.4561434
#>  [421] 0.2567683 0.2551395 0.5324002 0.2521018 0.2209839 0.7593263 0.2342109
#>  [428] 0.4102869 0.1849745 0.1744191 0.3471156 0.3224647 0.2878811 0.2412085
#>  [435] 0.5409495 0.3795570 0.2226602 0.3023647 0.3341640 0.3504684 0.2364109
#>  [442] 0.3326349 0.2845805 0.2477730 0.3741456 0.2892374 0.4357142 0.1574702
#>  [449] 0.3803867 0.3777940 0.7296404 0.1934153 0.3664613 0.2906467 0.4796942
#>  [456] 0.3005456 0.1306370 0.6786773 0.2198315 0.2129966 0.3900019 0.4398897
#>  [463] 0.5168750 0.3882688 0.5587577 0.5160397 0.5645070 0.4343779 0.2951512
#>  [470] 0.3393519 0.3745157 0.2261857 0.3540033 0.2445686 0.2310976 0.3599247
#>  [477] 0.1459716 0.5261784 0.2165505 0.2481704 0.4181369 0.2029775 0.2483227
#>  [484] 0.3953485 0.3577258 0.4851495 0.3980246 0.1774977 0.1877132 0.3458057
#>  [491] 0.3043032 0.4966835 0.1770067 0.2822367 0.3245982 0.2058524 0.2084306
#>  [498] 0.3621437 0.2066687 0.3780457 0.1819220 0.3857585 0.2579221 0.3988180
#>  [505] 0.2350029 0.2011412 0.2968260 0.1898098 0.2538598 0.2442745 0.4376919
#>  [512] 0.2645574 0.2275106 0.1483748 0.3822418 0.2530391 0.2682152 0.2973896
#>  [519] 0.3519062 0.2939244 0.2409023 0.2750615 0.5991129 0.3700497 0.5078242
#>  [526] 0.2285454 0.2206466 0.4258493 0.3342028 0.3342293 0.4160383 0.1781075
#>  [533] 0.1711575 0.3118515 0.2815776 0.3451871 0.2238669 0.2990492 0.3289503
#>  [540] 0.3570073 0.6157046 0.3949053 0.2904825 0.8183651 0.2794604 0.7283907
#>  [547] 0.2926348 0.2903638 0.2856494 0.3077478 0.3223167 0.3444243 0.4140321
#>  [554] 0.2551723 0.6337642 0.2087349 0.5115363 0.1738287 0.2764462 0.2572168
#>  [561] 0.2292576 0.3558800 0.2241313 0.2300900 0.5989856 0.4005025 0.4535601
#>  [568] 0.1785437 0.3428083 0.2973094 0.1929010 0.6694335 0.3750980 0.2764660
#>  [575] 0.2901052 0.2414203 0.2456026 0.2242105 0.1953554 0.1992335 0.2938750
#>  [582] 0.2111940 0.2824038 0.4079889 0.3540517 0.4235291 0.4081405 0.1977336
#>  [589] 0.2267393 0.2283389 0.3022467 0.2183054 0.3146795 0.3037249 0.1789824
#>  [596] 0.3063360 0.5147824 0.3614700 0.3585505 0.3564820 0.2851610 0.4039411
#>  [603] 0.3061900 0.4736322 0.4461625 0.3742993 0.3852821 0.4523869 0.4936627
#>  [610] 0.2492593 0.3272547 0.3296138 0.2740192 0.2598824 0.3345377 0.2887888
#>  [617] 0.1905604 0.2088179 0.4539143 0.2870119 0.1804127 0.3849178 0.2977300
#>  [624] 0.2356156 0.6066858 0.2543048 0.3867375 0.2493220 0.2555559 0.2454861
#>  [631] 0.2458137 0.3360692 0.4463883 0.2697452 0.2454939 0.2356036 0.3873672
#>  [638] 0.3063783 0.3398539 0.3865889 0.2341139 0.3704350 0.3352370 0.3465553
#>  [645] 0.2049344 0.3302141 0.3241884 0.3914956 0.3644842 0.3010448 0.5260926
#>  [652] 0.3688402 0.2267355 0.3236709 0.3552439 0.2854832 0.4640280 0.2483209
#>  [659] 0.2171783 0.1925724 0.2353952 0.3291157 0.2676245 0.3006267 0.3385414
#>  [666] 0.3626306 0.2051312 0.3788715 0.2806087 0.2510294 0.4354677 0.3596347
#>  [673] 0.2533002 0.3768442 0.2125732 0.3910745 0.1612238 0.2237808 0.2119777
#>  [680] 0.3412642 0.1435708 0.2251997 0.2742788 0.4207199 0.2522470 0.2405492
#>  [687] 0.2793119 0.3839431 0.3825499 0.4106105 0.1882926 0.2499020 0.2079272
#>  [694] 0.2533356 0.3329350 0.2917697 0.2979964 0.3150222 0.4827588 0.2462770
#>  [701] 0.3554436 0.3440175 0.5570993 0.3220183 0.2030827 0.2507261 0.2457319
#>  [708] 0.3992910 0.3366375 0.3767911 0.2861685 0.2585436 0.3472683 0.3373399
#>  [715] 0.3296818 0.4160241 0.3487488 0.2437288 0.2920373 0.3407812 0.3510288
#>  [722] 0.4401053 0.5298522 0.3160171 0.2142320 0.1946846 0.1860472 0.6134415
#>  [729] 0.3062306 0.2918426 0.3769762 0.3691483 0.3953653 0.2714453 0.2859079
#>  [736] 0.2579562 0.3872284 0.3619358 0.2720584 0.3114491 0.2425573 0.2408433
#>  [743] 0.1985962 0.2313086 0.6294669 0.4242639 0.2217763 0.2555163 0.2782390
#>  [750] 0.4572043 0.3961321 0.5160591 0.6386168 0.2045328 0.4370205 0.2078999
#>  [757] 0.2066768 0.4797606 0.3531259 0.3335758 0.3503793 0.4618383 0.4200202
#>  [764] 0.2274808 0.1933010 0.2103872 0.6781605 0.2222768 0.4376015 0.1821125
#>  [771] 0.2793475 0.3251705 0.3560898 0.3182545 0.2789114 0.5496589 0.3566385
#>  [778] 0.2151748 0.3017832 0.2150117 0.3275842 0.3872242 0.2239177 0.2648679
#>  [785] 0.2523075 0.2303048 0.1997466 0.4843561 0.4927241 0.2596318 0.4559004
#>  [792] 0.4188281 0.3077098 0.2364161 0.3715372 0.2951858 0.4338708 0.5218059
#>  [799] 0.2383844 0.3233994 0.2902338 0.2738380 0.3225175 0.6185024 0.2440634
#>  [806] 0.5994657 0.3064213 0.2677700 0.3212774 0.2515404 0.2862471 0.2943893
#>  [813] 0.3348629 0.3194445 0.2713094 0.3642148 0.4088254 0.2395143 0.2265712
#>  [820] 0.4598139 0.3460750 0.3925037 0.2459996 0.2049745 0.2206135 0.3460594
#>  [827] 0.2058833 0.2116067 0.3850352 0.5585964 0.3155184 0.2257380 0.5675948
#>  [834] 0.4449274 0.3810152 0.3353304 0.4027341 0.2816662 0.4688496 0.2382109
#>  [841] 0.5104321 0.2948908 0.3250540 0.3121122 0.2816654 0.3858831 0.2823327
#>  [848] 0.2865050 0.2778387 0.2458501 0.3501827 0.2467558 0.1934282 0.2984639
#>  [855] 0.3616109 0.1705991 0.1780794 0.5499677 0.2831575 0.2614925 0.3414302
#>  [862] 0.3659869 0.2908202 0.3753029 0.3786805 0.2225059 0.2807750 0.1930947
#>  [869] 0.3834206 0.4893093 0.2054899 0.3640560 0.2751031 0.2533404 0.2733545
#>  [876] 0.6277438 0.2449605 0.4758901 0.2121018 0.3481979 0.2076346 0.4325601
#>  [883] 0.2293842 0.3686918 0.3101559 0.2275836 0.2144669 0.3186888 0.2958924
#>  [890] 0.2689512 0.2367497 0.3295400 0.4113033 0.2484457 0.2108231 0.2526940
#>  [897] 0.3742433 0.3183910 0.2424084 0.2046558 0.2482752 0.2865264 0.2607995
#>  [904] 0.3104845 0.3279785 0.4417318 0.4184108 0.2523829 0.2217816 0.2775894
#>  [911] 0.3716122 0.3017191 0.4395051 0.3551949 0.2073165 0.2213294 0.5333268
#>  [918] 0.2923417 0.2272681 0.2361890 0.2096557 0.3545886 0.4246906 0.4507319
#>  [925] 0.3304155 0.2450193 0.1649644 0.3111483 0.3903218 0.3347347 0.1876641
#>  [932] 0.2460402 0.3717936 0.3123591 0.4524334 0.3735416 0.2303053 0.2742520
#>  [939] 0.3551150 0.3482641 0.2060810 0.2324349 0.2319693 0.2960285 0.2104732
#>  [946] 0.4818626 0.2382991 0.4613437 0.2983197 0.5043389 0.2261897 0.3630687
#>  [953] 0.2514392 0.1999485 0.3158716 0.5099062 0.3517752 0.3008818 0.4268368
#>  [960] 0.3420329 0.5288472 0.3496433 0.2877435 0.3503294 0.2759548 0.4792495
#>  [967] 0.2900203 0.4530831 0.2626658 0.2669477 0.4698170 0.2541985 0.2772182
#>  [974] 0.2631262 0.2242461 0.1893918 0.4401928 0.2954785 0.3851665 0.2336821
#>  [981] 0.2335092 0.4529887 0.2747180 0.3076035 0.3613408 0.2931724 0.2460288
#>  [988] 0.2685336 0.3056453 0.2436106 0.2704427 0.3092097 0.3374462 0.3591599
#>  [995] 0.2613202 0.1939786 0.2036932 0.2393495 0.1626725 0.4031749 0.1981438
#> [1002] 0.4959630 0.3166999 0.2920715 0.3646561 0.2216307 0.3819096 0.4292352
#> [1009] 0.2888989 0.3737103 0.4168867 0.3620557 0.3323203 0.3458263 0.2165179
#> [1016] 0.3998509 0.2480165 0.3899270 0.3686822 0.2500195 0.2064750 0.1844305
#> [1023] 0.2625861 0.2491684 0.3216630 0.2545749 0.4411789 0.4130893 0.4748008
#> [1030] 0.2532340 0.2593336 0.3273889 0.2224418 0.3971422 0.3192983 0.4704563
#> [1037] 0.3229653 0.5739203 0.3720976 0.2243618 0.2225623 0.1930832 0.4479456
#> [1044] 0.5015570 0.2815489 0.2940102 0.1744107 0.3605281 0.2422795 0.2559725
#> [1051] 0.2517847 0.1721713 0.2790517 0.3223734 0.4886246 0.4407244 0.3964146
#> [1058] 0.2726941 0.3625645 0.2628655 0.3769446 0.2956171 0.5087956 0.3888394
#> [1065] 0.3576501 0.5133396 0.2128795 0.2470387 0.3037582 0.4436828 0.1782744
#> [1072] 0.2693368 0.2092455 0.4363715 0.3458656 0.3954336 0.2023145 0.3975507
#> [1079] 0.3295611 0.2919392 0.2550263 0.2060614 0.3050064 0.2280023 0.1993496
#> [1086] 0.1943675 0.2919818 0.3363612 0.2551225 0.2035806 0.2679558 0.2802717
#> [1093] 0.2181825 0.3060288 0.2544408 0.3288121 0.2702123 0.3087299 0.3238017
#> [1100] 0.3085545 0.1979766 0.2633950 0.3719336 0.2518881 0.2700649 0.3945715
#> [1107] 0.3496405 0.2855139 0.2234113 0.2027838 0.2278335 0.3479629 0.3579146
#> [1114] 0.4377219 0.3440307 0.4690571 0.3501642 0.3200579 0.2687710 0.2169806
#> [1121] 0.3178491 0.5736403 0.3720857 0.4075015 0.2577947 0.3512413 0.5478680
#> [1128] 0.2173534 0.5503581 0.5228776 0.2634202 0.2224832 0.3430077 0.3024251
#> [1135] 0.2304217 0.3637149 0.2605770 0.2447841 0.3720768 0.4411027 0.3320162
#> [1142] 0.2953801 0.2548322 0.2488167 0.2662126 0.4519248 0.6333443 0.3836738
#> [1149] 0.2466292 0.3338871 0.2093261 0.3487379 0.3308684 0.3800751 0.2634128
#> [1156] 0.2817611 0.2355360 0.2867831 0.2008935 0.2337930 0.2242838 0.2137936
#> [1163] 0.2971318 0.1972581 0.3336916 0.2771239 0.3445342 0.2619416 0.2949695
#> [1170] 0.2488131 0.3126342 0.3197506 0.3763476 0.2352104 0.1975903 0.3716570
#> [1177] 0.2076971 0.3500424 0.3444344 0.2733584 0.2754358 0.3806693 0.4522192
#> [1184] 0.2993296 0.2601587 0.2797878 0.3355496 0.3668838 0.4307142 0.4394932
#> [1191] 0.4111942 0.2791762 0.2992198 0.4322784 0.2295811 0.4238645 0.2697567
#> [1198] 0.4269490 0.5209688 0.5251447 0.3573007 0.1866259 0.3084569 0.2654935
#> [1205] 0.3059305 0.2278122 0.3883747 0.2733219 0.1777195 0.5378674 0.4278076
#> [1212] 0.4153349 0.2758028 0.4052745 0.5208966 0.2678707 0.3064040 0.2277000
#> [1219] 0.3598190 0.5409865 0.5460647 0.5115604 0.2489118 0.3442715 0.2984355
#> [1226] 0.2394595 0.3508571 0.2653181 0.3926189 0.3977145 0.5401593 0.5089887
#> [1233] 0.2831890 0.2309212 0.2199945 0.2683459 0.5171698 0.3049813 0.5030487
#> [1240] 0.1659349 0.5092703 0.2162336 0.3089616 0.3098581 0.4519595 0.4862722
#> [1247] 0.2666316 0.4920791 0.2440805 0.2584867 0.2662480 0.4132317 0.2223464
#> [1254] 0.2745625 0.3747607 0.2502577 0.2342360 0.3628165 0.1967155 0.3707221
#> [1261] 0.2915591 0.1868176 0.2827101 0.2051703 0.2682403 0.2043711 0.6054033
#> [1268] 0.2719994 0.2954661 0.2902803 0.3687909 0.1948453 0.5316772 0.2362821
#> [1275] 0.2853957 0.4134103 0.3314184 0.4540281 0.2761074 0.4362191 0.3186935
#> [1282] 0.3429144 0.3267213 0.4103407 0.2646419 0.3610437 0.4132625 0.2669975
#> [1289] 0.5411532 0.3066154 0.2979705 0.2182021 0.2185284 0.3546013 0.3135057
#> [1296] 0.1960735 0.3179338 0.2843216 0.2210091 0.4142584 0.2434583 0.3065402
#> [1303] 0.3449533 0.2434492 0.2944523 0.2420736 0.3460009 0.2197238 0.2393643
#> [1310] 0.2467727 0.4188491 0.2317628 0.2317540 0.3825381 0.2716863 0.5208461
#> [1317] 0.3907648 0.3771020 0.3328608 0.3464135 0.3942357 0.4712236 0.6606868
#> [1324] 0.4853354 0.3632315 0.2757658 0.1993667 0.4874996 0.4835560 0.2964697
#> [1331] 0.2285643 0.2397496 0.2629136 0.2118875 0.4252771 0.6988483 0.2209037
#> [1338] 0.2571847 0.5034885 0.4650977 0.4516321 0.2270586 0.4581465 0.2771243
#> [1345] 0.4438315 0.2379894 0.4085600 0.5640146 0.2334685 0.1787296 0.2691395
#> [1352] 0.2846191 0.3206635 0.2637148 0.2805933 0.2079763 0.2495344 0.1777981
#> [1359] 0.3185522 0.3799316 0.2626974 0.2941971 0.2356168 0.3061525 0.4658834
#> [1366] 0.3686223 0.3219219 0.3109888 0.2704211 0.2753291 0.4808620 0.1857003
#> [1373] 0.2316543 0.2498222 0.2499971 0.5112245 0.5139121 0.2398726 0.2512878
#> [1380] 0.2150497 0.4995240 0.5087931 0.5093222 0.2728200 0.3426765 0.2726130
#> [1387] 0.1812259 0.2531735 0.2476566 0.2399371 0.1921787 0.2749535 0.3154526
#> [1394] 0.2963692 0.2158231 0.3214964 0.2833013 0.3905554 0.3918380 0.4373419
#> [1401] 0.4399187 0.3462502 0.2817798 0.2763781 0.4439060 0.2890562 0.2455889
#> [1408] 0.2643525 0.4424337 0.3801711 0.2017038 0.3027377 0.2855282 0.2228850
#> [1415] 0.2589513 0.3020464 0.3423478 0.2024512 0.3137894 0.3359170 0.2631398
#> [1422] 0.3372624 0.2467103 0.3619866 0.5323986 0.3104668 0.3558773 0.5561796
#> [1429] 0.1983090 0.2231564 0.2892211 0.2057700 0.3551439 0.3468372 0.3086685
#> [1436] 0.2257387 0.3808622 0.3458808 0.2556577 0.2275107 0.5152476 0.5338598
#> [1443] 0.2801458 0.4038762 0.2242941 0.3284200 0.2325176 0.3511316 0.2927352
#> [1450] 0.2170010 0.4091701 0.2901046 0.1973000 0.4172297 0.4767514 0.3271130
#> [1457] 0.4060788 0.1871844 0.3346792 0.2392015 0.2325221 0.2147880 0.3826635
#> [1464] 0.3753604 0.2754219 0.2822705 0.3841726 0.2889449 0.4116546 0.4090699
#> [1471] 0.3477002 0.2675747 0.2212736 0.3162190 0.3547540 0.6988336 0.2524240
#> [1478] 0.2165007 0.4808706 0.3818471 0.2816243 0.4617121 0.3784552 0.3265495
#> [1485] 0.1981875 0.2061722 0.2355851 0.2253773 0.2699402 0.2862489 0.3409725
#> [1492] 0.4003343 0.3288266 0.3030020 0.2558087 0.2354651 0.1975138 0.3285133
#> [1499] 0.2966862 0.4309830 0.4451310 0.2986674 0.2756675 0.3058548 0.2099186
#> [1506] 0.2769243 0.2532038 0.2005482 0.4742934 0.4243766 0.3344936 0.2882425
#> [1513] 0.3444047 0.3003273 0.4147824 0.2824714 0.2035013 0.3055572 0.2668917
#> [1520] 0.2102857 0.4971403 0.3227474 0.2905232 0.2326280 0.2453108 0.3033392
#> [1527] 0.3443231 0.3124495 0.2214716 0.5638429 0.2544435 0.4953704 0.4111187
#> [1534] 0.2373381 0.4129438 0.2531026 0.2564451 0.2709667 0.3379010 0.1879924
#> [1541] 0.2697583 0.4095309 0.2493691 0.3435385 0.3913730 0.4226222 0.2654110
#> [1548] 0.3034662 0.2651951 0.5656138 0.2413474 0.4988354 0.1930090 0.3632858
#> [1555] 0.4600997 0.1727803 0.2403029 0.3172771 0.2729698 0.3958109 0.2217322
#> [1562] 0.3160656 0.2738658 0.2700251 0.2665331 0.2747413 0.3443685 0.3180989
#> [1569] 0.3471067 0.3567940 0.2611536 0.3502208 0.4252023 0.3605985 0.4257403
#> [1576] 0.3184753 0.5178826 0.2485653 0.2952267 0.2903009 0.5398488 0.3037364
#> [1583] 0.2115749 0.3670861 0.3002596 0.5503740 0.2448303 0.2496936 0.2849628
#> [1590] 0.3097960 0.2343810 0.3241458 0.5231027 0.2719025 0.2358163 0.4386922
#> [1597] 0.2773226 0.4222094 0.2393466 0.2962987 0.2868773 0.2128428 0.3573518
#> [1604] 0.2146252 0.3100117 0.2599189 0.3031364 0.3728074 0.5624160 0.2260377
#> [1611] 0.3107243 0.2359086 0.1864420 0.3115686 0.2236147 0.3862449 0.3552186
#> [1618] 0.2996762 0.7717254 0.2848729 0.2388211 0.4178554 0.4654834 0.2114276
#> [1625] 0.4311255 0.3292861 0.2089465 0.1916786 0.3302349 0.4854519 0.2590917
#> [1632] 0.2699167 0.2858109 0.3313164 0.3582345 0.2599039 0.2735536 0.4842279
#> [1639] 0.2341422 0.1992902 0.1677549 0.2369538 0.4036247 0.5805432 0.3227386
#> [1646] 0.2499511 0.4237469 0.2982185 0.2992941 0.2864929 0.2941414 0.2914898
#> [1653] 0.1534682 0.2553905 0.4619000 0.3457223 0.3621569 0.3716235 0.3878061
#> [1660] 0.3007803 0.2350679 0.2976644 0.2831561 0.2373706 0.3654223 0.3564093
#> [1667] 0.2414708 0.3755106 0.2915468 0.3433778 0.2784332 0.2816320 0.2802159
#> [1674] 0.3046945 0.2398570 0.1612087 0.3101459 0.3346005 0.4031874 0.2844078
#> [1681] 0.2846148 0.2084970 0.2635721 0.5909050 0.2989274 0.4789930 0.2794316
#> [1688] 0.3044257 0.1514343 0.2514400 0.2493701 0.3215001 0.2129241 0.3160331
#> [1695] 0.2725179 0.2369692 0.3746345 0.1739709 0.3330958 0.3467366 0.3232874
#> [1702] 0.5326464 0.3902584 0.2627363 0.1956959 0.3461858 0.3645017 0.2756760
#> [1709] 0.3575726 0.3092664 0.2332220 0.1676751 0.2455536 0.2046548 0.3182299
#> [1716] 0.4209600 0.3251535 0.5581686 0.2854123 0.4269129 0.3239916 0.6153473
#> [1723] 0.4105484 0.4530906 0.4960567 0.1765097 0.3247919 0.2482989 0.2267656
#> [1730] 0.1880308 0.2164243 0.5309615 0.5771075 0.2510796 0.2474277 0.2293781
#> [1737] 0.2675023 0.4968483 0.2172384 0.2024017 0.4518962 0.2154029 0.3032762
#> [1744] 0.2593849 0.2721680 0.2514168 0.2280534 0.2514002 0.3519045 0.2736480
#> [1751] 0.3589219 0.3363975 0.5524951 0.2205723 0.2658137 0.2180689 0.4066948
#> [1758] 0.2428140 0.4661718 0.3592659 0.2942617 0.2647933 0.3602318 0.3487857
#> [1765] 0.3591919 0.4135226 0.2718495 0.2298365 0.4608522 0.2895409 0.4261060
#> [1772] 0.3103911 0.1882857 0.2069719 0.2825056 0.3523372 0.6819140 0.4001708
#> [1779] 0.3976448 0.2749962 0.3089876 0.5181241 0.3907906 0.2378221 0.2006811
#> [1786] 0.3883328 0.2642112 0.2554810 0.2636867 0.2910107 0.3145375 0.1895021
#> [1793] 0.3046499 0.4377701 0.3777119 0.3998777 0.3253304 0.1699039 0.3722130
#> [1800] 0.3280607 0.3103268 0.4790225 0.2547829 0.2857634 0.3273082 0.2474286
#> [1807] 0.1837568 0.4699096 0.4025866 0.4817529 0.3674721 0.2178939 0.1950323
#> [1814] 0.2640546 0.3861576 0.5284306 0.2288528 0.2355235 0.3880870 0.3265599
#> [1821] 0.2853021 0.2449369 0.3225014 0.3686945 0.2125640 0.2959727 0.2492163
#> [1828] 0.2729525 0.3368849 0.3611803 0.1831990 0.1930084 0.2337780 0.3214939
#> [1835] 0.1890885 0.4368327 0.2331079 0.2012960 0.3035822 0.2849147 0.2554582
#> [1842] 0.4018776 0.3650079 0.3087409 0.1897012 0.4313086 0.4872105 0.2955787
#> [1849] 0.2900735 0.2943234 0.3745297 0.2590101 0.2220306 0.2389222 0.2706501
#> [1856] 0.4311187 0.6020913 0.2207030 0.3054952 0.3016042 0.2267832 0.3236983
#> [1863] 0.3795572 0.2855642 0.2900555 0.3480067 0.2305935 0.3869494 0.4569994
#> [1870] 0.2120431 0.2672664 0.2919372 0.5020200 0.4599603 0.2498183 0.1949850
#> [1877] 0.2559728 0.3985822 0.2150614 0.3738193 0.4977833 0.3151913 0.4168657
#> [1884] 0.3409852 0.2478496 0.2027537 0.2249207 0.3729794 0.6586260 0.3710469
#> [1891] 0.4174785 0.3483944 0.2318150 0.4333151 0.3154710 0.3311662 0.3439557
#> [1898] 0.3888338 0.3190140 0.2497558 0.2533390 0.2572355 0.3325712 0.4290169
#> [1905] 0.3366775 0.2687543 0.2859086 0.1943601 0.2637026 0.3678774 0.1763948
#> [1912] 0.2410255 0.2866781 0.2465646 0.2938857 0.2720340 0.2674334 0.3530870
#> [1919] 0.3291195 0.4452943 0.3450685 0.2493179 0.2238770 0.1792052 0.3988445
#> [1926] 0.1792698 0.6066472 0.3229428 0.5122933 0.2955752 0.3089698 0.1988918
#> [1933] 0.3927101 0.1672848 0.1962432 0.3347786 0.2970560 0.2975128 0.4765553
#> [1940] 0.3072290 0.2059332 0.2468603 0.3175997 0.2014682 0.3491895 0.4058579
#> [1947] 0.3525458 0.6304216 0.2603472 0.2667907 0.2796171 0.2250963 0.2864060
#> [1954] 0.2353182 0.3520111 0.2316173 0.2416356 0.3190004 0.2458949 0.6344673
#> [1961] 0.3031417 0.2262617 0.2882025 0.3399451 0.3480656 0.2044347 0.2694240
#> [1968] 0.2213304 0.3468716 0.2500962 0.3030127 0.2536466 0.4703339 0.3375988
#> [1975] 0.2188401 0.2183755 0.2120155 0.2698062 0.2305251 0.2231883 0.3718953
#> [1982] 0.2304032 0.2463904 0.3554902 0.2092863 0.3958235 0.1727385 0.2641636
#> [1989] 0.2632716 0.3168194 0.2843228 0.3308838 0.3561827 0.3556120 0.2516393
#> [1996] 0.2418675 0.2645688 0.3103079 0.3733005 0.3882401 0.2308453 0.2058768
#> [2003] 0.2973756 0.3700568 0.2611780 0.5957812 0.1945031 0.2693983 0.3394636
#> [2010] 0.3693900 0.3686618 0.2813014 0.2503455 0.4355701 0.2890105 0.2559974
#> [2017] 0.7448030 0.2810638 0.3749748 0.3144118 0.2808492 0.2552067 0.5299137
#> [2024] 0.1912482 0.1862885 0.4592664 0.2583835 0.1981880 0.2788934 0.3562611
#> [2031] 0.2456288 0.3281178 0.2083765 0.7401379 0.2179129 0.6333790 0.4173963
#> [2038] 0.5488795 0.1915914 0.2453827 0.3766516 0.2924298 0.2573655 0.3422838
#> [2045] 0.3620717 0.2670010 0.2746544 0.6813519 0.3217803 0.4541043 0.2609676
#> [2052] 0.2619874 0.2687459 0.2937401 0.2582381 0.4269412 0.2804080 0.3151499
#> [2059] 0.2170014 0.4195212 0.2577111 0.2258807 0.2110320 0.2170489 0.2172782
#> [2066] 0.2825386 0.4286717 0.4693608 0.2988874 0.3217518 0.3812509 0.3919903
#> [2073] 0.2254729 0.1910644 0.2376693 0.2709742 0.3097409 0.3742064 0.3264317
#> [2080] 0.5797911 0.2807604 0.1723883 0.3877863 0.1922247 0.2214354 0.4783723
#> [2087] 0.2998500 0.2912029 0.2451829 0.2715455 0.4419313 0.4653549 0.3907646
#> [2094] 0.3393572 0.2979929 0.1915648 0.5281138 0.2782318 0.1777818 0.2223793
#> [2101] 0.3577285 0.2296977 0.2751772 0.5803183 0.2164985 0.3144451 0.2519907
#> [2108] 0.2414104 0.3263767 0.2340260 0.2833986 0.2535861 0.2351738 0.2202361
#> [2115] 0.3118659 0.2481772 0.3606514 0.4627541 0.4809562 0.2362651 0.4414326
#> [2122] 0.2478697 0.3778470 0.3487291 0.3993818 0.3646184 0.3540005 0.4136138
#> [2129] 0.2571357 0.2326813 0.3057405 0.3568112 0.2656427 0.2194956 0.2394689
#> [2136] 0.2380260 0.2969206 0.3321917 0.4060377 0.3021268 0.2104678 0.2922766
#> [2143] 0.2479234 0.2565571 0.2174291 0.2302296 0.1936819 0.4335228 0.4260476
#> [2150] 0.2778723 0.5758456 0.3146056 0.2674019 0.2337643 0.1889206 0.3393095
#> [2157] 0.2812879 0.5438429 0.3539890 0.4012846 0.2762583 0.2737060 0.4355054
#> [2164] 0.3703850 0.3144431 0.2984699 0.3276283 0.2436284 0.3911800 0.3431359
#> [2171] 0.3073227 0.4382048 0.2163174 0.3325133 0.4753646 0.5050466 0.3392450
#> [2178] 0.3039665 0.2477178 0.3486229 0.3709540 0.3004282 0.2467258 0.3609907
#> [2185] 0.4779615 0.3187736 0.3069184 0.2366291 0.2391373 0.3696636 0.2280772
#> [2192] 0.2184223 0.2653303 0.3118526 0.2953441 0.3172877 0.3199264 0.2742252
#> [2199] 0.3563849 0.3266246 0.4083382 0.3154038 0.3448759 0.4228840 0.5745612
#> [2206] 0.2744062 0.2155440 0.2629428 0.2453029 0.3408930 0.2457982 0.3059468
#> [2213] 0.3007641 0.2399259 0.1812010 0.4054646 0.2921400 0.3032559 0.3144956
#> [2220] 0.4621875 0.5410657 0.2875591 0.2326771 0.2264546 0.2375431 0.3062664
#> [2227] 0.2984509 0.5949980 0.2116251 0.2937013 0.2974673 0.2260822 0.2585506
#> [2234] 0.3767709 0.3429556 0.2112355 0.3246665 0.2150485 0.2260867 0.2543936
#> [2241] 0.5368075 0.8822968 0.2575801 0.3761758 0.1734172 0.2364920 0.4203595
#> [2248] 0.2823348 0.3457223 0.2284853 0.3275650 0.2748569 0.3441269 0.3575605
#> [2255] 0.3251940 0.4094881 0.3011070 0.5090562 0.2518055 0.4818030 0.2330024
#> [2262] 0.2468448 0.2720636 0.5350482 0.2173539 0.2622033 0.2852284 0.2433871
#> [2269] 0.3185290 0.2374045 0.4003214 0.1671767 0.4356447 0.3592591 0.2555857
#> [2276] 0.4240593 0.5123085 0.2167986 0.2818099 0.2206958 0.3252519 0.3923711
#> [2283] 0.3678629 0.2448069 0.1672112 0.2814378 0.2683419 0.2872159 0.4145307
#> [2290] 0.3015614 0.3999411 0.4163723 0.4370782 0.2252347 0.5116827 0.4535076
#> [2297] 0.2002154 0.6559033 0.2914543 0.2375419 0.2605764 0.5622009 0.2659173
#> [2304] 0.4875367 0.2688029 0.2310203 0.7343483 0.2845617 0.4091888 0.5412602
#> [2311] 0.2234157 0.2505465 0.4097177 0.4064079 0.4453164 0.4278712 0.2156471
#> [2318] 0.2276945 0.2515689 0.5551850 0.2108489 0.3067896 0.2760737 0.3183439
#> [2325] 0.5047891 0.2717420 0.2626614 0.3171842 0.2544653 0.2283596 0.3952166
#> [2332] 0.2111557 0.1921718 0.4760343 0.2480584 0.5438559 0.1414992 0.2835792
#> [2339] 0.2164022 0.3340576 0.2558515 0.2366519 0.4341587 0.2421495 0.4726383
#> [2346] 0.2541368 0.1924047 0.2625969 0.4412387 0.3827854 0.3178096 0.3596524
#> [2353] 0.2007635 0.3360089 0.3592897 0.2506131 0.2677695 0.2862254 0.4541469
#> [2360] 0.2856155 0.2179624 0.4028512 0.4364168 0.2767383 0.2749599 0.3705831
#> [2367] 0.4496356 0.3822892 0.3107257 0.1667525 0.2562254 0.2815189 0.2095512
#> [2374] 0.2439564 0.2652404 0.2180620 0.5648718 0.1974153 0.2959244 0.3156362
#> [2381] 0.4145048 0.1987028 0.3411045 0.2424641 0.4254167 0.3638016 0.2267066
#> [2388] 0.2952285 0.4136774 0.1898929 0.2025657 0.4018517 0.2324534 0.2732065
#> [2395] 0.3388236 0.5284542 0.2690281 0.3229575 0.2805890 0.3680033 0.2856455
#> [2402] 0.3033967 0.3459200 0.3623595 0.5162164 0.2352727 0.4072490 0.2590772
#> [2409] 0.2619177 0.3681726 0.3197105 0.4548655 0.3047649 0.3391973 0.3751043
#> [2416] 0.2515353 0.2190538 0.2564685 0.2687701 0.5645683 0.2374454 0.2818331
#> [2423] 0.3124994 0.2881084 0.2570380 0.2360841 0.3287968 0.2673924 0.2866070
#> [2430] 0.3103844 0.3122028 0.3358444 0.3258092 0.3848870 0.2638438 0.4960452
#> [2437] 0.4775249 0.3101319 0.2898926 0.2617446 0.4872999 0.3435628 0.4163277
#> [2444] 0.3096783 0.2484576 0.1950270 0.4373764 0.4386713 0.2794730 0.4892568
#> [2451] 0.2620803 0.1895633 0.3364908 0.2854146 0.6349609 0.5586323 0.2020846
#> [2458] 0.5058286 0.3635489 0.2413453 0.1971168 0.2948193 0.2706635 0.3040695
#> [2465] 0.2922252 0.3017902 0.3993082 0.3598028 0.2715752 0.2267727 0.2872741
#> [2472] 0.4515845 0.4033980 0.4098016 0.3709397 0.4257474 0.3616652 0.3346093
#> [2479] 0.3754848 0.3447573 0.2959706 0.2095303 0.2426403 0.4811396 0.2909593
#> [2486] 0.3742614 0.2464448 0.3788902 0.3930387 0.3775920 0.3475680 0.2476845
#> [2493] 0.2970904 0.2358774 0.4980198 0.3442509 0.3789085 0.4054543 0.4234756
#> [2500] 0.2415747 0.2478946 0.2362544 0.5789100 0.6535870 0.4892521 0.4553045
#> [2507] 0.1908590 0.3963542 0.2910184 0.3193767 0.3523295 0.2292114 0.1579748
#> [2514] 0.6019315 0.2980828 0.4227437 0.4346614 0.3295664 0.2822109 0.2107561
#> [2521] 0.4635883 0.2291819 0.2760832 0.2733531 0.3941280 0.2269833 0.3249649
#> [2528] 0.2774983 0.1884275 0.2652486 0.2586467 0.3470460 0.2447380 0.3597981
#> [2535] 0.3059413 0.2383347 0.5190088 0.4451520 0.2728058 0.2051914 0.2379821
#> [2542] 0.2503097 0.2081575 0.2343937 0.2611820 0.2983717 0.3110699 0.3720068
#> [2549] 0.2757432 0.1974031 0.2332097 0.2390445 0.2831775 0.2670872 0.2555939
#> [2556] 0.5703716 0.3432600 0.3717390 0.3334066 0.3505698 0.2753156 0.1601754
#> [2563] 0.2865495 0.4475883 0.4447441 0.3760713 0.2386961 0.4958237 0.3863463
#> [2570] 0.2703810 0.3579540 0.3049065 0.6159155 0.3186830 0.3943192 0.3444624
#> [2577] 0.4334540 0.3668173 0.5525139 0.3648622 0.2760463 0.3037125 0.2745401
#> [2584] 0.2397663 0.2479980 0.2002758 0.3924986 0.1781565 0.3480742 0.2303164
#> [2591] 0.2948831 0.2209497 0.3025212 0.2520610 0.2837164 0.2945426 0.2695935
#> [2598] 0.2440089 0.3028019 0.3530742 0.4291843 0.4062093 0.3382544 0.3495312
#> [2605] 0.2433851 0.2274600 0.3315625 0.2559370 0.2808923 0.3001484 0.3165009
#> [2612] 0.2697741 0.2601001 0.3340130 0.2494342 0.3543272 0.1795382 0.2535539
#> [2619] 0.2901533 0.5810100 0.3193677 0.1906279 0.4029777 0.2649423 0.2240264
#> [2626] 0.5204181 0.1981140 0.3277644 0.3182720 0.3273126 0.3282420 0.4362819
#> [2633] 0.3740993 0.3103930 0.4046117 0.1927293 0.4431417 0.2883076 0.2223893
#> [2640] 0.3900280 0.2781364 0.2426148 0.2891814 0.3178613 0.1876563 0.3254646
#> [2647] 0.2813407 0.4934200 0.4485239 0.5958009 0.4090570 0.4551275 0.2690108
#> [2654] 0.2984050 0.4902596 0.2939786 0.2374995 0.2713592 0.2453117 0.3694035
#> [2661] 0.2610974 0.4297850 0.2616988 0.4500934 0.1908668 0.2760478 0.2753251
#> [2668] 0.4519014 0.3937919 0.3922561 0.2172140 0.4416519 0.3637817 0.1826210
#> [2675] 0.3432687 0.3289612 0.3749190 0.2730368 0.5681104 0.3666160 0.3492742
#> [2682] 0.2886566 0.2518446 0.1746497 0.2541858 0.5862697 0.1602680 0.3307843
#> [2689] 0.3998582 0.3686128 0.2239754 0.2745497 0.2522549 0.3201816 0.2572934
#> [2696] 0.3467510 0.3605847 0.4118065 0.4373365 0.2094456 0.2424390 0.3120856
#> [2703] 0.3082597 0.1731004 0.2910445 0.3785128 0.3546416 0.3218726 0.1991039
#> [2710] 0.5099788 0.2229913 0.4124693 0.2111358 0.3465374 0.3434171 0.2891303
#> [2717] 0.3912601 0.3513137 0.2855360 0.2950146 0.3078871 0.3029499 0.2605541
#> [2724] 0.3795255 0.3748058 0.3028282 0.2089987 0.4656665 0.2277279 0.2279150
#> [2731] 0.4208095 0.2360148 0.1932954 0.2134465 0.2327766 0.3884561 0.2368336
#> [2738] 0.4568548 0.3901863 0.3138734 0.1680633 0.2929586 0.5785859 0.3118627
#> [2745] 0.2266008 0.3093705 0.4607458 0.2994181 0.5432334 0.3759691 0.2868503
#> [2752] 0.2809986 0.3314547 0.1895963 0.3482412 0.3135559 0.1734570 0.4106157
#> [2759] 0.2948282 0.2646456 0.4520901 0.2414188 0.4907526 0.4785792 0.2673809
#> [2766] 0.3111017 0.2599143 0.5115827 0.3268047 0.2088148 0.2593013 0.3489480
#> [2773] 0.1650923 0.2992534 0.4097024 0.2904960 0.3064542 0.4107812 0.4518048
#> [2780] 0.2430878 0.3270677 0.3968584 0.3521686 0.4590517 0.6909301 0.4191544
#> [2787] 0.3724116 0.2410053 0.2497992 0.6437471 0.2471810 0.1987794 0.3959728
#> [2794] 0.2122129 0.3259861 0.3138523 0.4362141 0.1896708 0.1758434 0.3229638
#> [2801] 0.3676837 0.2774306 0.3410634 0.5161534 0.4169809 0.3985606 0.2164285
#> [2808] 0.3011387 0.2605541 0.2161748 0.2470751 0.2746251 0.2448527 0.4486887
#> [2815] 0.2959295 0.2923718 0.4195345 0.3537966 0.5315024 0.2734612 0.2447444
#> [2822] 0.2606912 0.1871272 0.5537912 0.2924570 0.2773884 0.1730739 0.3157912
#> [2829] 0.3016052 0.2848924 0.2504765 0.2300153 0.3594967 0.2764258 0.2484668
#> [2836] 0.3738707 0.2436526 0.2642674 0.1707733 0.2121423 0.6135164 0.2723736
#> [2843] 0.2708373 0.2264246 0.4599445 0.1701006 0.3229131 0.3747622 0.2427594
#> [2850] 0.1900492 0.3012555 0.2893106 0.2498010 0.2836431 0.4361061 0.3417236
#> [2857] 0.3485564 0.2168533 0.2849083 0.3160294 0.2804265 0.2825833 0.3303940
#> [2864] 0.2184872 0.3464779 0.2527937 0.2908249 0.2882568 0.2981911 0.2682078
#> [2871] 0.4147605 0.3270911 0.3797541 0.3970947 0.4226702 0.3689948 0.2542176
#> [2878] 0.4313189 0.1813392 0.2505964 0.2546729 0.2731478 0.2261158 0.2625176
#> [2885] 0.4352677 0.3280062 0.2753173 0.2702280 0.3738800 0.2537948 0.1836181
#> [2892] 0.3232051 0.3540242 0.5971987 0.4113848 0.4451492 0.2359204 0.2221515
#> [2899] 0.1753466 0.2388157 0.3168009 0.3509036 0.3181156 0.4247005 0.4794293
#> [2906] 0.3537275 0.2329987 0.3199135 0.2131757 0.2648234 0.3547078 0.3037418
#> [2913] 0.3797099 0.5170457 0.3705868 0.3646925 0.2746159 0.1902839 0.2665386
#> [2920] 0.2579969 0.2121396 0.2573087 0.4097001 0.2642130 0.2783775 0.3173054
#> [2927] 0.3038037 0.4022875 0.3166562 0.2759646 0.2094144 0.1961598 0.4769512
#> [2934] 0.5447420 0.2065455 0.2606103 0.3699513 0.3865909 0.5185301 0.3525387
#> [2941] 0.3291416 0.3481254 0.3596902 0.2979809 0.3377675 0.2515609 0.3925015
#> [2948] 0.3736527 0.5742679 0.2631780 0.2271528 0.2254439 0.2914708 0.7017807
#> [2955] 0.1645987 0.3563691 0.4430341 0.5801587 0.1661351 0.4343541 0.1995414
#> [2962] 0.3748455 0.1943573 0.3167434 0.2337839 0.3444480 0.2354496 0.2984556
#> [2969] 0.2419888 0.3184690 0.3675092 0.3390103 0.3684094 0.2978060 0.4156162
#> [2976] 0.2542751 0.2630094 0.5002311 0.3267114 0.2245783 0.3725353 0.5944576
#> [2983] 0.1937326 0.3195679 0.3264650 0.2469324 0.3584535 0.2474708 0.1951702
#> [2990] 0.7586325 0.2376294 0.1891738 0.3238332 0.2968103 0.3023395 0.2876315
#> [2997] 0.2642730 0.3356788 0.5625073 0.3642870 0.3230890 0.4104855 0.4901297
#> [3004] 0.2947209 0.3930685 0.3274714 0.2545127 0.2842058 0.1961811 0.3810468
#> [3011] 0.4051736 0.3052334 0.4325481 0.2539014 0.3093885 0.3351238 0.4003573
#> [3018] 0.3891171 0.1563959 0.4036385 0.4686247 0.1678562 0.2931631 0.3447613
#> [3025] 0.2421567 0.2982971 0.3793140 0.1882612 0.2320993 0.1892989 0.2706345
#> [3032] 0.3585981 0.2590163 0.2837636 0.5034118 0.3371744 0.3477894 0.3067409
#> [3039] 0.3611271 0.3211006 0.3107137 0.3376231 0.1822694 0.3588971 0.4744030
#> [3046] 0.4063616 0.3829363 0.5308173 0.2616346 0.2263413 0.2015906 0.2378521
#> [3053] 0.3924956 0.3303237 0.3221810 0.2427717 0.2300008 0.3681851 0.1900077
#> [3060] 0.4201724 0.6698275 0.3573920 0.3057785 0.2712007 0.2123745 0.4625182
#> [3067] 0.4799062 0.4035550 0.4213917 0.2642537 0.3214306 0.3422220 0.2666087
#> [3074] 0.2267729 0.4929322 0.2556544 0.5808676 0.2978970 0.2399509 0.4978825
#> [3081] 0.2168524 0.1810901 0.3623662 0.4656547 0.3231085 0.4145739 0.4473103
#> [3088] 0.4352238 0.8374275 0.2280606 0.3892436 0.3264398 0.2095247 0.2444817
#> [3095] 0.6638232 0.2309935 0.2633963 0.4269694 0.1863377 0.2785955 0.3632017
#> [3102] 0.2288051 0.3143194 0.4819983 0.4894082 0.3623624 0.4132695 0.2899047
#> [3109] 0.2505705 0.3360953 0.2343780 0.2376213 0.2702568 0.3821301 0.5626487
#> [3116] 0.3504178 0.2204264 0.2813794 0.2578908 0.2609987 0.2622023 0.2761935
#> [3123] 0.3254509 0.2583264 0.2265918 0.1924229 0.1987694 0.2921872 0.4002231
#> [3130] 0.3802714 0.3537081 0.2724383 0.2738566 0.2731701 0.3678487 0.5029046
#> [3137] 0.2336870 0.4231592 0.2759474 0.2786120 0.2913231 0.3628048 0.2570926
#> [3144] 0.2399263 0.3150830 0.3034750 0.2982993 0.2249273 0.3867213 0.5437216
#> [3151] 0.2736309 0.3630033 0.5260166 0.2088691 0.2852678 0.4836387 0.2278904
#> [3158] 0.2672158 0.2574224 0.2867529 0.2371910 0.2931813 0.2233879 0.3634690
#> [3165] 0.4204946 0.4310303 0.5380350 0.2708093 0.4623999 0.4121150 0.2457727
#> [3172] 0.2986131 0.3157085 0.4765579 0.3195704 0.2884524 0.2399347 0.2073589
#> [3179] 0.3906390 0.1982189 0.4489968 0.2140787 0.4408773 0.2027437 0.2633264
#> [3186] 0.3591288 0.3010768 0.3251190 0.1829542 0.2944818 0.2338792 0.4946125
#> [3193] 0.3675385 0.2375796 0.3439842 0.2616702 0.3364017 0.3462329 0.2263710
#> [3200] 0.2687765 0.2738104 0.3609405 0.3943718 0.2761465 0.4173181 0.3634800
#> [3207] 0.2039413 0.2384095 0.3268504 0.2121396 0.3508674 0.4552334 0.3606258
#> [3214] 0.3181387 0.5483848 0.3664155 0.2931424 0.2970817 0.3688742 0.5632031
#> [3221] 0.2739199 0.2434395 0.3062839 0.3145567 0.5050060 0.2872703 0.3389779
#> [3228] 0.2704167 0.2844921 0.4135982 0.2049165 0.1878341 0.1946999 0.3539640
#> [3235] 0.2740128 0.2591919 0.2977828 0.2578116 0.3354456 0.3482654 0.3068585
#> [3242] 0.2279152 0.2293126 0.1623901 0.3312953 0.4294394 0.4882985 0.1992991
#> [3249] 0.4238049 0.2508643 0.3178717 0.2068983 0.4545317 0.2973260 0.2084562
#> [3256] 0.2656222 0.1806148 0.3017576 0.4027559 0.5694816 0.2659534 0.3414673
#> [3263] 0.3554816 0.1997377 0.2051667 0.3729728 0.2847679 0.2841154 0.2521406
#> [3270] 0.4999402 0.2321582 0.3199687 0.1643044 0.2028394 0.2850755 0.2516258
#> [3277] 0.2654834 0.2051257 0.3304682 0.3107214 0.2492388 0.2539857 0.3700381
#> [3284] 0.5705540 0.4787091 0.3036674 0.2435120 0.3717636 0.2598655 0.3105043
#> [3291] 0.2942925 0.2468994 0.2183001 0.3352375 0.1953341 0.3190450 0.2504523
#> [3298] 0.1782974 0.2101858 0.2164387 0.3344250 0.2533129 0.3544283 0.2233010
#> [3305] 0.4418056 0.2565097 0.3136038 0.3667084 0.5149659 0.2955149 0.3772553
#> [3312] 0.2414649 0.3223547 0.5087103 0.4720424 0.2136456 0.4045948 0.3130680
#> [3319] 0.3937975 0.2874849 0.2568907 0.3055520 0.6218748 0.2040909 0.3605884
#> [3326] 0.3179465 0.4010167 0.4651804 0.5532824 0.3403709 0.3499913 0.4235829
#> [3333] 0.2594771