Euclidean Depth

Computes the euclidean depth of a point or vectors of points with respect to a multivariate data set.

depthEuclid(u, X)

Arguments

u
Numerical vector or matrix whose depth is to be calculated. Dimension has to be the same as that of the observations.

X
The data as a matrix, data frame or list. If it is a matrix or data frame, then each row is viewed as one multivariate observation. If it is a list, all components must be numerical vectors of equal length (coordinates of observations).

…
currently not supported.

Details

Calculation of Euclidean depth is exact.

Returns the depth of multivariate point u with respect to data set X.

Examples

x <- matrix(rnorm(9999), nc = 3)
depthEuclid(x, x)

#> Depth method:  Euclidean 
#>    [1] 0.25322372 0.51798436 0.72883690 0.87640878 0.69522902 0.39903947
#>    [7] 0.14649424 0.43286286 0.18123001 0.18224498 0.57552991 0.33924705
#>   [13] 0.23713401 0.23615072 0.71278360 0.37758496 0.16484936 0.36688803
#>   [19] 0.39587135 0.31496539 0.16317514 0.35784669 0.58549686 0.28101120
#>   [25] 0.39690687 0.16022427 0.34587037 0.44645053 0.20377527 0.29837687
#>   [31] 0.43323620 0.40262965 0.27587826 0.24109997 0.15053068 0.24609260
#>   [37] 0.25592278 0.31713056 0.63644350 0.22016511 0.14371947 0.19651975
#>   [43] 0.16320117 0.14966365 0.70531099 0.26138196 0.18572869 0.08853890
#>   [49] 0.41134288 0.39812712 0.62780832 0.29169864 0.17956496 0.07432992
#>   [55] 0.15203897 0.51427982 0.14301299 0.14820239 0.36231694 0.23472385
#>   [61] 0.10405788 0.17248577 0.53596810 0.21219148 0.25832802 0.32598696
#>   [67] 0.21024041 0.40700253 0.43670400 0.24791363 0.48938401 0.12262634
#>   [73] 0.20013477 0.37758917 0.09000373 0.45203976 0.87348966 0.39586577
#>   [79] 0.30764817 0.26000922 0.49892567 0.16053850 0.29423753 0.40394317
#>   [85] 0.53604410 0.59119479 0.35209863 0.36165691 0.70822211 0.29642424
#>   [91] 0.66879874 0.15603085 0.20727786 0.12154586 0.92882363 0.34920534
#>   [97] 0.25360690 0.18585983 0.14646423 0.77997309 0.18975489 0.16039020
#>  [103] 0.55866322 0.31498400 0.54003731 0.20991670 0.27261853 0.14020811
#>  [109] 0.19159355 0.06619503 0.27560607 0.32938995 0.10876405 0.08714024
#>  [115] 0.17039231 0.29131219 0.12953673 0.33591430 0.10470168 0.24043034
#>  [121] 0.25720252 0.31892227 0.12181079 0.27262463 0.17414793 0.22039588
#>  [127] 0.19010421 0.58507934 0.29086664 0.49753958 0.46237201 0.44913199
#>  [133] 0.39915527 0.18932348 0.11217171 0.31582450 0.42372061 0.61424716
#>  [139] 0.51570111 0.29185390 0.20587557 0.29718468 0.28061243 0.20161389
#>  [145] 0.52399253 0.62145750 0.18293246 0.51103809 0.21139291 0.45770326
#>  [151] 0.18901868 0.96554053 0.62300810 0.26725378 0.46142102 0.33032142
#>  [157] 0.69719352 0.84007409 0.12251562 0.31253379 0.84815779 0.30943298
#>  [163] 0.35258109 0.63537507 0.36147701 0.18367199 0.39987128 0.41077058
#>  [169] 0.40474140 0.51871073 0.13958919 0.14758139 0.32057112 0.23779388
#>  [175] 0.18483901 0.12810427 0.17818557 0.18406351 0.09527623 0.09308447
#>  [181] 0.48591788 0.60640202 0.24057395 0.33834556 0.12376316 0.54840751
#>  [187] 0.18826542 0.66871685 0.79628889 0.13637015 0.15764783 0.16114625
#>  [193] 0.72683756 0.16456446 0.53276137 0.24799868 0.14426763 0.15816662
#>  [199] 0.18251545 0.43415311 0.24521447 0.89415005 0.44970300 0.32418652
#>  [205] 0.43595990 0.30721696 0.38404640 0.74952311 0.17299537 0.51998436
#>  [211] 0.31095020 0.38962936 0.47500316 0.26835105 0.27907299 0.12008565
#>  [217] 0.71007995 0.33428279 0.59906278 0.47582262 0.09736083 0.43663882
#>  [223] 0.39150161 0.26460571 0.29510574 0.40385849 0.08763328 0.18706089
#>  [229] 0.21410666 0.35134438 0.19087132 0.54225113 0.37049978 0.17432585
#>  [235] 0.69273651 0.91692615 0.39127616 0.41366809 0.37898122 0.13863694
#>  [241] 0.28115222 0.13407080 0.16623138 0.35436948 0.39475752 0.41433156
#>  [247] 0.54152301 0.76430331 0.25918148 0.26521393 0.19789769 0.45270237
#>  [253] 0.24501584 0.46145071 0.17954044 0.25658354 0.79075816 0.17530672
#>  [259] 0.41016138 0.26556282 0.53580597 0.16103139 0.12685846 0.45759548
#>  [265] 0.32070967 0.32501480 0.48995134 0.32389266 0.67792872 0.29571825
#>  [271] 0.28597981 0.43672626 0.45228559 0.72781288 0.22364640 0.15035026
#>  [277] 0.51050336 0.52872434 0.31804399 0.04154229 0.82200746 0.72655707
#>  [283] 0.29500589 0.85843744 0.13082693 0.20265376 0.40150859 0.28261246
#>  [289] 0.27906372 0.17860485 0.67839317 0.33785503 0.40232747 0.31155838
#>  [295] 0.27257588 0.88952672 0.64221102 0.35426101 0.30895749 0.40340660
#>  [301] 0.37064321 0.41532256 0.14232008 0.27969331 0.11868285 0.43329233
#>  [307] 0.34619297 0.28863431 0.55111156 0.28866634 0.30900686 0.21207411
#>  [313] 0.23325457 0.62620591 0.47387120 0.46754613 0.22419299 0.61031435
#>  [319] 0.07992656 0.10532495 0.15681171 0.24264140 0.21859408 0.08016894
#>  [325] 0.21210019 0.48907004 0.11234651 0.52857357 0.61048137 0.43238234
#>  [331] 0.46307664 0.80077479 0.52505638 0.14865313 0.17747971 0.41344685
#>  [337] 0.38741338 0.22457880 0.37099914 0.20905607 0.25099470 0.23889196
#>  [343] 0.12792777 0.40534229 0.23157171 0.57977385 0.57599143 0.13906465
#>  [349] 0.33638475 0.21377300 0.19637185 0.17822681 0.32754734 0.31482900
#>  [355] 0.21525216 0.78859818 0.09935291 0.51593995 0.37966391 0.14406705
#>  [361] 0.50423517 0.12625714 0.29519150 0.37809069 0.27879508 0.28529162
#>  [367] 0.23747993 0.15657563 0.79372131 0.36054741 0.52641728 0.76079663
#>  [373] 0.18758055 0.23417354 0.43743893 0.55106027 0.45879419 0.44862222
#>  [379] 0.21938248 0.32349459 0.31581553 0.31367745 0.36572759 0.17683935
#>  [385] 0.19358890 0.46925289 0.32553932 0.30969066 0.45556012 0.23325586
#>  [391] 0.20556200 0.32509789 0.37737500 0.16698235 0.22788475 0.13575620
#>  [397] 0.54532846 0.28078355 0.38919488 0.57714578 0.18882214 0.21986594
#>  [403] 0.73295465 0.36737131 0.25998617 0.28710787 0.36949121 0.48453313
#>  [409] 0.25780103 0.17720883 0.36682045 0.20886477 0.61468935 0.43997560
#>  [415] 0.23576716 0.19165759 0.14179112 0.30422282 0.35420317 0.21403500
#>  [421] 0.29145358 0.16407719 0.11114433 0.59683693 0.19867474 0.45496780
#>  [427] 0.25396718 0.20495161 0.20047208 0.62240347 0.71909787 0.48440608
#>  [433] 0.45700758 0.16064496 0.57304034 0.49637447 0.15318811 0.11016800
#>  [439] 0.67193212 0.12694141 0.38472064 0.27330317 0.48541547 0.13111269
#>  [445] 0.90287797 0.65792674 0.53443878 0.19282733 0.16777828 0.65504956
#>  [451] 0.37496472 0.38176418 0.27542971 0.55379877 0.68823333 0.44370401
#>  [457] 0.72353574 0.45386710 0.13609491 0.13211972 0.23995526 0.31707829
#>  [463] 0.13479601 0.41500898 0.51817608 0.45595368 0.22709242 0.53571706
#>  [469] 0.23802485 0.60535364 0.39261418 0.78444371 0.27160056 0.64113572
#>  [475] 0.25789192 0.19663652 0.20021011 0.05553961 0.43360714 0.14405229
#>  [481] 0.37962507 0.37100658 0.34854929 0.77081276 0.19817957 0.15907938
#>  [487] 0.27490727 0.33015169 0.74733979 0.25982671 0.38754572 0.26807688
#>  [493] 0.17690565 0.20657014 0.32874750 0.48978828 0.66360218 0.18059654
#>  [499] 0.36911874 0.27879012 0.46266283 0.49870968 0.68498519 0.22782831
#>  [505] 0.71592557 0.22058642 0.48989396 0.20409553 0.31903198 0.21346382
#>  [511] 0.15368670 0.32595851 0.18510178 0.17764486 0.19155078 0.33369021
#>  [517] 0.23212847 0.58303474 0.25139372 0.23234634 0.24533197 0.13095591
#>  [523] 0.33289522 0.78921074 0.93381314 0.11427569 0.72905321 0.37915724
#>  [529] 0.53063990 0.28772787 0.20952264 0.06515593 0.23730474 0.40294011
#>  [535] 0.33609438 0.25738592 0.25439593 0.27555618 0.27590317 0.24632989
#>  [541] 0.51889124 0.65535298 0.31690253 0.68689550 0.51580726 0.19681897
#>  [547] 0.24401394 0.56587611 0.49724425 0.18795763 0.33046646 0.22282126
#>  [553] 0.39220487 0.14716874 0.21083851 0.15028834 0.32233736 0.42884281
#>  [559] 0.14196914 0.33738772 0.10930589 0.21292036 0.15047890 0.34247738
#>  [565] 0.17874514 0.12710298 0.12932221 0.20571444 0.40961359 0.44382490
#>  [571] 0.71035873 0.16232084 0.20548078 0.55740823 0.51403941 0.25222355
#>  [577] 0.55442406 0.36715770 0.86101538 0.12817353 0.30617414 0.49588324
#>  [583] 0.70731891 0.41087685 0.21121254 0.15846586 0.30887428 0.20595236
#>  [589] 0.10179466 0.19583409 0.15435097 0.17600763 0.35821986 0.42759644
#>  [595] 0.40611188 0.59776784 0.52819022 0.57556692 0.71996007 0.39003858
#>  [601] 0.21439340 0.23458020 0.52594242 0.11029760 0.46462544 0.16749358
#>  [607] 0.28143436 0.59716204 0.16836745 0.20882443 0.30581359 0.35446975
#>  [613] 0.23691880 0.37612223 0.44406664 0.14897607 0.13322843 0.30245172
#>  [619] 0.45390630 0.32100118 0.87870172 0.37300462 0.40106607 0.34686460
#>  [625] 0.19922117 0.15045672 0.47789672 0.11367698 0.38890619 0.30482819
#>  [631] 0.42031766 0.35541838 0.16604048 0.67592151 0.13213366 0.16187657
#>  [637] 0.12747195 0.15638533 0.15258269 0.18243022 0.22644357 0.91740924
#>  [643] 0.67121792 0.45727815 0.05594584 0.59203192 0.23315016 0.28658133
#>  [649] 0.39713231 0.32215158 0.37099618 0.34035043 0.37548062 0.38492326
#>  [655] 0.22213320 0.18201924 0.13645291 0.29753253 0.14636181 0.23771184
#>  [661] 0.31358606 0.19696416 0.55035386 0.21254166 0.22030603 0.26577968
#>  [667] 0.36763602 0.09748947 0.27203871 0.22756114 0.27582817 0.55553410
#>  [673] 0.22761547 0.33398491 0.59753560 0.92033969 0.21215788 0.37148546
#>  [679] 0.42609003 0.20846652 0.39853142 0.50486755 0.15808698 0.38004976
#>  [685] 0.21836734 0.57831540 0.09674318 0.42147195 0.16554526 0.32724649
#>  [691] 0.23126241 0.10015811 0.18593847 0.25284678 0.75647045 0.41557558
#>  [697] 0.26739099 0.19713177 0.44191272 0.36198776 0.51810315 0.12814903
#>  [703] 0.49638603 0.66633464 0.10689311 0.47456352 0.31551665 0.16997572
#>  [709] 0.21688372 0.42556315 0.23975413 0.34482040 0.46418144 0.19171275
#>  [715] 0.24659454 0.31341005 0.16533134 0.34989095 0.15070421 0.66747728
#>  [721] 0.15400926 0.35327008 0.27009266 0.48499056 0.88723531 0.20192943
#>  [727] 0.20428368 0.43074249 0.19678223 0.64307800 0.14325646 0.40086242
#>  [733] 0.18652393 0.21880730 0.32861220 0.28722090 0.35549831 0.66037833
#>  [739] 0.52827837 0.31807204 0.51865624 0.25512926 0.67051890 0.16113859
#>  [745] 0.48238236 0.77700132 0.25857343 0.21340434 0.14542434 0.59598961
#>  [751] 0.11866843 0.11447243 0.64049346 0.19945893 0.55468482 0.30467349
#>  [757] 0.63491147 0.20738675 0.30136183 0.29678210 0.74095568 0.20020261
#>  [763] 0.57422239 0.48311099 0.30553480 0.31981147 0.28992976 0.57255643
#>  [769] 0.12549323 0.21679054 0.42440124 0.62709878 0.74718448 0.20886421
#>  [775] 0.06507760 0.18340974 0.45458636 0.30650290 0.79845090 0.25175082
#>  [781] 0.45644636 0.35984740 0.76655323 0.12654152 0.49984648 0.22406973
#>  [787] 0.35114540 0.35569165 0.34370690 0.33188179 0.27078856 0.69001986
#>  [793] 0.20801409 0.09869382 0.37510938 0.06755779 0.58670585 0.73066289
#>  [799] 0.54425922 0.36602046 0.41726426 0.33544022 0.54692235 0.46432124
#>  [805] 0.17448429 0.31794970 0.34060457 0.49927454 0.34988710 0.25126395
#>  [811] 0.49350331 0.26999022 0.12277378 0.10995940 0.21576909 0.23031566
#>  [817] 0.36937830 0.15321306 0.46058449 0.75399031 0.44232577 0.15641930
#>  [823] 0.48762104 0.18046937 0.64342664 0.21689412 0.09628456 0.09422909
#>  [829] 0.22907435 0.38100627 0.09443808 0.16614219 0.18050268 0.10161784
#>  [835] 0.45439464 0.33935015 0.21776504 0.57239113 0.44293267 0.55079635
#>  [841] 0.16766948 0.11339109 0.41886152 0.20257824 0.17683367 0.18959037
#>  [847] 0.31237143 0.42128956 0.14964378 0.20460684 0.68479972 0.18651826
#>  [853] 0.31739861 0.27342547 0.59027693 0.16428363 0.23590687 0.53224460
#>  [859] 0.20038746 0.37799232 0.17800483 0.20439130 0.23736921 0.41013677
#>  [865] 0.33716536 0.24298357 0.40960746 0.27867445 0.29068047 0.59996850
#>  [871] 0.39277537 0.16941644 0.15253261 0.81724815 0.20879527 0.40021047
#>  [877] 0.59728152 0.14037693 0.26021476 0.18755184 0.12562046 0.39460894
#>  [883] 0.84706020 0.23552663 0.60351563 0.39440269 0.41566185 0.52459097
#>  [889] 0.47323166 0.19880743 0.34170107 0.36705399 0.08448272 0.17518340
#>  [895] 0.22825225 0.50104419 0.40954877 0.34356107 0.50825373 0.25638886
#>  [901] 0.60450068 0.41606897 0.61754392 0.98266837 0.20711725 0.18566876
#>  [907] 0.21397562 0.36708343 0.26964322 0.21126493 0.11283090 0.22452550
#>  [913] 0.34961145 0.17905775 0.21050290 0.46298746 0.09874564 0.18940425
#>  [919] 0.63938552 0.44030798 0.41360024 0.58054575 0.34945145 0.41416256
#>  [925] 0.14842261 0.45729079 0.32989199 0.22313303 0.65701905 0.15605263
#>  [931] 0.33887201 0.23893741 0.67600574 0.21666240 0.07736806 0.11207801
#>  [937] 0.34579510 0.20921434 0.32671628 0.51539370 0.16764879 0.43018902
#>  [943] 0.25087803 0.28294001 0.16909506 0.60372936 0.52914490 0.07046005
#>  [949] 0.51236376 0.34928546 0.23091016 0.38470420 0.18104065 0.19644250
#>  [955] 0.22070523 0.28670475 0.15917701 0.15079180 0.12501759 0.65600150
#>  [961] 0.10634193 0.70811342 0.57152434 0.28694612 0.25714834 0.74459293
#>  [967] 0.36218838 0.15143041 0.86556217 0.20357456 0.66833896 0.51584690
#>  [973] 0.13166821 0.40292685 0.47729945 0.20296977 0.22684437 0.69325318
#>  [979] 0.19401421 0.56312771 0.18472634 0.20950822 0.22489891 0.32524524
#>  [985] 0.17405723 0.21949696 0.20240423 0.53244984 0.87659150 0.20917487
#>  [991] 0.47622576 0.16254908 0.14856591 0.16685877 0.11158736 0.51641178
#>  [997] 0.33404553 0.55681538 0.27671063 0.75960308 0.50351822 0.39524148
#> [1003] 0.27999733 0.19923733 0.55288563 0.38000104 0.08729295 0.42817623
#> [1009] 0.40324526 0.12138796 0.47727093 0.11738070 0.27315737 0.68931093
#> [1015] 0.07393789 0.26404142 0.28394711 0.34413361 0.53072743 0.52016970
#> [1021] 0.58592085 0.22689759 0.46702690 0.23353364 0.50411897 0.36426217
#> [1027] 0.05577992 0.26254984 0.52736565 0.46050781 0.37144276 0.12579245
#> [1033] 0.21685437 0.41693092 0.35791696 0.35946413 0.80327777 0.60462047
#> [1039] 0.43149842 0.36100299 0.11781094 0.51970894 0.21381308 0.41470941
#> [1045] 0.74866513 0.43084922 0.17707584 0.92805818 0.34793124 0.54215850
#> [1051] 0.14885489 0.24009336 0.29231778 0.35088002 0.29292644 0.35910207
#> [1057] 0.48268278 0.41526528 0.45519299 0.24069558 0.21594043 0.12093405
#> [1063] 0.33949254 0.40054275 0.10441965 0.33153612 0.36248007 0.19960865
#> [1069] 0.14040229 0.31025484 0.37356285 0.18961803 0.34426058 0.20913611
#> [1075] 0.09399316 0.14113285 0.24338199 0.14312729 0.39839324 0.28282779
#> [1081] 0.88837927 0.21465210 0.21453448 0.09135553 0.11678754 0.61667124
#> [1087] 0.14683707 0.38051028 0.72975483 0.52445870 0.60374536 0.47066283
#> [1093] 0.32965864 0.15012897 0.10467280 0.11345209 0.12559237 0.40986982
#> [1099] 0.47269463 0.33089740 0.34511194 0.26498885 0.13367645 0.17829356
#> [1105] 0.89769898 0.32997928 0.31977082 0.39367113 0.34158652 0.26449693
#> [1111] 0.17447567 0.26723313 0.08621655 0.65711697 0.12609988 0.32438161
#> [1117] 0.27416167 0.28154567 0.60583077 0.38805068 0.21701304 0.63528539
#> [1123] 0.49359982 0.50342678 0.12781628 0.31952112 0.13570326 0.57658933
#> [1129] 0.46767419 0.20329678 0.23417099 0.44503320 0.26278946 0.57734686
#> [1135] 0.33605845 0.26171377 0.10752872 0.14669085 0.85150648 0.24777390
#> [1141] 0.47650857 0.19847729 0.09008776 0.10849825 0.24257964 0.20412339
#> [1147] 0.38496495 0.68186343 0.46106862 0.35533267 0.20546376 0.11930914
#> [1153] 0.14294933 0.40945871 0.16026033 0.36206225 0.90580704 0.25525278
#> [1159] 0.69052904 0.31821941 0.55561060 0.12597181 0.74583617 0.11961655
#> [1165] 0.19003697 0.59522303 0.27676121 0.87342179 0.21125929 0.10844999
#> [1171] 0.44978288 0.48513767 0.42406242 0.18890539 0.22825243 0.29743892
#> [1177] 0.32276196 0.18676293 0.27454308 0.11942717 0.34342569 0.20902202
#> [1183] 0.38611098 0.56717163 0.29286174 0.47679521 0.46860418 0.08439236
#> [1189] 0.58307567 0.25467435 0.73374310 0.31131717 0.06794452 0.19412525
#> [1195] 0.30679928 0.18443112 0.76621671 0.23174353 0.28604584 0.24960582
#> [1201] 0.22470618 0.21817683 0.20612502 0.59177933 0.19211022 0.50266766
#> [1207] 0.32551432 0.12743675 0.58449384 0.27035667 0.43613447 0.93046557
#> [1213] 0.19981035 0.16714173 0.11481501 0.35890585 0.61124209 0.10099119
#> [1219] 0.28533914 0.81256747 0.16010470 0.14351740 0.12866398 0.65147687
#> [1225] 0.44179185 0.60686524 0.22488723 0.31009764 0.27180875 0.22855941
#> [1231] 0.29521013 0.22211817 0.47956735 0.23964897 0.50617155 0.42901276
#> [1237] 0.31670342 0.11275348 0.45960435 0.20447730 0.55839572 0.23948431
#> [1243] 0.85298427 0.45137224 0.13334293 0.31943020 0.21388561 0.40222751
#> [1249] 0.81785443 0.30671546 0.07928432 0.24366793 0.61667812 0.12545211
#> [1255] 0.21768876 0.24241948 0.59091323 0.38566733 0.30059884 0.27643332
#> [1261] 0.24626573 0.12715929 0.23556755 0.43321360 0.18955087 0.27609219
#> [1267] 0.25004679 0.48748032 0.17657942 0.12011644 0.25219818 0.43511396
#> [1273] 0.69146116 0.22624182 0.35153046 0.30518044 0.18619706 0.56658452
#> [1279] 0.71422199 0.18914724 0.14831231 0.23743475 0.13460317 0.34673278
#> [1285] 0.13818137 0.33806218 0.41540001 0.43492905 0.22460653 0.17738607
#> [1291] 0.14251623 0.28798066 0.58030714 0.45746560 0.20833830 0.55861428
#> [1297] 0.54144629 0.31907580 0.11703914 0.76794349 0.53055405 0.14928712
#> [1303] 0.45758812 0.34240943 0.13386246 0.09307945 0.36922998 0.53124414
#> [1309] 0.39980055 0.35849230 0.25679203 0.29002023 0.17204312 0.13050510
#> [1315] 0.17245963 0.31488639 0.24130669 0.13941910 0.37055314 0.76622154
#> [1321] 0.35489792 0.82178716 0.20436882 0.22906399 0.49278248 0.15748454
#> [1327] 0.56348342 0.22362692 0.44645551 0.44980555 0.31625914 0.11697472
#> [1333] 0.17794322 0.41695214 0.48221782 0.11245188 0.69684315 0.61463330
#> [1339] 0.15090352 0.53080240 0.15173524 0.36888801 0.20948792 0.10514533
#> [1345] 0.26166907 0.47117611 0.74811309 0.13476490 0.51401547 0.18873840
#> [1351] 0.22976362 0.44418462 0.28350567 0.30709547 0.60349493 0.93237438
#> [1357] 0.23145403 0.38805430 0.18863342 0.32840347 0.37619601 0.36851678
#> [1363] 0.17358216 0.17513427 0.25465886 0.27607826 0.38455442 0.64186250
#> [1369] 0.29766838 0.34118956 0.21333083 0.36885456 0.12637046 0.27183832
#> [1375] 0.06535958 0.36187554 0.06735576 0.60421040 0.20496317 0.26924590
#> [1381] 0.50687499 0.09375016 0.28838997 0.54040898 0.19570276 0.20435158
#> [1387] 0.36570587 0.55659733 0.13479018 0.14219935 0.28891999 0.57018496
#> [1393] 0.31886898 0.31728488 0.15868127 0.15632072 0.14521807 0.10638808
#> [1399] 0.30513213 0.30005654 0.39013582 0.27883066 0.39052738 0.18384115
#> [1405] 0.20271442 0.22385307 0.36336330 0.40419098 0.49182957 0.49529936
#> [1411] 0.31455328 0.49062699 0.39423495 0.21555835 0.23551072 0.31870031
#> [1417] 0.17615820 0.31822660 0.30333009 0.46534121 0.13879383 0.26045537
#> [1423] 0.57008423 0.17518219 0.43453280 0.16190410 0.16455333 0.09282159
#> [1429] 0.24447050 0.33986779 0.54368733 0.12639008 0.47678557 0.23152238
#> [1435] 0.38462614 0.34349691 0.24483503 0.18666103 0.20912004 0.24449470
#> [1441] 0.91700393 0.19934222 0.32676798 0.12502790 0.51131961 0.14103342
#> [1447] 0.30341426 0.24414842 0.17852322 0.60342509 0.08853404 0.48630983
#> [1453] 0.26826012 0.42743582 0.51444811 0.18977713 0.18948029 0.32780497
#> [1459] 0.16819081 0.39624720 0.17185271 0.14232482 0.13026599 0.52027187
#> [1465] 0.26129990 0.47750131 0.53400152 0.14846498 0.15153237 0.49611729
#> [1471] 0.53993886 0.37507127 0.47457953 0.67763162 0.74391746 0.66286955
#> [1477] 0.28166317 0.97644492 0.50278501 0.32568151 0.49076200 0.21880632
#> [1483] 0.43179830 0.38782964 0.45072587 0.60281710 0.28023512 0.40900324
#> [1489] 0.10196919 0.27961373 0.46071855 0.30925101 0.38387574 0.38985141
#> [1495] 0.40630200 0.22897319 0.20300447 0.81626122 0.32185416 0.19361000
#> [1501] 0.17925254 0.21804560 0.15860142 0.21041820 0.52528269 0.35289544
#> [1507] 0.21632291 0.17226584 0.17466954 0.12907206 0.18712821 0.77309860
#> [1513] 0.43306822 0.29346316 0.19298876 0.64099241 0.37894722 0.46506243
#> [1519] 0.31414767 0.48592585 0.18209444 0.24392603 0.27802272 0.19010054
#> [1525] 0.38491815 0.43020857 0.89295752 0.36289515 0.11727094 0.83293375
#> [1531] 0.34438817 0.37804903 0.51511064 0.20214839 0.39563465 0.96755085
#> [1537] 0.22763154 0.15659240 0.17918755 0.41009572 0.64683350 0.16897824
#> [1543] 0.27096783 0.60497115 0.66629711 0.22370698 0.09568960 0.15066879
#> [1549] 0.62620969 0.09724234 0.32721637 0.17632563 0.29048077 0.61268937
#> [1555] 0.39312893 0.23980330 0.53413320 0.40130764 0.10715080 0.80478961
#> [1561] 0.12878448 0.19125547 0.15573568 0.45580793 0.36890010 0.18519794
#> [1567] 0.30047266 0.11803332 0.16637359 0.11938489 0.14656124 0.72608920
#> [1573] 0.73772979 0.39187464 0.45571186 0.07645493 0.71805670 0.15886240
#> [1579] 0.25379505 0.51731670 0.25982654 0.32889172 0.43617126 0.29823841
#> [1585] 0.26107467 0.31735587 0.26331774 0.34318184 0.14747811 0.39334852
#> [1591] 0.35907215 0.12317099 0.13949962 0.11719246 0.15423141 0.15773457
#> [1597] 0.19556974 0.88594186 0.63968391 0.30970709 0.24642996 0.27907612
#> [1603] 0.85136356 0.41493525 0.12253969 0.46900665 0.33823338 0.23429673
#> [1609] 0.36427066 0.23325608 0.30522714 0.70360574 0.14395996 0.76985969
#> [1615] 0.66006513 0.20928744 0.14363731 0.17379542 0.17876135 0.62620696
#> [1621] 0.12315331 0.39296075 0.43337680 0.36305131 0.10567319 0.28247488
#> [1627] 0.25795654 0.37013544 0.23211976 0.48098358 0.67448266 0.08945985
#> [1633] 0.27646678 0.28045867 0.38111771 0.86229076 0.25681027 0.71293456
#> [1639] 0.42388337 0.58920102 0.35680994 0.19632615 0.35705703 0.18530732
#> [1645] 0.09047991 0.35309416 0.39510566 0.26414662 0.89199816 0.15904187
#> [1651] 0.28865281 0.42092933 0.16029576 0.17207535 0.79011975 0.26971290
#> [1657] 0.78867379 0.29113174 0.55865594 0.27498573 0.33243007 0.44300933
#> [1663] 0.29123867 0.45147778 0.28016642 0.25530843 0.54540018 0.39470848
#> [1669] 0.11501831 0.27636743 0.18000489 0.34757135 0.42858104 0.30022425
#> [1675] 0.62760163 0.10386964 0.21757669 0.39631291 0.09782622 0.16725837
#> [1681] 0.25374832 0.28603676 0.27628980 0.54381215 0.43316186 0.25900665
#> [1687] 0.26933095 0.81250166 0.33176632 0.46133881 0.76540877 0.13222090
#> [1693] 0.13350305 0.22091382 0.29586883 0.29313886 0.65204294 0.19686392
#> [1699] 0.47354523 0.32241941 0.37506013 0.80416498 0.25851186 0.64150940
#> [1705] 0.38255833 0.31648269 0.34184904 0.30172027 0.35147368 0.22814690
#> [1711] 0.14197450 0.69065566 0.41706910 0.28925115 0.09356719 0.15909679
#> [1717] 0.21640677 0.36898738 0.30668004 0.27154655 0.46549226 0.35298371
#> [1723] 0.39833646 0.16065596 0.15087286 0.29511278 0.28590709 0.27108582
#> [1729] 0.32283147 0.43072716 0.16675172 0.41838870 0.25511344 0.60831637
#> [1735] 0.18458367 0.45209819 0.52424259 0.65355917 0.25514399 0.24226192
#> [1741] 0.53419610 0.18131207 0.10316894 0.16781652 0.16975941 0.14084023
#> [1747] 0.75152444 0.50253773 0.29773334 0.38816354 0.72105464 0.42900770
#> [1753] 0.29163545 0.50339177 0.25863207 0.11730078 0.57603278 0.58651927
#> [1759] 0.19532954 0.55518479 0.21259905 0.88739877 0.10232072 0.21120279
#> [1765] 0.24524716 0.43518077 0.11251113 0.25169417 0.35890505 0.50526846
#> [1771] 0.29191253 0.91424446 0.88417349 0.45694729 0.18459699 0.22874166
#> [1777] 0.26041913 0.08278781 0.12537431 0.19361433 0.15351441 0.25572039
#> [1783] 0.58898399 0.57908899 0.22932372 0.24739323 0.31373771 0.16432495
#> [1789] 0.34020871 0.32046372 0.37394907 0.91789583 0.31074197 0.34226636
#> [1795] 0.25801799 0.12221024 0.29732239 0.15399897 0.23676215 0.14423289
#> [1801] 0.17608670 0.22248141 0.21756872 0.24162998 0.22001905 0.77956025
#> [1807] 0.24020877 0.24444607 0.83794996 0.16503777 0.23722886 0.12981009
#> [1813] 0.24888332 0.25798712 0.68313192 0.19264369 0.04650926 0.32805655
#> [1819] 0.19244019 0.33895972 0.41620366 0.58315296 0.25714467 0.46720038
#> [1825] 0.49038080 0.26234374 0.53701181 0.06674824 0.40173686 0.42140215
#> [1831] 0.24259328 0.64989562 0.35600468 0.60676600 0.30107636 0.45514261
#> [1837] 0.32819393 0.27950232 0.27451947 0.33669826 0.28392961 0.31691354
#> [1843] 0.24500080 0.70577487 0.20297142 0.34708371 0.12297459 0.47299772
#> [1849] 0.21950962 0.21089692 0.10867903 0.34711684 0.11634319 0.16790694
#> [1855] 0.58361726 0.11825436 0.30886620 0.21779057 0.41306248 0.36474490
#> [1861] 0.16761260 0.40939129 0.26957716 0.30107488 0.44197647 0.24017412
#> [1867] 0.32228637 0.70510398 0.35516497 0.81046433 0.20819024 0.24593718
#> [1873] 0.20976027 0.62281015 0.11675724 0.82127219 0.18076011 0.17456858
#> [1879] 0.32637737 0.24226680 0.59026362 0.55730734 0.71401976 0.72479491
#> [1885] 0.25237942 0.21412385 0.49244374 0.30250197 0.78983707 0.32448879
#> [1891] 0.25688594 0.15843299 0.31289628 0.20655303 0.22832001 0.39272068
#> [1897] 0.75885370 0.25832737 0.27863047 0.07870181 0.22888974 0.33813988
#> [1903] 0.68088191 0.16117278 0.37870581 0.34058551 0.44716624 0.16562270
#> [1909] 0.59973223 0.20901678 0.16196870 0.13494507 0.18016141 0.50311870
#> [1915] 0.19091677 0.53289286 0.14932718 0.13245799 0.38407118 0.27440154
#> [1921] 0.34763369 0.42947120 0.12350737 0.21848841 0.20388320 0.41161548
#> [1927] 0.32717281 0.48741598 0.51398265 0.12421125 0.13229260 0.42334206
#> [1933] 0.20170362 0.25891334 0.12504676 0.23455788 0.36587000 0.17944163
#> [1939] 0.16241540 0.32277794 0.22245527 0.81578580 0.21476533 0.24008789
#> [1945] 0.18284310 0.29765823 0.49379562 0.17544564 0.14185197 0.16511550
#> [1951] 0.55544501 0.34875532 0.31016617 0.17523548 0.42287803 0.79168528
#> [1957] 0.83424610 0.56605193 0.42099435 0.43757234 0.47782961 0.22345111
#> [1963] 0.45946328 0.28171640 0.42365352 0.13872983 0.47000482 0.18447905
#> [1969] 0.11012022 0.31447379 0.08930335 0.54581054 0.16776163 0.12193371
#> [1975] 0.35426642 0.29938047 0.21456946 0.15007756 0.22717797 0.13352164
#> [1981] 0.20050131 0.31406244 0.45815032 0.30005181 0.17606633 0.51734480
#> [1987] 0.48819245 0.16403605 0.39771218 0.51475484 0.20447011 0.22748796
#> [1993] 0.28397339 0.35031568 0.43543654 0.23528389 0.73088741 0.19312594
#> [1999] 0.12946343 0.42327646 0.48002037 0.33615872 0.19324413 0.64745756
#> [2005] 0.26756700 0.18654003 0.31074965 0.37990052 0.31492383 0.29439006
#> [2011] 0.70634889 0.25027409 0.47276434 0.41567003 0.66674262 0.16800816
#> [2017] 0.12596026 0.22996900 0.54816726 0.79734666 0.27985087 0.23174063
#> [2023] 0.35425830 0.33022167 0.20518991 0.18247482 0.10068173 0.53479285
#> [2029] 0.11896970 0.16179414 0.08955438 0.46740369 0.57409373 0.55749820
#> [2035] 0.32610199 0.51028044 0.33861467 0.25488579 0.31861438 0.13901420
#> [2041] 0.34235118 0.57460410 0.65988453 0.35482959 0.31099302 0.14725422
#> [2047] 0.32699207 0.12894555 0.38383728 0.21263582 0.61993803 0.40708663
#> [2053] 0.46374395 0.77358911 0.52532466 0.74406906 0.77164028 0.34971870
#> [2059] 0.44287017 0.29001576 0.21046215 0.26573799 0.13758880 0.18416382
#> [2065] 0.25728093 0.22995804 0.22993529 0.18724233 0.18109611 0.30862097
#> [2071] 0.33941558 0.19256743 0.45626031 0.24047687 0.17583104 0.13210173
#> [2077] 0.20998562 0.25089640 0.31325639 0.30016085 0.19725610 0.21503363
#> [2083] 0.36158328 0.18947709 0.26048524 0.64729404 0.32493752 0.43027431
#> [2089] 0.12982964 0.33910465 0.50350203 0.40020644 0.49171944 0.08463416
#> [2095] 0.26617289 0.23792966 0.44115490 0.18059170 0.09136553 0.22670690
#> [2101] 0.07705351 0.96182861 0.27253590 0.36162997 0.38009844 0.49676766
#> [2107] 0.34509577 0.50403396 0.25237918 0.84981464 0.78653874 0.69733186
#> [2113] 0.28817629 0.18433419 0.97958097 0.35477596 0.22699174 0.26778581
#> [2119] 0.27450529 0.22563199 0.21591808 0.23152249 0.63761225 0.21306062
#> [2125] 0.14124365 0.52349448 0.46942963 0.40788545 0.11600721 0.73834596
#> [2131] 0.17530125 0.77088588 0.50241250 0.20200995 0.47080867 0.10063267
#> [2137] 0.07498507 0.44468208 0.69841700 0.55860810 0.29084010 0.06464813
#> [2143] 0.31609321 0.46218436 0.64911536 0.18354870 0.46030903 0.19808258
#> [2149] 0.46526112 0.29481973 0.17987376 0.37933128 0.47317256 0.22226569
#> [2155] 0.07768635 0.15082878 0.22674659 0.11211901 0.20258805 0.20062813
#> [2161] 0.86194508 0.92976366 0.17089992 0.17553428 0.17325689 0.07693506
#> [2167] 0.23723107 0.83169375 0.32777200 0.38593009 0.42379021 0.15385959
#> [2173] 0.39018247 0.33939652 0.59671175 0.14624252 0.52677430 0.49462909
#> [2179] 0.50854671 0.22251475 0.67970360 0.20627492 0.22978454 0.15217767
#> [2185] 0.19499065 0.46967730 0.25839434 0.62839809 0.21242910 0.23952128
#> [2191] 0.47701454 0.25977479 0.17873681 0.33696944 0.63441470 0.13028346
#> [2197] 0.21368686 0.74516622 0.24401866 0.19972881 0.30439882 0.88691557
#> [2203] 0.17964020 0.32094792 0.27795026 0.33185380 0.32452325 0.17225770
#> [2209] 0.14981957 0.20975142 0.19682290 0.64626503 0.35767117 0.20391980
#> [2215] 0.45400371 0.28287187 0.22944226 0.53543751 0.89353585 0.19568857
#> [2221] 0.56507238 0.18450500 0.50482149 0.36840673 0.45476437 0.64729487
#> [2227] 0.13966687 0.31183669 0.17307951 0.59508175 0.30703223 0.67990995
#> [2233] 0.44666503 0.31579613 0.25649716 0.59543049 0.23298603 0.23147686
#> [2239] 0.26188762 0.37847950 0.36059001 0.70943820 0.13172527 0.45156887
#> [2245] 0.31948410 0.34501983 0.14231840 0.19934962 0.87704468 0.36880678
#> [2251] 0.78204894 0.21080512 0.15989876 0.19693315 0.51769927 0.13477761
#> [2257] 0.33255437 0.11235706 0.13589335 0.40715575 0.10228825 0.39422780
#> [2263] 0.23953953 0.65794818 0.15299743 0.73858004 0.11520983 0.78583820
#> [2269] 0.19794078 0.48746677 0.26084570 0.31632434 0.08384447 0.26420127
#> [2275] 0.26801764 0.19336713 0.35772282 0.44059872 0.61548448 0.26334128
#> [2281] 0.17197468 0.18992428 0.25898062 0.50157115 0.24365067 0.21700122
#> [2287] 0.59285797 0.21893108 0.19172002 0.28962082 0.19621085 0.29525235
#> [2293] 0.60363105 0.14798404 0.68485110 0.71294998 0.15796441 0.26450749
#> [2299] 0.18271827 0.07731769 0.40355674 0.20883048 0.69971255 0.15942421
#> [2305] 0.58504614 0.26050105 0.51211050 0.57421279 0.05520906 0.09825477
#> [2311] 0.54299351 0.15882580 0.08165454 0.13334741 0.40683639 0.64636424
#> [2317] 0.17456036 0.12310914 0.65718633 0.29371702 0.21922626 0.81344811
#> [2323] 0.12105172 0.10724751 0.42245449 0.53184702 0.59337937 0.23927176
#> [2329] 0.72377682 0.39793751 0.29432802 0.46631121 0.28844058 0.72872735
#> [2335] 0.68301941 0.48098098 0.41763578 0.78392622 0.81274348 0.11937325
#> [2341] 0.71926353 0.53818789 0.68710562 0.21122622 0.24246974 0.79290952
#> [2347] 0.25502090 0.38015281 0.32688057 0.26242198 0.25315879 0.60493626
#> [2353] 0.15496705 0.11635123 0.59226002 0.28231404 0.45585525 0.24014949
#> [2359] 0.24137511 0.19663724 0.24867166 0.22320374 0.93877934 0.44934326
#> [2365] 0.39999195 0.54717805 0.24279242 0.25867456 0.55420460 0.29428320
#> [2371] 0.13809622 0.30165847 0.55369956 0.36344698 0.21435729 0.59854090
#> [2377] 0.48354985 0.18999208 0.69075684 0.25056782 0.37495580 0.41274224
#> [2383] 0.28248556 0.83217201 0.17964497 0.37410335 0.52195332 0.21635929
#> [2389] 0.40025320 0.51446097 0.59134025 0.37651063 0.22788472 0.27659490
#> [2395] 0.22774205 0.24266020 0.75597084 0.42974970 0.24719856 0.12676747
#> [2401] 0.31674319 0.72484685 0.32653865 0.18763829 0.37926038 0.21140250
#> [2407] 0.24048073 0.17226517 0.77256524 0.17020034 0.16866550 0.13828073
#> [2413] 0.76954500 0.43378967 0.44870249 0.72252175 0.41596882 0.29945769
#> [2419] 0.12316351 0.55972180 0.39017943 0.15876787 0.50101335 0.94050737
#> [2425] 0.09922386 0.27904644 0.30921812 0.22342190 0.46462342 0.16518155
#> [2431] 0.30280342 0.30448373 0.10915091 0.33553237 0.19982167 0.86231769
#> [2437] 0.41172670 0.13982137 0.49938479 0.08329231 0.49474889 0.11572307
#> [2443] 0.10423202 0.08907899 0.35070924 0.15868765 0.44393346 0.27709665
#> [2449] 0.40565491 0.70642822 0.26246962 0.33192134 0.21035625 0.41662474
#> [2455] 0.19342496 0.21027848 0.11671213 0.57351312 0.23787870 0.22804689
#> [2461] 0.22394174 0.17403207 0.39809812 0.15844962 0.19598783 0.80313254
#> [2467] 0.22074385 0.19917266 0.50968641 0.76154555 0.27200189 0.49075315
#> [2473] 0.26721589 0.52456072 0.40470576 0.24398552 0.52041749 0.17813516
#> [2479] 0.12707684 0.14068767 0.26874307 0.14359661 0.39551426 0.39318902
#> [2485] 0.47849573 0.16713639 0.34098163 0.36169253 0.42756420 0.16172644
#> [2491] 0.22363789 0.26228455 0.64530903 0.53984234 0.11602268 0.27900961
#> [2497] 0.27666356 0.56983334 0.42888840 0.69472568 0.40594082 0.46363685
#> [2503] 0.24716689 0.72833021 0.36626231 0.62383069 0.59987331 0.46461501
#> [2509] 0.66255445 0.47288173 0.18035173 0.39657889 0.22586719 0.08353573
#> [2515] 0.72150643 0.70725317 0.84402337 0.68932705 0.19335670 0.16796974
#> [2521] 0.45007400 0.59704307 0.41255884 0.41551358 0.46720106 0.12245618
#> [2527] 0.10029780 0.53625465 0.22157863 0.23501340 0.13054537 0.18975789
#> [2533] 0.36228127 0.07124840 0.21177648 0.21397640 0.19112030 0.66670727
#> [2539] 0.23608692 0.38422308 0.39803210 0.12685106 0.23087306 0.27256187
#> [2545] 0.23549443 0.26764295 0.28387900 0.33147285 0.74645665 0.15889482
#> [2551] 0.40638724 0.28807639 0.47947799 0.30657111 0.10303973 0.81324033
#> [2557] 0.22782430 0.33572812 0.32342573 0.34779116 0.16892138 0.19752998
#> [2563] 0.23180176 0.16440265 0.58393771 0.10382886 0.48415819 0.24080160
#> [2569] 0.11568285 0.28216947 0.25361293 0.51997771 0.14814355 0.38502017
#> [2575] 0.84965049 0.19299750 0.42051976 0.21369925 0.51897031 0.37578154
#> [2581] 0.11734963 0.56653888 0.57829986 0.72843290 0.29529933 0.24736415
#> [2587] 0.42664305 0.13462045 0.37907249 0.32912754 0.22298277 0.40404208
#> [2593] 0.30883683 0.13108348 0.16575360 0.41441045 0.19903535 0.50434960
#> [2599] 0.52758905 0.68131328 0.11400956 0.22765490 0.28766396 0.12774835
#> [2605] 0.51040997 0.80722559 0.18220839 0.46636679 0.83397109 0.20060813
#> [2611] 0.34228077 0.19941301 0.26609121 0.22194904 0.16395234 0.62884482
#> [2617] 0.06258103 0.18908971 0.40009434 0.26280765 0.23275776 0.52255122
#> [2623] 0.29194076 0.36903890 0.26530838 0.23531955 0.38749057 0.14882646
#> [2629] 0.29100415 0.25125037 0.94027159 0.96026422 0.66688926 0.59928790
#> [2635] 0.20851732 0.76731792 0.32527931 0.22362616 0.13098556 0.13677687
#> [2641] 0.39456538 0.19552871 0.40686345 0.79177379 0.14542720 0.19951904
#> [2647] 0.17165853 0.14535553 0.20927660 0.23927467 0.27954624 0.17462188
#> [2653] 0.20675767 0.11909193 0.13379249 0.19174451 0.42465358 0.19660759
#> [2659] 0.88223097 0.20210343 0.38282392 0.26413687 0.61123989 0.44514714
#> [2665] 0.42546034 0.35840295 0.45779763 0.61643403 0.30244116 0.15224541
#> [2671] 0.18700252 0.72366186 0.22904303 0.35738269 0.26786705 0.88885913
#> [2677] 0.78527832 0.50971950 0.37227437 0.13217384 0.42629781 0.18123516
#> [2683] 0.15188341 0.47624949 0.21043559 0.26716455 0.51064342 0.69648490
#> [2689] 0.70723335 0.16824184 0.35730259 0.69690828 0.22631695 0.21072262
#> [2695] 0.26128491 0.56253951 0.20135081 0.19980854 0.47218264 0.30605616
#> [2701] 0.20838229 0.31369318 0.43992124 0.11343155 0.15801520 0.17908670
#> [2707] 0.11145939 0.16580116 0.37952034 0.48843240 0.31260363 0.31407533
#> [2713] 0.13935286 0.21097981 0.15183979 0.20264361 0.41483721 0.65478113
#> [2719] 0.15833308 0.28428281 0.19580588 0.43314979 0.30904894 0.34774589
#> [2725] 0.50195348 0.19490023 0.10947146 0.60450054 0.23896129 0.41342037
#> [2731] 0.11526660 0.38403238 0.64926259 0.49075003 0.34062661 0.33868353
#> [2737] 0.39027316 0.32094911 0.43165226 0.18032367 0.16712937 0.64342872
#> [2743] 0.58202110 0.31640410 0.45166129 0.10367543 0.40423872 0.19606354
#> [2749] 0.17413328 0.29138122 0.62173584 0.07164556 0.22181086 0.27291763
#> [2755] 0.23893689 0.39628674 0.58777401 0.43026964 0.53349919 0.77564113
#> [2761] 0.33671102 0.28422534 0.32383504 0.13905075 0.33576618 0.14273424
#> [2767] 0.51680671 0.78784516 0.17006084 0.56830122 0.11758167 0.17484814
#> [2773] 0.47021792 0.30126611 0.19427669 0.42283844 0.68191284 0.16784089
#> [2779] 0.17061577 0.13760489 0.50320939 0.16752849 0.11053679 0.24288791
#> [2785] 0.43893985 0.82514827 0.17888078 0.55630279 0.24772084 0.17050396
#> [2791] 0.19716655 0.10261999 0.60154142 0.53098491 0.21647610 0.11681742
#> [2797] 0.35801026 0.33366164 0.10880849 0.22081383 0.25131313 0.60280868
#> [2803] 0.07230768 0.37177702 0.55012564 0.27497813 0.41436809 0.54349719
#> [2809] 0.18943718 0.27251377 0.41848744 0.77027833 0.32467880 0.12799308
#> [2815] 0.20035719 0.25802072 0.61489077 0.18668079 0.51854188 0.22105102
#> [2821] 0.29396445 0.25861335 0.28561585 0.17230134 0.11106856 0.18889018
#> [2827] 0.44251490 0.08533448 0.14806323 0.45213026 0.29667931 0.08068275
#> [2833] 0.24511662 0.11608590 0.12926561 0.11744230 0.60950469 0.59468262
#> [2839] 0.41471129 0.71444200 0.13919054 0.22111277 0.24182944 0.35829601
#> [2845] 0.54758373 0.38434376 0.24769317 0.24218182 0.12433082 0.44847355
#> [2851] 0.50008240 0.43040887 0.39950688 0.16291437 0.63614151 0.31326276
#> [2857] 0.53914467 0.31065498 0.20235853 0.29316513 0.16658600 0.51628662
#> [2863] 0.44416873 0.35512568 0.20516945 0.18909107 0.47354918 0.25212840
#> [2869] 0.48378253 0.28460216 0.32612861 0.44536491 0.20436482 0.19144563
#> [2875] 0.21886166 0.63752691 0.07700744 0.79955140 0.88565413 0.45335141
#> [2881] 0.24380932 0.65929282 0.60444119 0.26870980 0.44393720 0.22691632
#> [2887] 0.37207371 0.91418973 0.27658413 0.11647619 0.43819367 0.34925324
#> [2893] 0.51588083 0.41029702 0.16066799 0.27724828 0.69176058 0.49551325
#> [2899] 0.27195762 0.32598133 0.38322940 0.18039113 0.42180465 0.32471770
#> [2905] 0.20348849 0.56647660 0.15234956 0.18413397 0.16576200 0.76886862
#> [2911] 0.10876517 0.14345420 0.65067268 0.19888563 0.54386663 0.23117486
#> [2917] 0.16062776 0.14294986 0.25635622 0.14502904 0.23808289 0.73722569
#> [2923] 0.37768063 0.07455590 0.35642059 0.74889797 0.36794937 0.14251534
#> [2929] 0.24950077 0.24196000 0.13578794 0.17118959 0.28774824 0.25888578
#> [2935] 0.44249734 0.30930668 0.19178640 0.37629160 0.24264668 0.29940566
#> [2941] 0.47537270 0.83190990 0.73958035 0.08963242 0.13894834 0.34393660
#> [2947] 0.58559449 0.20762199 0.18899005 0.76115909 0.32165778 0.18637434
#> [2953] 0.22028996 0.34055924 0.52509707 0.35653444 0.11012382 0.35730001
#> [2959] 0.45056188 0.28380828 0.22133791 0.22560954 0.59788021 0.34151784
#> [2965] 0.25611840 0.28095523 0.19096286 0.37236427 0.43908386 0.30379369
#> [2971] 0.27313076 0.10196204 0.40909619 0.46047394 0.72850662 0.67105225
#> [2977] 0.37698985 0.24301287 0.43334254 0.61438233 0.62542812 0.13873117
#> [2983] 0.62649073 0.32794301 0.22061073 0.13162825 0.06902696 0.24053378
#> [2989] 0.60833313 0.32421346 0.38280567 0.15111897 0.22639622 0.55273990
#> [2995] 0.29975266 0.28502086 0.26841637 0.22604717 0.76040631 0.12655183
#> [3001] 0.92042156 0.73038169 0.13276499 0.25833985 0.10193000 0.64447598
#> [3007] 0.09024493 0.29708255 0.24731172 0.16577724 0.17467454 0.37878574
#> [3013] 0.22946908 0.32233359 0.22542854 0.71805720 0.24764962 0.09646854
#> [3019] 0.15160017 0.36117448 0.29233364 0.26879198 0.22494251 0.23982403
#> [3025] 0.15512499 0.16441199 0.85699020 0.47407512 0.14461914 0.15262906
#> [3031] 0.23400687 0.09013190 0.29279631 0.19017435 0.40361094 0.41543183
#> [3037] 0.32479594 0.09633753 0.18160362 0.18336660 0.79632855 0.26633897
#> [3043] 0.77050110 0.35849506 0.30669886 0.16878997 0.19950557 0.35814157
#> [3049] 0.07296195 0.33392971 0.44091267 0.36802423 0.70083991 0.20651506
#> [3055] 0.22768825 0.22697620 0.27331450 0.26684810 0.26344685 0.25239076
#> [3061] 0.15229899 0.34995530 0.15588546 0.27024767 0.23397209 0.77054176
#> [3067] 0.19981229 0.36183105 0.95064241 0.53895898 0.62952270 0.16341646
#> [3073] 0.11942547 0.38310803 0.10442619 0.28576834 0.29217666 0.61560560
#> [3079] 0.12949775 0.12539134 0.11213069 0.10803129 0.18763699 0.29817012
#> [3085] 0.14047525 0.09922989 0.54624221 0.12929036 0.60449846 0.11653829
#> [3091] 0.31109660 0.18755957 0.11089087 0.37130192 0.16643495 0.36815957
#> [3097] 0.62254043 0.66313273 0.62071382 0.11771534 0.15795435 0.17590682
#> [3103] 0.34660500 0.22098846 0.30755998 0.12856841 0.09224369 0.47213863
#> [3109] 0.15137895 0.16791052 0.25259793 0.31570384 0.53708019 0.72381709
#> [3115] 0.39782404 0.25616024 0.44235646 0.27049289 0.77332081 0.96499142
#> [3121] 0.32452003 0.15213907 0.10998987 0.53843414 0.63631448 0.23421207
#> [3127] 0.13498961 0.19671870 0.27247759 0.12758153 0.24830728 0.42807691
#> [3133] 0.96304942 0.78083709 0.19504709 0.11878136 0.45917299 0.14562209
#> [3139] 0.83071655 0.30513120 0.40043593 0.39105725 0.29313943 0.55899081
#> [3145] 0.14484300 0.20958234 0.51442704 0.45975328 0.43088873 0.10266514
#> [3151] 0.38820722 0.46661906 0.16350226 0.80320842 0.19852897 0.07376291
#> [3157] 0.33644637 0.14788995 0.18640751 0.60736968 0.67237074 0.37198666
#> [3163] 0.47837887 0.70810203 0.19147964 0.16341010 0.19560349 0.30486018
#> [3169] 0.26489273 0.32409134 0.22872427 0.34316030 0.14940473 0.59279003
#> [3175] 0.88957255 0.45895639 0.23411681 0.16827451 0.16538485 0.36759688
#> [3181] 0.30401559 0.55348085 0.32457061 0.33980875 0.21085086 0.39263653
#> [3187] 0.13280939 0.22512901 0.19293039 0.19204220 0.11447536 0.24804853
#> [3193] 0.43952958 0.58770536 0.19569159 0.24147037 0.34161606 0.37450757
#> [3199] 0.12520936 0.23046190 0.22393261 0.39346981 0.77283595 0.32608252
#> [3205] 0.22934039 0.19429746 0.49615231 0.37864139 0.34453262 0.35631622
#> [3211] 0.89523338 0.73874366 0.66257117 0.14951908 0.63641298 0.22945311
#> [3217] 0.11438613 0.55939443 0.18117963 0.63999050 0.58427159 0.57273341
#> [3223] 0.13619317 0.59740880 0.20317751 0.39330324 0.46851478 0.47250206
#> [3229] 0.29075166 0.79594499 0.54129879 0.14978362 0.10821157 0.45484163
#> [3235] 0.19833017 0.10431478 0.43111731 0.33113395 0.46778626 0.23030249
#> [3241] 0.21834986 0.39519150 0.75698307 0.15757660 0.26252879 0.03840485
#> [3247] 0.41408788 0.19505059 0.36059122 0.11419754 0.10374917 0.57513074
#> [3253] 0.20013847 0.60306286 0.40990166 0.21089274 0.35101679 0.18104218
#> [3259] 0.23616719 0.34523725 0.24140125 0.57116453 0.17762507 0.14219761
#> [3265] 0.80102370 0.64038387 0.41978227 0.31588381 0.48042305 0.39945724
#> [3271] 0.18000547 0.86106053 0.24874309 0.54030667 0.16014187 0.07270507
#> [3277] 0.27023022 0.63153352 0.39000749 0.27477799 0.26610349 0.48154419
#> [3283] 0.39672346 0.96901935 0.36759509 0.31404418 0.75065256 0.38652316
#> [3289] 0.48211590 0.35564638 0.46144962 0.37473667 0.15889037 0.33755308
#> [3295] 0.19916813 0.93568794 0.12393483 0.50868815 0.82882271 0.20635666
#> [3301] 0.52285275 0.35325916 0.35315407 0.06589176 0.25601449 0.70724015
#> [3307] 0.19910802 0.36048380 0.44372784 0.39057097 0.71471160 0.20847850
#> [3313] 0.31518699 0.46095655 0.40111474 0.30113601 0.27473140 0.15394937
#> [3319] 0.13558754 0.44220901 0.90460131 0.37138060 0.26203061 0.52715223
#> [3325] 0.23981825 0.17980164 0.56448988 0.29838636 0.44426519 0.14059422
#> [3331] 0.37158316 0.12426696 0.18078724