LP Depth

Computes the LP depth of a point or vectors of points with respect to a multivariate data set.

depthLP(u, X, pdim = 2, la = 1, lb = 1, threads = -1, func = NULL)

Arguments

u
Numerical vector or matrix whose depth is to be calculated. Dimension has to be the same as that of the observations.

X
The data as a matrix, data frame or list. If it is a matrix or data frame, then each row is viewed as one multivariate observation. If it is a list, all components must be numerical vectors of equal length (coordinates of observations).

pdim
dimension used in calculating depth function.

la
slope the weighing function.

lb
intercept in the weighing function.

threads
number of threads used in parallel computations. Default value -1 means that all possible cores will be used.

func
the weighing function. Currently it is not supported.

…
currently not supported.

Details

Returns the depth of multivariate point u with respect to data set X.

Examples

x <- matrix(rnorm(3000), ncol = 3)

# Same results
depthLP(x, x, pdim = 2)

#> Depth method:  LP 
#>    [1] 0.2427909 0.2210003 0.2123047 0.2585794 0.2366711 0.2166749 0.2124972
#>    [8] 0.2422013 0.2642428 0.2464211 0.2734009 0.2334929 0.2445814 0.2666057
#>   [15] 0.2414226 0.2449516 0.1658119 0.2587673 0.2295377 0.2759334 0.2422841
#>   [22] 0.1908795 0.2656814 0.2713232 0.2147239 0.2492344 0.2404177 0.2460758
#>   [29] 0.2711478 0.2350186 0.2532019 0.2647727 0.2137332 0.2748685 0.2657102
#>   [36] 0.2363854 0.2504816 0.2045110 0.2264310 0.2725837 0.2303020 0.2468039
#>   [43] 0.2654800 0.2632997 0.2648663 0.2189811 0.2264021 0.2538707 0.1939586
#>   [50] 0.2678311 0.2328281 0.2481562 0.2218440 0.2655542 0.2509939 0.2474405
#>   [57] 0.2510227 0.2235068 0.2742882 0.2460356 0.2669005 0.2640867 0.2606566
#>   [64] 0.2705822 0.2086168 0.2359348 0.2586834 0.2511770 0.2369095 0.2409663
#>   [71] 0.2407552 0.2208602 0.2517243 0.2556597 0.2345192 0.2579024 0.2495773
#>   [78] 0.1935323 0.2170380 0.2477526 0.2602486 0.2384803 0.2582493 0.2785070
#>   [85] 0.1947990 0.2726895 0.2263305 0.1991252 0.2133532 0.2369210 0.2569223
#>   [92] 0.2393429 0.2319994 0.2559954 0.2289448 0.2253435 0.2326315 0.2274263
#>   [99] 0.2565704 0.2684465 0.2452280 0.2437053 0.2319253 0.2119294 0.2379424
#>  [106] 0.2500889 0.2308001 0.2110234 0.2421892 0.2774632 0.2707648 0.2201762
#>  [113] 0.2618104 0.2188020 0.2361872 0.1974597 0.2720374 0.2542501 0.2613792
#>  [120] 0.2463708 0.2486890 0.2487785 0.2692425 0.2323073 0.2594996 0.1812254
#>  [127] 0.2167786 0.2120395 0.2504126 0.2137202 0.1750506 0.2644856 0.2583864
#>  [134] 0.2512825 0.2284530 0.2691603 0.1951843 0.2508252 0.2548831 0.1997117
#>  [141] 0.2577292 0.2152833 0.2679854 0.2473055 0.1875252 0.2223513 0.2740600
#>  [148] 0.2723911 0.2152264 0.2571684 0.2293930 0.2579524 0.1892162 0.2660610
#>  [155] 0.2189154 0.2181918 0.2722239 0.2611330 0.2545776 0.2405022 0.2463152
#>  [162] 0.2393806 0.2379589 0.2474807 0.2634410 0.2037990 0.2251923 0.2294402
#>  [169] 0.2653029 0.2422396 0.2608968 0.2317617 0.2018036 0.2389633 0.2566773
#>  [176] 0.2009144 0.1904892 0.2470974 0.2639625 0.2635849 0.2363741 0.2152307
#>  [183] 0.2274388 0.2351761 0.2273224 0.1939096 0.2551433 0.2410499 0.2574118
#>  [190] 0.2593962 0.2233488 0.2587828 0.2439310 0.2194574 0.2308845 0.2389533
#>  [197] 0.2409149 0.2449951 0.2036194 0.2188844 0.2015304 0.2563211 0.2496642
#>  [204] 0.2024262 0.2463792 0.2540214 0.2705487 0.2446333 0.2601216 0.2710774
#>  [211] 0.2290297 0.2465367 0.2516347 0.2384689 0.2547409 0.2638929 0.2213189
#>  [218] 0.2480041 0.2042622 0.2327858 0.2617756 0.2120984 0.2682769 0.2408285
#>  [225] 0.2656665 0.1995093 0.2742744 0.2707354 0.2130694 0.2247019 0.2584350
#>  [232] 0.2550113 0.2490789 0.2619868 0.2653059 0.2289631 0.2644792 0.2255060
#>  [239] 0.2383777 0.2540350 0.1986332 0.2350788 0.2664843 0.2381088 0.2733004
#>  [246] 0.1981933 0.2100240 0.2430711 0.2360615 0.2157161 0.2259571 0.2719370
#>  [253] 0.2606836 0.2298035 0.2752096 0.2356620 0.2444068 0.2710280 0.2721293
#>  [260] 0.2758994 0.2502625 0.2251094 0.2525991 0.2020338 0.2611833 0.2471446
#>  [267] 0.2349443 0.2368611 0.2311822 0.2635061 0.2033044 0.2318004 0.2696925
#>  [274] 0.2247140 0.2595155 0.2432385 0.2124928 0.2526917 0.2531696 0.2289650
#>  [281] 0.2647507 0.2313486 0.2651033 0.1892821 0.2637137 0.2095913 0.2222509
#>  [288] 0.2051260 0.2512918 0.2335468 0.2516308 0.2482678 0.2592236 0.2046489
#>  [295] 0.2355097 0.2433450 0.2168692 0.2356111 0.2542481 0.2498728 0.2325535
#>  [302] 0.2278307 0.2432945 0.2590285 0.2322940 0.2261547 0.1942252 0.2594415
#>  [309] 0.1757537 0.2229746 0.2500927 0.2363405 0.2346959 0.2151565 0.2550626
#>  [316] 0.2447086 0.2663709 0.2431394 0.2637795 0.2609304 0.2713625 0.2469144
#>  [323] 0.2712617 0.2625375 0.2483731 0.2558782 0.2566420 0.1941746 0.2330483
#>  [330] 0.2412808 0.2626029 0.2547293 0.2522102 0.1739802 0.2678149 0.2185255
#>  [337] 0.2039772 0.2522963 0.2700265 0.1936106 0.2669215 0.2215942 0.2379059
#>  [344] 0.2269109 0.2607632 0.2395776 0.2078714 0.2452793 0.2250614 0.2337266
#>  [351] 0.2130498 0.1824480 0.2344852 0.2454576 0.2687553 0.2527754 0.2606437
#>  [358] 0.2304811 0.1830306 0.2571555 0.1966369 0.2476017 0.1978864 0.2394728
#>  [365] 0.2764031 0.2360768 0.2141324 0.2588049 0.2516530 0.2681965 0.2159310
#>  [372] 0.2604291 0.2549147 0.2235564 0.2557647 0.2325370 0.2009829 0.2733717
#>  [379] 0.2340037 0.2293575 0.2613467 0.2305136 0.2462396 0.2512344 0.2174448
#>  [386] 0.2433461 0.2683880 0.2454097 0.2770853 0.2277277 0.2071337 0.2482776
#>  [393] 0.2703881 0.2491500 0.2215631 0.2461024 0.2504260 0.2442839 0.2391116
#>  [400] 0.2229563 0.1897928 0.2367568 0.2536103 0.2640520 0.2604801 0.2555110
#>  [407] 0.2218831 0.2537359 0.1857510 0.2539241 0.2256168 0.2653007 0.2563688
#>  [414] 0.2410858 0.1892046 0.2298668 0.2202528 0.2456819 0.2340012 0.2564236
#>  [421] 0.2638576 0.2647188 0.2741934 0.2698102 0.2607370 0.2514361 0.2045626
#>  [428] 0.2509440 0.2403331 0.2710545 0.2460732 0.2548259 0.2696750 0.2463612
#>  [435] 0.2144655 0.2315791 0.2478441 0.2474326 0.2352449 0.2554245 0.2410957
#>  [442] 0.2306968 0.2586007 0.2169796 0.2460024 0.2305849 0.2179367 0.2291196
#>  [449] 0.2072448 0.2315287 0.2436197 0.2191231 0.2423534 0.2147810 0.2636288
#>  [456] 0.2388070 0.2550915 0.2079927 0.2459706 0.2611219 0.2600330 0.2600479
#>  [463] 0.2578366 0.2098948 0.2398636 0.2191763 0.1646189 0.2609287 0.2177501
#>  [470] 0.2613709 0.1872482 0.2294508 0.2322526 0.2353084 0.2245547 0.2478688
#>  [477] 0.2281832 0.2624358 0.2570760 0.2760997 0.2288089 0.2780750 0.2438748
#>  [484] 0.2153240 0.2541137 0.2462272 0.1909154 0.2475102 0.2633952 0.2144366
#>  [491] 0.2126021 0.2522038 0.2568049 0.2476993 0.2343218 0.2330575 0.2252594
#>  [498] 0.2669858 0.2436734 0.2220180 0.2403617 0.2662902 0.2402948 0.2738543
#>  [505] 0.2388655 0.2339455 0.2656990 0.2074354 0.2530969 0.2211438 0.2260380
#>  [512] 0.2317407 0.2650573 0.2291674 0.2230660 0.1982989 0.2551655 0.2537867
#>  [519] 0.2179432 0.2280763 0.2405791 0.2010147 0.2496041 0.2665212 0.2171313
#>  [526] 0.2709826 0.2390750 0.2775925 0.2362186 0.2660946 0.2153276 0.2653025
#>  [533] 0.2246347 0.2503122 0.2343621 0.2318206 0.2756471 0.2604965 0.2701056
#>  [540] 0.2216093 0.2539556 0.2406582 0.2179210 0.2115836 0.2375254 0.2646333
#>  [547] 0.2723897 0.2606912 0.2139962 0.2469734 0.2605046 0.2269779 0.2569399
#>  [554] 0.2458399 0.2299934 0.2565733 0.2398311 0.2290079 0.2222047 0.2555897
#>  [561] 0.2301097 0.2330354 0.2526942 0.2289820 0.2182700 0.2221681 0.2445622
#>  [568] 0.2593422 0.2607806 0.2773496 0.2168926 0.2665262 0.2286429 0.2163447
#>  [575] 0.2480339 0.2537473 0.1805767 0.2418140 0.2701197 0.1655483 0.2270231
#>  [582] 0.2366296 0.2762042 0.2135204 0.2392060 0.2536816 0.2410140 0.2288808
#>  [589] 0.2607341 0.2068315 0.2383713 0.1985451 0.2035294 0.2293576 0.1955319
#>  [596] 0.2439454 0.2284396 0.2410367 0.2674815 0.2369477 0.2378387 0.2056102
#>  [603] 0.1911474 0.2518763 0.2149633 0.2534446 0.2324020 0.2213405 0.2229601
#>  [610] 0.2499051 0.2507526 0.2327361 0.2703575 0.2441275 0.1930882 0.2773606
#>  [617] 0.2427500 0.2228632 0.2619439 0.1797325 0.2413019 0.2688531 0.2238598
#>  [624] 0.2499339 0.2045650 0.2728115 0.2297258 0.2376593 0.2659160 0.2347955
#>  [631] 0.2216335 0.2461830 0.2091525 0.2742095 0.2655819 0.1920641 0.2061693
#>  [638] 0.2345219 0.2267037 0.2559772 0.2140029 0.2650299 0.2598993 0.2453823
#>  [645] 0.2644062 0.2252552 0.2286380 0.2267269 0.2287589 0.2472402 0.2350674
#>  [652] 0.2718122 0.2435404 0.2738958 0.1916276 0.2296969 0.1796600 0.2217719
#>  [659] 0.2608519 0.2240201 0.2534990 0.2334448 0.2513065 0.2301333 0.2615795
#>  [666] 0.1998085 0.2448011 0.2616074 0.2090021 0.2392772 0.2361373 0.2556746
#>  [673] 0.2432675 0.2109642 0.2428972 0.2314294 0.2262786 0.2409106 0.2525626
#>  [680] 0.2463034 0.2360655 0.2224152 0.1960012 0.2589059 0.2535157 0.2708696
#>  [687] 0.1899624 0.1899026 0.2163777 0.2414860 0.2453195 0.2546406 0.2520706
#>  [694] 0.2533524 0.2301559 0.2390934 0.2329267 0.2241520 0.2605377 0.2716965
#>  [701] 0.1954717 0.2720147 0.2538393 0.2359471 0.2519960 0.2090026 0.2431617
#>  [708] 0.2546305 0.2608987 0.2686477 0.2436205 0.2146007 0.2430986 0.1894953
#>  [715] 0.2143878 0.2310127 0.2631566 0.2048310 0.2349074 0.1975547 0.2703692
#>  [722] 0.2108293 0.2714788 0.2477974 0.2346540 0.1960712 0.2174796 0.2610929
#>  [729] 0.2424801 0.2486014 0.2030820 0.2068625 0.2559142 0.2287077 0.2646058
#>  [736] 0.2596008 0.2456259 0.2598309 0.2158733 0.2506192 0.2575498 0.2230852
#>  [743] 0.2581686 0.2350254 0.2406998 0.2364808 0.2476188 0.2193616 0.2571394
#>  [750] 0.2266775 0.2659634 0.2691027 0.2395662 0.2338783 0.2249483 0.2072727
#>  [757] 0.2097853 0.2601533 0.2342706 0.2399822 0.2127475 0.2327150 0.2466058
#>  [764] 0.2300484 0.1737881 0.2429742 0.1940016 0.2090511 0.2174444 0.2523429
#>  [771] 0.2191560 0.2713136 0.2284911 0.2390948 0.2720453 0.2136452 0.2457490
#>  [778] 0.2369520 0.2430666 0.1827728 0.2502924 0.2249533 0.2048112 0.2135260
#>  [785] 0.2640992 0.2619630 0.2447890 0.2008870 0.2624444 0.2507495 0.2137187
#>  [792] 0.2312302 0.2497326 0.2436007 0.2089537 0.2436662 0.2233707 0.2497888
#>  [799] 0.2389659 0.1573483 0.2029376 0.2687425 0.2073743 0.2420331 0.2335416
#>  [806] 0.2704235 0.2072003 0.2521586 0.2679009 0.2436737 0.2616712 0.1949550
#>  [813] 0.2447690 0.2365638 0.2694325 0.2194435 0.2415995 0.2739425 0.2420484
#>  [820] 0.2313645 0.2452345 0.2436932 0.2667436 0.2129185 0.2243274 0.2444578
#>  [827] 0.1952932 0.2276269 0.2436390 0.2424167 0.2383632 0.2510456 0.2421080
#>  [834] 0.2566417 0.2176548 0.2578006 0.2469970 0.2540510 0.2177444 0.2321666
#>  [841] 0.2464540 0.2591665 0.2621496 0.2055367 0.2169105 0.2551476 0.2574035
#>  [848] 0.2695791 0.2482425 0.2550558 0.2721558 0.2167706 0.2350901 0.2741623
#>  [855] 0.2310673 0.2320817 0.2373834 0.2253783 0.2302870 0.2317684 0.2365341
#>  [862] 0.2753847 0.2341423 0.1969268 0.2168149 0.2069227 0.2600990 0.2406928
#>  [869] 0.2105584 0.2651195 0.2678517 0.2274964 0.2405343 0.2541948 0.2306149
#>  [876] 0.2550911 0.2567941 0.2629562 0.2700171 0.2646674 0.2473429 0.2729682
#>  [883] 0.2489599 0.2725031 0.2147414 0.2062292 0.2712514 0.1908762 0.2751433
#>  [890] 0.1641540 0.2631182 0.2087531 0.1955070 0.2493885 0.2592843 0.1956537
#>  [897] 0.2620728 0.2517706 0.2472103 0.2706150 0.2324381 0.2518682 0.2418493
#>  [904] 0.1925784 0.2666974 0.2324145 0.2768905 0.2617888 0.2521473 0.2043205
#>  [911] 0.2737578 0.2283362 0.1731870 0.2270398 0.2428617 0.2644004 0.2001743
#>  [918] 0.2416230 0.2274863 0.2423911 0.2393983 0.2649663 0.2105740 0.2559121
#>  [925] 0.2766768 0.2603459 0.2670493 0.2656743 0.2199760 0.2330662 0.2235536
#>  [932] 0.2399971 0.2261592 0.2223345 0.2342331 0.2269985 0.1839315 0.2436470
#>  [939] 0.2177731 0.2569388 0.2615613 0.1921531 0.2536456 0.2572684 0.2085204
#>  [946] 0.2451435 0.2388195 0.2197342 0.1997267 0.1861565 0.2307367 0.2587420
#>  [953] 0.2243660 0.2508279 0.2581577 0.2049405 0.2441315 0.2130906 0.2505837
#>  [960] 0.2446023 0.2127359 0.2259061 0.1913900 0.2492523 0.2452090 0.2448753
#>  [967] 0.2478824 0.2382320 0.2088942 0.1954502 0.2256387 0.2329501 0.2651193
#>  [974] 0.2674216 0.2196285 0.2343386 0.2052326 0.2554373 0.2267345 0.2237119
#>  [981] 0.2461754 0.2760164 0.2638061 0.2005586 0.2262684 0.2725062 0.2009649
#>  [988] 0.2312145 0.1922929 0.1888379 0.2113077 0.2489326 0.2395515 0.2315986
#>  [995] 0.2534421 0.2475331 0.2665770 0.1976765 0.2253189 0.2678548