Mahalanobis Depth

Computes the mahalanobis depth of a point or vectors of points with respect to a multivariate data set.

depthMah(u, X, cov = NULL, mean = NULL, threads = -1)

Arguments

u
Numerical vector or matrix whose depth is to be calculated. Dimension has to be the same as that of the observations.

X
The data as a matrix, data frame or list. If it is a matrix or data frame, then each row is viewed as one multivariate observation. If it is a list, all components must be numerical vectors of equal length (coordinates of observations).

cov
custom covariance matrix passed. If NULL standard calculations will be based on standard covariance estimator.

mean
custom mean vector. If null --- mean average will be used.

threads
number of threads used in parallel computations. Default value -1 means that all possible cores will be used.

Details

Calculation of Mahalanobis depth is exact.

Returns the depth of multivariate point u with respect to data set X.

Examples

x <- matrix(rnorm(9999), nc = 3)
depthMah(x, x)

#> Depth method:  Mahalanobis 
#>    [1] 0.36257565 0.29801168 0.13298877 0.50808837 0.45834294 0.70441739
#>    [7] 0.20899556 0.26328488 0.24505270 0.19268268 0.42450391 0.23413926
#>   [13] 0.16892663 0.10131757 0.17759627 0.50829544 0.10002971 0.49888705
#>   [19] 0.32454049 0.27594933 0.27416846 0.27540634 0.43357483 0.27951678
#>   [25] 0.15209978 0.49770950 0.17613094 0.09013626 0.17417079 0.36536863
#>   [31] 0.22152585 0.38748426 0.19198029 0.14478600 0.25995205 0.80820667
#>   [37] 0.14627579 0.29056162 0.23120174 0.27686515 0.57846864 0.20092073
#>   [43] 0.19383625 0.17029593 0.24378903 0.57326675 0.62120473 0.76136451
#>   [49] 0.37958772 0.42936371 0.44157052 0.17075733 0.52492002 0.82173722
#>   [55] 0.12597461 0.14498414 0.66524052 0.24741175 0.26392480 0.56224342
#>   [61] 0.23156635 0.78344926 0.16901943 0.46447091 0.20168376 0.50326881
#>   [67] 0.24222449 0.26470675 0.28444646 0.13812382 0.15611300 0.70143619
#>   [73] 0.43548022 0.17544466 0.25893206 0.17087068 0.20368780 0.54070565
#>   [79] 0.17022109 0.16356168 0.18374324 0.31064217 0.19229417 0.67767250
#>   [85] 0.21688290 0.24538615 0.08108388 0.23602196 0.17187305 0.23124803
#>   [91] 0.13527746 0.64111517 0.29609881 0.31529821 0.17563967 0.18347169
#>   [97] 0.47979675 0.44309465 0.41965127 0.10815103 0.24459977 0.58675121
#>  [103] 0.23947458 0.14800120 0.30253706 0.56597223 0.19604588 0.28656681
#>  [109] 0.93475230 0.31883560 0.30755478 0.32499364 0.42412276 0.19979464
#>  [115] 0.23281654 0.25291425 0.22920162 0.41764101 0.64489505 0.21914724
#>  [121] 0.16162782 0.52275905 0.19033737 0.39007603 0.38067198 0.12755294
#>  [127] 0.35924385 0.26461101 0.15068912 0.44894346 0.24311314 0.64172274
#>  [133] 0.35311092 0.19323270 0.38067589 0.31838849 0.16798189 0.52319406
#>  [139] 0.10533575 0.76057136 0.35340647 0.08571077 0.35564759 0.61825371
#>  [145] 0.38629314 0.28644703 0.57983621 0.17835882 0.45760155 0.54969107
#>  [151] 0.61724539 0.80131336 0.52609645 0.20563905 0.34195097 0.17703991
#>  [157] 0.16752517 0.76114982 0.34558912 0.31356694 0.24801243 0.44886783
#>  [163] 0.19455947 0.23964277 0.34360524 0.19357813 0.35307354 0.16136351
#>  [169] 0.52212489 0.34646689 0.13333306 0.18912103 0.09319821 0.19089889
#>  [175] 0.42491575 0.23482636 0.30479828 0.45258670 0.27275945 0.23715134
#>  [181] 0.41056112 0.61573917 0.23263246 0.31264175 0.31376148 0.84265949
#>  [187] 0.82563220 0.24813893 0.15742065 0.83358234 0.32794625 0.28998362
#>  [193] 0.26109053 0.13561947 0.20005924 0.49948080 0.31363066 0.13703947
#>  [199] 0.28170054 0.32575214 0.25628272 0.21041569 0.13518954 0.23921977
#>  [205] 0.18271329 0.24637137 0.20156406 0.16886221 0.88066820 0.48023728
#>  [211] 0.37722795 0.20621163 0.14021360 0.12467999 0.20857779 0.16195445
#>  [217] 0.17974715 0.16407567 0.13326823 0.35866344 0.37823679 0.39065707
#>  [223] 0.14310524 0.23810541 0.34255566 0.41182426 0.19803773 0.24209996
#>  [229] 0.51619531 0.32184749 0.34142118 0.20273923 0.76746432 0.12817364
#>  [235] 0.42444881 0.71554839 0.26216724 0.80955237 0.39784103 0.25569006
#>  [241] 0.81738595 0.11577289 0.28403804 0.48665019 0.23148179 0.27501394
#>  [247] 0.52814004 0.34708060 0.54673847 0.15979723 0.08707705 0.33376347
#>  [253] 0.39753234 0.43529452 0.53498445 0.71984405 0.97583110 0.36998587
#>  [259] 0.19691160 0.17803160 0.15951955 0.27281565 0.25040108 0.21829704
#>  [265] 0.55682321 0.29139968 0.49679063 0.64457846 0.23982568 0.18184695
#>  [271] 0.34535649 0.48565031 0.36154300 0.11055384 0.17849046 0.12410671
#>  [277] 0.24839680 0.15438781 0.26193968 0.39169619 0.23240792 0.21295966
#>  [283] 0.14185486 0.55164771 0.49937141 0.44442281 0.22756775 0.28192249
#>  [289] 0.09075798 0.33364059 0.19313361 0.34518294 0.58592561 0.19351157
#>  [295] 0.86384695 0.59866313 0.35687405 0.25883313 0.38116918 0.39286922
#>  [301] 0.22058787 0.26173027 0.53364662 0.18215962 0.38599160 0.15162942
#>  [307] 0.55422041 0.23958569 0.34274979 0.23147517 0.75002737 0.23240834
#>  [313] 0.20657446 0.22676690 0.22372271 0.23008712 0.40211448 0.18693703
#>  [319] 0.10730800 0.14870997 0.20846448 0.29100738 0.36844724 0.23468356
#>  [325] 0.95508676 0.47293999 0.55026565 0.59673646 0.39761067 0.60720033
#>  [331] 0.83691729 0.16483414 0.13409539 0.93490969 0.46798679 0.18260457
#>  [337] 0.51776493 0.11645663 0.63198758 0.48214867 0.28385599 0.34711467
#>  [343] 0.47540917 0.40995353 0.29091634 0.45770024 0.21649499 0.21633073
#>  [349] 0.38399085 0.36387749 0.22205997 0.16564171 0.36070739 0.76436634
#>  [355] 0.42705098 0.07348674 0.86873965 0.57105093 0.26900137 0.38603317
#>  [361] 0.22440786 0.29240822 0.70820700 0.11784583 0.42919443 0.24297914
#>  [367] 0.64989765 0.21856405 0.34851769 0.20925412 0.23058579 0.65261858
#>  [373] 0.13031490 0.20114142 0.43013705 0.25707192 0.26167777 0.12532675
#>  [379] 0.18248674 0.28504026 0.82095224 0.40048059 0.92755068 0.13841545
#>  [385] 0.23606750 0.82017722 0.33964191 0.23940783 0.22051735 0.25264853
#>  [391] 0.24337059 0.18248431 0.36449666 0.35780192 0.12572330 0.35436858
#>  [397] 0.95673180 0.15634123 0.12648303 0.22547715 0.23696535 0.14191338
#>  [403] 0.37949561 0.27385802 0.12431176 0.66879523 0.17538418 0.11476111
#>  [409] 0.20672757 0.26876294 0.44247522 0.34752296 0.41401574 0.46906101
#>  [415] 0.21880242 0.73141381 0.28057789 0.13888151 0.69427735 0.47123496
#>  [421] 0.48182991 0.41713159 0.36236062 0.66606315 0.38720847 0.11491279
#>  [427] 0.39612314 0.51275907 0.34811917 0.48443896 0.24303786 0.43990468
#>  [433] 0.60562826 0.12085812 0.12260152 0.12389355 0.30592003 0.20480535
#>  [439] 0.57867614 0.23545277 0.33783771 0.23488139 0.67918485 0.25439450
#>  [445] 0.11301299 0.82558227 0.34582847 0.43921235 0.33141644 0.18362598
#>  [451] 0.92390127 0.34549021 0.32858412 0.32420755 0.16978940 0.60454769
#>  [457] 0.33854400 0.13937781 0.37226204 0.57525373 0.35141153 0.25969624
#>  [463] 0.73506925 0.12063782 0.83859899 0.60521055 0.38628672 0.26441225
#>  [469] 0.64745677 0.26633433 0.49751913 0.71547773 0.36807499 0.64394351
#>  [475] 0.15439190 0.29238989 0.28332768 0.15583094 0.41422116 0.08735200
#>  [481] 0.14173836 0.16072452 0.09741659 0.41515718 0.16833500 0.12977662
#>  [487] 0.61524966 0.27464536 0.15374620 0.43264492 0.17600498 0.22238866
#>  [493] 0.25056862 0.32037050 0.16180275 0.23794954 0.52696115 0.09602910
#>  [499] 0.11944861 0.28889884 0.20443689 0.19596304 0.15925807 0.34764194
#>  [505] 0.35809785 0.54876256 0.45377992 0.23623571 0.45739566 0.61139053
#>  [511] 0.91023579 0.19413473 0.12316326 0.32713663 0.32531886 0.16705748
#>  [517] 0.34261570 0.19751957 0.58264346 0.29480251 0.34247028 0.38990212
#>  [523] 0.54243381 0.67044009 0.26792584 0.39170890 0.33645984 0.09682084
#>  [529] 0.27025960 0.13973673 0.10276955 0.42802782 0.39458014 0.32829506
#>  [535] 0.49658419 0.11856171 0.24828218 0.42126152 0.39925600 0.20671104
#>  [541] 0.18859109 0.39785143 0.21305261 0.29439614 0.25090970 0.40706293
#>  [547] 0.12446395 0.35815563 0.25130237 0.11536459 0.20859919 0.27088496
#>  [553] 0.33114753 0.47891747 0.20851355 0.10759712 0.34423414 0.31804194
#>  [559] 0.99826160 0.85112228 0.30856176 0.30552350 0.37221481 0.12880224
#>  [565] 0.54898780 0.28451651 0.35874650 0.12057782 0.11903771 0.17336187
#>  [571] 0.16522737 0.09832215 0.11110429 0.49372774 0.19340324 0.20386117
#>  [577] 0.18644027 0.64808394 0.21079176 0.17429010 0.31976154 0.17458940
#>  [583] 0.22684741 0.21524121 0.26825973 0.15730239 0.12133369 0.33153071
#>  [589] 0.52028450 0.20033691 0.21530900 0.11239106 0.20333394 0.40975069
#>  [595] 0.23643169 0.33901494 0.26283156 0.98597393 0.27075845 0.17141165
#>  [601] 0.47840710 0.35842435 0.11256513 0.23032305 0.73519628 0.22071604
#>  [607] 0.11167548 0.16595069 0.39853215 0.55284810 0.13132466 0.52686465
#>  [613] 0.19692873 0.30611435 0.11136037 0.14809090 0.12753660 0.19787356
#>  [619] 0.38651286 0.24978520 0.36765164 0.37480607 0.18705621 0.28320690
#>  [625] 0.08702136 0.35739703 0.23443814 0.08370271 0.08597304 0.16124976
#>  [631] 0.17835084 0.36122266 0.28282733 0.34115822 0.33781511 0.52557533
#>  [637] 0.16845122 0.36337727 0.61921259 0.36039814 0.10196192 0.14544809
#>  [643] 0.63797700 0.48033945 0.81399499 0.29393646 0.25710601 0.34518192
#>  [649] 0.11762041 0.15375441 0.75469605 0.31908705 0.15779247 0.24497789
#>  [655] 0.31655384 0.26256903 0.28444385 0.11790429 0.23285425 0.44721232
#>  [661] 0.24577958 0.18689075 0.29043977 0.40681433 0.70524293 0.42471996
#>  [667] 0.46633691 0.24949606 0.34524854 0.19758384 0.15060707 0.16155264
#>  [673] 0.59345902 0.20999002 0.55103700 0.28243977 0.50641005 0.93450426
#>  [679] 0.31207526 0.46780673 0.18389214 0.09220222 0.37802826 0.12088222
#>  [685] 0.21742693 0.17184499 0.60024658 0.58115808 0.19167249 0.10742399
#>  [691] 0.23964507 0.66786985 0.91177856 0.18998178 0.12897319 0.16461969
#>  [697] 0.43501994 0.32180834 0.36450549 0.32142915 0.27429995 0.25056416
#>  [703] 0.16958560 0.11245947 0.14117520 0.27799061 0.73118499 0.19798936
#>  [709] 0.35340819 0.36174021 0.46840564 0.65393618 0.24134098 0.52550314
#>  [715] 0.61425484 0.44237902 0.82627658 0.11286829 0.09990663 0.31705792
#>  [721] 0.52436146 0.25163076 0.32582003 0.39635035 0.43536612 0.13719237
#>  [727] 0.20047913 0.39456835 0.23460758 0.73650524 0.11788156 0.64169515
#>  [733] 0.18911004 0.39482104 0.10806473 0.43177517 0.31052170 0.29350830
#>  [739] 0.12333054 0.52845135 0.19727813 0.38194768 0.15827141 0.96518495
#>  [745] 0.23123244 0.44095754 0.30181388 0.81234713 0.29702796 0.29057213
#>  [751] 0.20658602 0.09833810 0.33392749 0.19879241 0.08809939 0.19798966
#>  [757] 0.53661888 0.13346544 0.65479821 0.15142346 0.23931256 0.16704141
#>  [763] 0.15522218 0.40098397 0.43672914 0.24126027 0.29868842 0.41912375
#>  [769] 0.42834698 0.47416532 0.42485000 0.37886670 0.56098647 0.17726268
#>  [775] 0.35077148 0.80829770 0.26294047 0.15305741 0.23575161 0.61451374
#>  [781] 0.51104056 0.48421362 0.57289674 0.23663973 0.55119757 0.25495123
#>  [787] 0.41509937 0.28898490 0.15210299 0.87769772 0.45390349 0.15617666
#>  [793] 0.17401928 0.57169386 0.10465055 0.19199528 0.38636882 0.29097443
#>  [799] 0.53030435 0.43564424 0.31312396 0.83900541 0.20606909 0.41923809
#>  [805] 0.14967259 0.81358333 0.61081111 0.29010449 0.34503188 0.48373061
#>  [811] 0.09122133 0.49833019 0.32742327 0.43861415 0.30956327 0.17290076
#>  [817] 0.26675380 0.46660913 0.26635267 0.39646221 0.73239041 0.40175138
#>  [823] 0.16873422 0.30579070 0.06398197 0.43645785 0.61715397 0.21106300
#>  [829] 0.30421720 0.16734173 0.34994544 0.42207439 0.16946352 0.30232993
#>  [835] 0.10343049 0.31595733 0.30850047 0.36720792 0.18648536 0.78896187
#>  [841] 0.87914813 0.26788946 0.45042302 0.19088496 0.19934423 0.29160375
#>  [847] 0.65629783 0.13083182 0.14737838 0.36880958 0.11497315 0.34713318
#>  [853] 0.58170257 0.24926385 0.73590127 0.51517468 0.31690508 0.12600043
#>  [859] 0.25616225 0.35728196 0.15761124 0.11518703 0.33382375 0.61290726
#>  [865] 0.46643200 0.13976443 0.78200868 0.34769896 0.22505155 0.36783134
#>  [871] 0.45099955 0.39748730 0.30367994 0.26297470 0.46889075 0.13763716
#>  [877] 0.91225656 0.18262830 0.16481633 0.44711891 0.50414041 0.36788973
#>  [883] 0.29554211 0.23372222 0.78885930 0.35704603 0.61954036 0.29521403
#>  [889] 0.19754579 0.34786307 0.23705355 0.13622788 0.12570962 0.26983630
#>  [895] 0.41908144 0.31734123 0.55895418 0.21429763 0.13213823 0.15126985
#>  [901] 0.24175920 0.31166972 0.58135058 0.46607410 0.23766098 0.41366006
#>  [907] 0.20680962 0.37536909 0.27568431 0.29050837 0.17297358 0.16177979
#>  [913] 0.15009809 0.39958563 0.77388230 0.22526412 0.33946169 0.27742960
#>  [919] 0.60451626 0.26014291 0.32129645 0.17949169 0.28141023 0.12102724
#>  [925] 0.44611280 0.50652027 0.14005330 0.15140887 0.30366792 0.39817951
#>  [931] 0.09136358 0.53392768 0.16667049 0.26859375 0.68938961 0.29612755
#>  [937] 0.29289801 0.63211692 0.71401059 0.59922202 0.36481382 0.58660982
#>  [943] 0.22349152 0.45047976 0.15662951 0.29430136 0.46168738 0.23325059
#>  [949] 0.44905567 0.34634447 0.19901638 0.24758089 0.48563933 0.38077640
#>  [955] 0.44613437 0.57731407 0.62179969 0.33827901 0.24707308 0.19999058
#>  [961] 0.72718845 0.06963462 0.60749426 0.19538236 0.73942763 0.48598516
#>  [967] 0.51318895 0.52531076 0.10821907 0.34208858 0.45762877 0.78518968
#>  [973] 0.55550496 0.39563325 0.64622209 0.18686261 0.16296981 0.24002140
#>  [979] 0.52392190 0.55402936 0.17737805 0.21898387 0.29428688 0.10817443
#>  [985] 0.20633922 0.12245471 0.78490261 0.20684277 0.35727523 0.17455340
#>  [991] 0.18833970 0.29202387 0.28834922 0.23875712 0.15326266 0.56912119
#>  [997] 0.17992012 0.58018895 0.82858026 0.66396584 0.59148782 0.69278691
#> [1003] 0.20102632 0.31790262 0.29429058 0.19419470 0.47049445 0.63818556
#> [1009] 0.39740399 0.26505510 0.13033919 0.37008159 0.24636045 0.26653754
#> [1015] 0.11306063 0.21139528 0.30371937 0.81331095 0.19847428 0.41339054
#> [1021] 0.43588410 0.37946883 0.18525827 0.51431784 0.30173062 0.17607114
#> [1027] 0.20064153 0.28781778 0.39548765 0.70462261 0.78799545 0.34820272
#> [1033] 0.39849638 0.31184623 0.11190649 0.15082662 0.17727618 0.22749690
#> [1039] 0.19661468 0.26248160 0.16121374 0.25260051 0.61615765 0.20502085
#> [1045] 0.11319191 0.21206245 0.25008235 0.56757792 0.33835188 0.77230181
#> [1051] 0.42391057 0.16172579 0.34473224 0.29627043 0.22473728 0.75643956
#> [1057] 0.51526865 0.27935415 0.28461997 0.69747391 0.34069317 0.23693393
#> [1063] 0.39354706 0.39989306 0.46423006 0.24000697 0.22753013 0.24027596
#> [1069] 0.37573398 0.41184548 0.22766052 0.23546411 0.79307920 0.37478478
#> [1075] 0.44232760 0.15423995 0.29014607 0.90142863 0.07230573 0.28765172
#> [1081] 0.52520883 0.71922856 0.12297980 0.33427106 0.23258964 0.37542696
#> [1087] 0.22164316 0.56603366 0.24322617 0.42864103 0.16856606 0.20148929
#> [1093] 0.37208788 0.27722357 0.82284623 0.16176763 0.33916833 0.33144594
#> [1099] 0.15048127 0.32004222 0.35144239 0.18309140 0.25711191 0.16861738
#> [1105] 0.36396062 0.61742227 0.30667329 0.45358032 0.26318869 0.43396989
#> [1111] 0.41569768 0.16848834 0.49193257 0.27912569 0.34966229 0.30255714
#> [1117] 0.45453208 0.85784198 0.19640269 0.24425321 0.63126903 0.32225565
#> [1123] 0.49153830 0.22671335 0.35560638 0.88834771 0.35176483 0.18811464
#> [1129] 0.40892411 0.31288895 0.42493283 0.48875821 0.21617669 0.41916999
#> [1135] 0.18807973 0.60283789 0.28385030 0.26524085 0.23224189 0.09956630
#> [1141] 0.17690519 0.22140752 0.39551456 0.13984530 0.50676227 0.70559683
#> [1147] 0.21389458 0.41180894 0.37130020 0.09646709 0.31348635 0.06374675
#> [1153] 0.15954823 0.31534748 0.42570296 0.45712414 0.14352344 0.57529284
#> [1159] 0.29649759 0.21927873 0.65111876 0.18121822 0.12607661 0.72661513
#> [1165] 0.14565988 0.25530414 0.18802172 0.34963454 0.57959960 0.38522664
#> [1171] 0.95813472 0.18742575 0.61379521 0.28116693 0.29487591 0.18272077
#> [1177] 0.37213032 0.18489033 0.15304596 0.32403321 0.15029041 0.23848479
#> [1183] 0.71316925 0.18292077 0.16513342 0.55253268 0.28788555 0.20242029
#> [1189] 0.21608351 0.22533008 0.26386183 0.12684354 0.29622243 0.27455417
#> [1195] 0.31622104 0.44018364 0.16863619 0.23490903 0.55868371 0.69155740
#> [1201] 0.69070916 0.44923482 0.45866875 0.47141796 0.10013196 0.25902691
#> [1207] 0.48551117 0.38545237 0.31676428 0.73463062 0.10907061 0.32816230
#> [1213] 0.27530400 0.39898537 0.54612997 0.23861483 0.34576203 0.45671648
#> [1219] 0.34168386 0.31489732 0.85608227 0.58689631 0.50916077 0.09585686
#> [1225] 0.87003376 0.09551344 0.14256938 0.13681873 0.18368573 0.19899611
#> [1231] 0.64649492 0.42390473 0.23808854 0.45714034 0.27438944 0.35753573
#> [1237] 0.26401846 0.51562229 0.20958021 0.23190208 0.28022324 0.24300882
#> [1243] 0.56549858 0.16790378 0.31248222 0.35589709 0.32008623 0.34336098
#> [1249] 0.29691790 0.22854267 0.27162474 0.25959844 0.13142225 0.79501003
#> [1255] 0.61186941 0.19176861 0.27120800 0.41759832 0.56382617 0.16175240
#> [1261] 0.57165629 0.30676987 0.07636733 0.70694998 0.64337667 0.18392228
#> [1267] 0.31710416 0.44371729 0.13335959 0.27899457 0.34699214 0.14372436
#> [1273] 0.39856265 0.15036932 0.12424525 0.22719450 0.05975808 0.47489797
#> [1279] 0.23519253 0.65596623 0.14242642 0.40827341 0.09536907 0.58015218
#> [1285] 0.43117350 0.72182211 0.37759944 0.61542930 0.79196657 0.53397662
#> [1291] 0.71214881 0.14807741 0.33081649 0.17724493 0.51685700 0.17602548
#> [1297] 0.46056003 0.09469858 0.19878310 0.34858641 0.42210734 0.69021213
#> [1303] 0.13918528 0.21479145 0.42442342 0.55447988 0.83850879 0.28800368
#> [1309] 0.48090300 0.31110275 0.25294878 0.27542331 0.46615365 0.32497608
#> [1315] 0.20510471 0.14807316 0.13836791 0.28604292 0.49618839 0.72071479
#> [1321] 0.27409744 0.30771067 0.44460559 0.12077719 0.16117783 0.38294803
#> [1327] 0.21124549 0.36307137 0.46548362 0.19461635 0.40455633 0.35508187
#> [1333] 0.52185750 0.42604490 0.61612276 0.73710762 0.82167708 0.33734619
#> [1339] 0.24057244 0.69366368 0.79009055 0.41065297 0.17384445 0.35576361
#> [1345] 0.19978788 0.10577908 0.40326161 0.54893937 0.38814144 0.65435749
#> [1351] 0.11066982 0.20645429 0.18391022 0.19271831 0.33136333 0.27163407
#> [1357] 0.18560745 0.55898874 0.49517087 0.93981856 0.48527043 0.51854201
#> [1363] 0.45157969 0.20214520 0.08392120 0.56513501 0.55323233 0.64561414
#> [1369] 0.20718339 0.22555544 0.49690249 0.33301503 0.06917971 0.37799719
#> [1375] 0.32497242 0.15348893 0.26558782 0.33024391 0.10831499 0.12020060
#> [1381] 0.96665008 0.21177369 0.84626676 0.28240724 0.21941940 0.30103704
#> [1387] 0.24645845 0.10863284 0.58945146 0.19549785 0.06541092 0.18440210
#> [1393] 0.15278491 0.11790263 0.31825212 0.43978164 0.23740703 0.20608451
#> [1399] 0.13051184 0.34970258 0.30933704 0.50349787 0.70792404 0.28798589
#> [1405] 0.28037652 0.34010186 0.42208194 0.84384681 0.27343253 0.30883375
#> [1411] 0.79679554 0.15150033 0.12639056 0.12389731 0.67056279 0.39501715
#> [1417] 0.17863388 0.43350618 0.14750260 0.14105777 0.28686528 0.23022402
#> [1423] 0.38971665 0.19894933 0.46973040 0.42164618 0.15151674 0.19998788
#> [1429] 0.34212743 0.48603267 0.49363032 0.49963888 0.28464283 0.31762319
#> [1435] 0.24063166 0.86206901 0.19583946 0.49019551 0.08917273 0.66268229
#> [1441] 0.23068204 0.25771032 0.39573224 0.42035105 0.16053234 0.31507457
#> [1447] 0.37498975 0.21924024 0.29336332 0.14716689 0.13680529 0.31766106
#> [1453] 0.12410750 0.32937711 0.13758596 0.32772587 0.50115840 0.58678278
#> [1459] 0.14422580 0.50674902 0.19995658 0.48140076 0.33877957 0.37607946
#> [1465] 0.75506345 0.47292041 0.47403497 0.29773658 0.35537050 0.71051145
#> [1471] 0.26418765 0.79009695 0.73870116 0.41822082 0.19696215 0.49137848
#> [1477] 0.71669671 0.43437311 0.80684287 0.53562510 0.41174881 0.12932910
#> [1483] 0.17667614 0.36029958 0.42825994 0.37948099 0.39805912 0.14802009
#> [1489] 0.60453917 0.63728402 0.80242379 0.21180206 0.46938260 0.75484692
#> [1495] 0.41672057 0.31745605 0.18614094 0.23583109 0.29460644 0.27294532
#> [1501] 0.50714575 0.36504411 0.81646397 0.36335089 0.55633043 0.33706444
#> [1507] 0.11634127 0.31926286 0.25317632 0.20189293 0.43436368 0.55711028
#> [1513] 0.30608536 0.15759470 0.51968865 0.17824887 0.14031019 0.27150412
#> [1519] 0.93936725 0.11451402 0.64859282 0.17714971 0.23354053 0.25721017
#> [1525] 0.23255110 0.43950516 0.34693078 0.52012865 0.10837821 0.28742352
#> [1531] 0.16272503 0.31350074 0.32972447 0.19248375 0.16670552 0.26151535
#> [1537] 0.58327367 0.06947607 0.31521109 0.68728499 0.31464720 0.23944552
#> [1543] 0.16710972 0.25707023 0.82863918 0.32105100 0.29061595 0.39322902
#> [1549] 0.46574686 0.55275929 0.20458012 0.55065705 0.34365659 0.10123866
#> [1555] 0.33059062 0.63229523 0.25960287 0.41843599 0.14637464 0.34955482
#> [1561] 0.60130772 0.05736401 0.12876187 0.42874853 0.31099390 0.22620633
#> [1567] 0.11765694 0.49084538 0.36025857 0.43857272 0.81365427 0.21088493
#> [1573] 0.46209931 0.38744144 0.44586441 0.72148760 0.22796256 0.29481344
#> [1579] 0.35888253 0.10082167 0.36740205 0.53226034 0.44672572 0.14994041
#> [1585] 0.17743287 0.38559291 0.42968296 0.43127327 0.21877627 0.35802588
#> [1591] 0.33747163 0.69710364 0.19573745 0.32161552 0.41754489 0.27209404
#> [1597] 0.46858054 0.35166421 0.49393248 0.78392557 0.39758537 0.33602724
#> [1603] 0.28005233 0.40824235 0.72879849 0.30303536 0.48536581 0.44077155
#> [1609] 0.24733897 0.35365058 0.41317262 0.35332860 0.65377399 0.21097660
#> [1615] 0.41283711 0.23683832 0.20566326 0.43255955 0.56084548 0.21482652
#> [1621] 0.59512145 0.20267798 0.54880942 0.21443003 0.76321235 0.20116309
#> [1627] 0.70153897 0.19010793 0.31326400 0.79348293 0.49564939 0.27884336
#> [1633] 0.31274711 0.19628127 0.53472175 0.20401752 0.40237033 0.46650369
#> [1639] 0.31730356 0.45477827 0.49568316 0.09870759 0.23140408 0.56438481
#> [1645] 0.25426958 0.31239910 0.47537421 0.71599224 0.18508435 0.09989029
#> [1651] 0.44131605 0.23690750 0.88230384 0.18612057 0.32421172 0.64112785
#> [1657] 0.68845797 0.19195738 0.14345233 0.60231182 0.32496588 0.42826304
#> [1663] 0.31560895 0.26500482 0.28189518 0.27221833 0.22685549 0.17000868
#> [1669] 0.21951001 0.20540704 0.30506978 0.75451708 0.65417540 0.45824164
#> [1675] 0.62627544 0.23484509 0.40797387 0.15283010 0.35573497 0.33978564
#> [1681] 0.15840307 0.81545018 0.28569715 0.40269497 0.13461655 0.13617026
#> [1687] 0.46121798 0.38062375 0.22442880 0.46172926 0.25540463 0.23831462
#> [1693] 0.25530731 0.80294301 0.17123727 0.31619528 0.36509377 0.44242020
#> [1699] 0.31522652 0.40524394 0.29512755 0.34513077 0.58238499 0.15627893
#> [1705] 0.43086224 0.16028965 0.12962024 0.10498939 0.15519666 0.12518129
#> [1711] 0.33055221 0.42443861 0.24783712 0.22904802 0.09718057 0.11885289
#> [1717] 0.39722761 0.47727948 0.29666914 0.27502729 0.25678134 0.33095685
#> [1723] 0.26588335 0.14007932 0.53669277 0.29519315 0.31256031 0.15099601
#> [1729] 0.24052471 0.09873489 0.13758841 0.09769798 0.50990711 0.37307212
#> [1735] 0.49405130 0.27474627 0.18925657 0.47997147 0.30319213 0.17212657
#> [1741] 0.57714914 0.52951005 0.27443784 0.44699529 0.23777346 0.35238578
#> [1747] 0.91210101 0.19543061 0.20518087 0.21241528 0.50381021 0.37356271
#> [1753] 0.49139915 0.25656658 0.32442005 0.15874411 0.58867784 0.35145946
#> [1759] 0.49239673 0.12920081 0.32592161 0.14868938 0.22244643 0.59919808
#> [1765] 0.09037169 0.16166542 0.35753002 0.27189296 0.24445748 0.53206074
#> [1771] 0.20589896 0.21082084 0.32997005 0.26790574 0.45094042 0.13505261
#> [1777] 0.28127669 0.61690385 0.35809760 0.23910494 0.41628262 0.20748147
#> [1783] 0.31753720 0.13962666 0.30531638 0.11020439 0.28601742 0.31001123
#> [1789] 0.40885903 0.53316186 0.26348937 0.12168577 0.30123942 0.37852830
#> [1795] 0.14163268 0.16257964 0.29516529 0.27855718 0.18078210 0.15568978
#> [1801] 0.56689363 0.19731996 0.13410001 0.23823006 0.71827131 0.32789147
#> [1807] 0.10217109 0.30525922 0.17087312 0.19229563 0.15042373 0.25845281
#> [1813] 0.09216000 0.63355307 0.80431263 0.20922082 0.48780808 0.72691131
#> [1819] 0.15549795 0.30028660 0.23863736 0.17627474 0.45803915 0.14650250
#> [1825] 0.12300966 0.17065179 0.16345330 0.38155340 0.20609850 0.10374335
#> [1831] 0.23621857 0.37199248 0.44994966 0.47348505 0.83928331 0.19501720
#> [1837] 0.31720950 0.18341795 0.16423940 0.14624100 0.22771571 0.34330694
#> [1843] 0.18336467 0.42317679 0.89280478 0.10652257 0.43968707 0.15616261
#> [1849] 0.26274652 0.28601832 0.52677546 0.41608862 0.20068537 0.40097694
#> [1855] 0.17265820 0.59431760 0.22880831 0.23048014 0.43534845 0.58464854
#> [1861] 0.52152497 0.29819101 0.29151525 0.19544883 0.94098300 0.30139092
#> [1867] 0.25162024 0.09896907 0.13861087 0.13183808 0.46038906 0.12386983
#> [1873] 0.13078875 0.48872525 0.78652135 0.34530482 0.16166365 0.91293978
#> [1879] 0.55565907 0.09082363 0.25672218 0.15014868 0.53913721 0.13119218
#> [1885] 0.27482209 0.44976087 0.31069500 0.32696385 0.29092238 0.09726876
#> [1891] 0.35628103 0.25657788 0.12044451 0.24600917 0.21846864 0.41870196
#> [1897] 0.66240414 0.37451421 0.52945489 0.14880839 0.10385516 0.15030696
#> [1903] 0.09638913 0.23596511 0.15089073 0.13579420 0.88649348 0.13720349
#> [1909] 0.36288244 0.20089215 0.14879615 0.64162423 0.56306657 0.25844143
#> [1915] 0.37543396 0.65506356 0.34200801 0.17568001 0.47717296 0.22700697
#> [1921] 0.17889637 0.21848550 0.22784401 0.11608961 0.29158431 0.28456675
#> [1927] 0.26926357 0.14180873 0.16270059 0.33555892 0.72784704 0.81008715
#> [1933] 0.10135780 0.49594947 0.22716011 0.84789152 0.11755749 0.20029208
#> [1939] 0.28115165 0.19077628 0.45272350 0.36196850 0.24437598 0.16998698
#> [1945] 0.21271912 0.29394604 0.50313041 0.32700058 0.52396795 0.59021531
#> [1951] 0.37639157 0.08700456 0.29356311 0.30030397 0.21228269 0.18911890
#> [1957] 0.40114296 0.30330410 0.53206585 0.50029417 0.21296863 0.39097244
#> [1963] 0.10547876 0.37732911 0.49230994 0.97472021 0.73496846 0.23831225
#> [1969] 0.47123445 0.11819820 0.38323229 0.19866604 0.72580307 0.32887179
#> [1975] 0.64899324 0.20574411 0.41236549 0.20262236 0.30452791 0.13257841
#> [1981] 0.17367528 0.57790870 0.61591248 0.46644740 0.45787150 0.23343535
#> [1987] 0.10933079 0.18352538 0.61565991 0.16775239 0.47000419 0.35882030
#> [1993] 0.50678136 0.27478404 0.56362959 0.14753871 0.22141925 0.36564983
#> [1999] 0.15282519 0.29128558 0.23711566 0.37067954 0.83980103 0.39822972
#> [2005] 0.14796989 0.12665390 0.21699855 0.69715243 0.26491526 0.73090339
#> [2011] 0.60581173 0.24609017 0.16806051 0.15653785 0.20413013 0.49248351
#> [2017] 0.09841603 0.21869389 0.27830280 0.29310992 0.57209069 0.12374463
#> [2023] 0.52902810 0.31589041 0.30548707 0.35587956 0.39687670 0.19791922
#> [2029] 0.53594106 0.33353073 0.28279568 0.42921742 0.14995498 0.15191597
#> [2035] 0.23505018 0.36348315 0.62375664 0.13456288 0.91697086 0.20181104
#> [2041] 0.21429921 0.33059357 0.32989996 0.18760530 0.44598201 0.10079450
#> [2047] 0.89615884 0.16036924 0.33464464 0.12751046 0.40312902 0.18206766
#> [2053] 0.25016535 0.55780934 0.78768452 0.27168150 0.09218584 0.93715791
#> [2059] 0.56211701 0.15346488 0.35851806 0.12208280 0.47900771 0.17264642
#> [2065] 0.27539926 0.73565384 0.15016394 0.11169090 0.26586038 0.49090383
#> [2071] 0.38062793 0.28688182 0.18921619 0.31693012 0.41563820 0.21959017
#> [2077] 0.87032447 0.28312587 0.20296432 0.11890342 0.27725135 0.54817152
#> [2083] 0.27746436 0.61867884 0.12199339 0.09925243 0.36120276 0.38850445
#> [2089] 0.40244352 0.57090691 0.10987655 0.16630621 0.15571413 0.31514415
#> [2095] 0.50166136 0.14471575 0.47933392 0.27891983 0.20663064 0.35625312
#> [2101] 0.19228972 0.87742698 0.75790989 0.23432187 0.34922080 0.13813992
#> [2107] 0.80136626 0.56186072 0.40567151 0.33485130 0.52852851 0.65085947
#> [2113] 0.45097998 0.33486578 0.18472916 0.14755561 0.10566666 0.27315441
#> [2119] 0.31540295 0.78975058 0.12602667 0.25501540 0.27346333 0.28450684
#> [2125] 0.33057446 0.15438518 0.17026037 0.26274078 0.32779915 0.38361442
#> [2131] 0.14641651 0.50381694 0.07339003 0.21533920 0.21144300 0.49589325
#> [2137] 0.38924838 0.47749116 0.13679804 0.15705744 0.20548621 0.43811467
#> [2143] 0.50110277 0.17295755 0.49795652 0.43817399 0.21770106 0.75045504
#> [2149] 0.49961070 0.12255486 0.20684520 0.13152423 0.56840253 0.43032235
#> [2155] 0.18628130 0.29053646 0.09690653 0.13842632 0.15083999 0.31020567
#> [2161] 0.60733858 0.16325855 0.05449016 0.56303666 0.28016206 0.24948483
#> [2167] 0.21729538 0.11663450 0.27167256 0.57091647 0.28109243 0.25002750
#> [2173] 0.47490681 0.82495593 0.17534940 0.22819653 0.22965855 0.32469974
#> [2179] 0.70129520 0.45965065 0.46554277 0.60892951 0.41575900 0.08532936
#> [2185] 0.25096449 0.25281147 0.21074327 0.13102020 0.46135425 0.24425763
#> [2191] 0.11389432 0.33929073 0.12255912 0.38349986 0.26330788 0.09590814
#> [2197] 0.42689785 0.12789323 0.52952973 0.30731013 0.38398721 0.10085228
#> [2203] 0.70693291 0.13653151 0.69149758 0.34333380 0.26738487 0.58531678
#> [2209] 0.61180892 0.26344801 0.13714037 0.18934596 0.26484914 0.27852540
#> [2215] 0.42673524 0.21107473 0.42999162 0.74614452 0.63607253 0.30794873
#> [2221] 0.22437050 0.09174624 0.19959672 0.97066345 0.10784145 0.09886555
#> [2227] 0.12845407 0.44616335 0.12591542 0.11632923 0.50431281 0.57215848
#> [2233] 0.29245373 0.27650657 0.22207460 0.44361397 0.13189923 0.15969982
#> [2239] 0.53662168 0.16705477 0.27770825 0.37298250 0.31224162 0.44720281
#> [2245] 0.36353926 0.12838056 0.32545283 0.16178494 0.13428478 0.19331330
#> [2251] 0.27943270 0.21943754 0.43265838 0.21086632 0.75303036 0.25787003
#> [2257] 0.54596141 0.47746955 0.43455654 0.36652151 0.26157222 0.36466982
#> [2263] 0.23474859 0.15697804 0.38727193 0.41335222 0.21162797 0.32460761
#> [2269] 0.41442127 0.11267531 0.31584867 0.13371113 0.23794307 0.33767364
#> [2275] 0.29130369 0.12149065 0.26357428 0.26233015 0.09836863 0.17608092
#> [2281] 0.15879815 0.30381434 0.17721259 0.17588622 0.09784884 0.25900234
#> [2287] 0.11680705 0.67068398 0.36033322 0.21174971 0.37626540 0.15194723
#> [2293] 0.26460463 0.21701059 0.43523135 0.38604839 0.20273093 0.94128661
#> [2299] 0.10917610 0.22362511 0.29815933 0.08717857 0.09944346 0.21639793
#> [2305] 0.24985552 0.61228402 0.14598615 0.31850370 0.63028809 0.18364342
#> [2311] 0.84820594 0.46470655 0.67028223 0.34678035 0.17518142 0.69438591
#> [2317] 0.38510826 0.11748182 0.27593154 0.52955830 0.16829298 0.41694900
#> [2323] 0.98888874 0.26810530 0.35415203 0.20095507 0.37712271 0.79402674
#> [2329] 0.17516675 0.11377197 0.16694311 0.89953532 0.34240804 0.56917527
#> [2335] 0.76659536 0.26147129 0.61302791 0.51979461 0.21276976 0.28381212
#> [2341] 0.17842204 0.20418027 0.28232601 0.69648892 0.13604094 0.62047221
#> [2347] 0.34960221 0.14392462 0.40308408 0.39483958 0.21065186 0.31198371
#> [2353] 0.15443120 0.22926916 0.63201367 0.58231326 0.38570565 0.09690662
#> [2359] 0.24799975 0.54394750 0.13883548 0.36179096 0.17339294 0.54006613
#> [2365] 0.42684288 0.05542468 0.08718395 0.30613449 0.38697134 0.37394847
#> [2371] 0.28703219 0.84797565 0.40282070 0.17924929 0.17455939 0.22368549
#> [2377] 0.25349933 0.68745285 0.48924365 0.40628956 0.35051622 0.33138560
#> [2383] 0.19335375 0.12543653 0.24998982 0.19764135 0.11194824 0.09637473
#> [2389] 0.24653913 0.08040297 0.10839411 0.42963447 0.65876798 0.83923274
#> [2395] 0.14269255 0.09788742 0.12909693 0.12770082 0.55593166 0.22872809
#> [2401] 0.61009189 0.33352364 0.15555515 0.55249866 0.23413606 0.52384543
#> [2407] 0.59993119 0.43852790 0.12265428 0.14197732 0.57647920 0.29866080
#> [2413] 0.14525629 0.29833139 0.30532482 0.26479963 0.69216317 0.30055628
#> [2419] 0.50465349 0.27380940 0.37145395 0.50465899 0.37896269 0.49659017
#> [2425] 0.30198657 0.12435642 0.29959539 0.49899847 0.16234759 0.15885835
#> [2431] 0.75430017 0.60046777 0.14933783 0.22577065 0.23002259 0.47953310
#> [2437] 0.16424434 0.34395523 0.32913177 0.24470194 0.12179335 0.23827606
#> [2443] 0.39570406 0.37121985 0.23835080 0.13914954 0.47942090 0.12784952
#> [2449] 0.23934613 0.46905394 0.66359619 0.21348076 0.27938156 0.58289203
#> [2455] 0.12182325 0.34064930 0.31264854 0.36857192 0.52945983 0.29353558
#> [2461] 0.72652605 0.64500641 0.51198772 0.15675882 0.59121688 0.20955882
#> [2467] 0.21137153 0.29654970 0.41020971 0.16390062 0.11018985 0.38279958
#> [2473] 0.43827820 0.12992686 0.13848173 0.11690757 0.76791842 0.17786531
#> [2479] 0.25005444 0.60907161 0.45357212 0.41705613 0.47991287 0.54044451
#> [2485] 0.53420050 0.25422721 0.13871022 0.11243124 0.41401443 0.24979507
#> [2491] 0.24962248 0.40378953 0.14797253 0.12831927 0.23588159 0.40847733
#> [2497] 0.49832040 0.27151143 0.67105843 0.66285950 0.36811361 0.69788135
#> [2503] 0.29821898 0.19096995 0.26915323 0.44074237 0.38439179 0.11120126
#> [2509] 0.25908657 0.51004596 0.70802383 0.12193784 0.49783467 0.23924430
#> [2515] 0.57592102 0.27439918 0.29689519 0.17433990 0.57295360 0.71770351
#> [2521] 0.86870385 0.54401668 0.43432488 0.16233716 0.37480053 0.29846736
#> [2527] 0.27703897 0.58574998 0.33233171 0.34620887 0.12240822 0.16543078
#> [2533] 0.54713431 0.62084084 0.17927434 0.44581326 0.27452757 0.38298237
#> [2539] 0.28080435 0.28842113 0.76712593 0.47722020 0.15853614 0.23729791
#> [2545] 0.06931625 0.27022013 0.44037701 0.43914062 0.35994811 0.14495972
#> [2551] 0.34899664 0.14077143 0.24961553 0.27951286 0.41156487 0.18245250
#> [2557] 0.50817628 0.73065671 0.49717420 0.69634997 0.35952780 0.29500771
#> [2563] 0.30662703 0.22757778 0.12267446 0.23881043 0.30069222 0.34605003
#> [2569] 0.28199574 0.32582483 0.20694400 0.11318570 0.23240868 0.60001956
#> [2575] 0.13045769 0.21283882 0.18614586 0.61202291 0.94745331 0.18118066
#> [2581] 0.13013414 0.15771004 0.22453565 0.19576355 0.49384655 0.28258642
#> [2587] 0.47665495 0.13782165 0.17828521 0.23121799 0.57562044 0.18222173
#> [2593] 0.32734959 0.20237006 0.13521427 0.41042866 0.33699535 0.33178836
#> [2599] 0.29294283 0.18359846 0.23248679 0.37626519 0.39534989 0.31258247
#> [2605] 0.09943816 0.36486745 0.36634948 0.29064471 0.31138204 0.33689867
#> [2611] 0.42102681 0.40354452 0.11941780 0.60727531 0.14518939 0.16790137
#> [2617] 0.29580207 0.31568471 0.38459574 0.19128025 0.23009022 0.73002346
#> [2623] 0.43755935 0.23077848 0.24441247 0.47161639 0.16663305 0.14019402
#> [2629] 0.64261964 0.61252464 0.31025584 0.51023147 0.50348288 0.43402588
#> [2635] 0.22888195 0.34621976 0.41889017 0.19488578 0.15795788 0.39948733
#> [2641] 0.20196209 0.11558300 0.54656545 0.33400938 0.40714436 0.78621691
#> [2647] 0.14869282 0.11604929 0.50385234 0.65655520 0.12085729 0.78879445
#> [2653] 0.14593743 0.32034514 0.17437732 0.11388387 0.32113174 0.50313295
#> [2659] 0.35484698 0.21738804 0.48411679 0.46105526 0.40655545 0.08941737
#> [2665] 0.30067529 0.31431868 0.28816129 0.61321073 0.63236278 0.59472464
#> [2671] 0.17780196 0.52775577 0.46322473 0.60658578 0.23235377 0.33599503
#> [2677] 0.30886341 0.20419540 0.29893091 0.51827356 0.18900752 0.15388697
#> [2683] 0.30060083 0.11806384 0.68977559 0.17791283 0.08542881 0.45335915
#> [2689] 0.24788209 0.13536273 0.11924259 0.39258063 0.28942802 0.55227951
#> [2695] 0.39552677 0.07672394 0.26193447 0.26522328 0.22723411 0.29469097
#> [2701] 0.45219582 0.13284786 0.23491323 0.12653165 0.61439923 0.57605935
#> [2707] 0.44145055 0.85778493 0.35325314 0.08930977 0.16860093 0.54736685
#> [2713] 0.70311916 0.18270884 0.57411721 0.14270712 0.40125908 0.19334415
#> [2719] 0.29609040 0.51264093 0.79196891 0.62800969 0.15955014 0.23024385
#> [2725] 0.49230512 0.14471439 0.53349152 0.26422289 0.17250759 0.18908930
#> [2731] 0.28027095 0.60025131 0.14060811 0.35939835 0.26532875 0.24898878
#> [2737] 0.38085894 0.10565258 0.34149149 0.27353799 0.28233849 0.19521504
#> [2743] 0.36741050 0.35604745 0.10007976 0.18844705 0.42475467 0.28539329
#> [2749] 0.13324373 0.39578832 0.14832365 0.36985218 0.21649738 0.44306367
#> [2755] 0.51736573 0.24800382 0.15587694 0.36133880 0.65549526 0.53224470
#> [2761] 0.24680754 0.13421166 0.13285113 0.54618219 0.88532600 0.41628196
#> [2767] 0.18732298 0.22192262 0.35475876 0.51250327 0.30896326 0.50782181
#> [2773] 0.29187903 0.43223289 0.44837105 0.56004452 0.16245878 0.41370598
#> [2779] 0.22521958 0.12517954 0.34927236 0.21207184 0.08990587 0.26961112
#> [2785] 0.23719826 0.71488005 0.12334178 0.28004102 0.33741118 0.32867600
#> [2791] 0.64615699 0.16009970 0.25907902 0.20255363 0.35313839 0.43917620
#> [2797] 0.09673764 0.38279638 0.51515672 0.56845533 0.31484945 0.24024837
#> [2803] 0.23401255 0.66256907 0.34855154 0.11389206 0.34688721 0.75100120
#> [2809] 0.61616632 0.43086031 0.29959656 0.35026498 0.22177201 0.18111811
#> [2815] 0.29107194 0.40295660 0.40081292 0.08768681 0.59681780 0.11384035
#> [2821] 0.16221182 0.27366894 0.19199829 0.23303258 0.37752746 0.14155747
#> [2827] 0.24660040 0.14187436 0.24905017 0.36654065 0.21572441 0.38039243
#> [2833] 0.43305236 0.17103057 0.24848745 0.24239966 0.09799518 0.50467399
#> [2839] 0.49397968 0.20963648 0.43360628 0.13923308 0.57407895 0.24822429
#> [2845] 0.20602185 0.27293637 0.19346455 0.14221334 0.12019700 0.22144696
#> [2851] 0.09902246 0.68830654 0.30990136 0.10665186 0.76576147 0.92040342
#> [2857] 0.30006790 0.13393600 0.39220496 0.12348958 0.38109914 0.36539243
#> [2863] 0.24391118 0.46934219 0.10151865 0.20364086 0.33106871 0.38595578
#> [2869] 0.13276065 0.80327970 0.10921180 0.41880566 0.28667961 0.50701470
#> [2875] 0.24174873 0.81122207 0.10963195 0.30089036 0.32817527 0.21185861
#> [2881] 0.82454822 0.79728845 0.48037370 0.35972573 0.12053832 0.24455806
#> [2887] 0.35789169 0.08627797 0.75545822 0.09635507 0.57332876 0.10500123
#> [2893] 0.26486866 0.48300617 0.35226344 0.33307421 0.22203981 0.35680174
#> [2899] 0.14198158 0.24709936 0.15195836 0.44347248 0.35648753 0.49134889
#> [2905] 0.45106856 0.55258362 0.14642814 0.21771860 0.49914470 0.32029463
#> [2911] 0.32076814 0.15914066 0.28777363 0.47004073 0.22973085 0.24119975
#> [2917] 0.29906194 0.15655439 0.42623962 0.90700474 0.35930897 0.25295653
#> [2923] 0.48013869 0.12370631 0.42447477 0.37110222 0.19082224 0.16725987
#> [2929] 0.51722876 0.36880751 0.32804487 0.20228426 0.32178999 0.13364975
#> [2935] 0.65963946 0.28689961 0.11471430 0.59969373 0.47089261 0.27445050
#> [2941] 0.09984323 0.52819169 0.20080754 0.33050790 0.70643871 0.10843028
#> [2947] 0.64025320 0.10874336 0.20289426 0.26550601 0.58227969 0.13959622
#> [2953] 0.49936507 0.40344991 0.24998695 0.92243388 0.21556145 0.30302617
#> [2959] 0.80026642 0.26556786 0.82582513 0.35834916 0.23681126 0.70810601
#> [2965] 0.34879389 0.25233727 0.30479960 0.24557018 0.25426457 0.24662709
#> [2971] 0.16089547 0.31588728 0.27103425 0.16951811 0.51594352 0.28597205
#> [2977] 0.74233791 0.57920317 0.33472448 0.20999932 0.27654380 0.51299127
#> [2983] 0.40119275 0.77987799 0.25639464 0.45517954 0.30759888 0.28844335
#> [2989] 0.50401463 0.61760518 0.19154233 0.09592599 0.30197045 0.44742517
#> [2995] 0.58693329 0.32240470 0.29892083 0.22617814 0.12369698 0.24898992
#> [3001] 0.33615149 0.44393621 0.28404040 0.45715796 0.19719792 0.29169186
#> [3007] 0.13678568 0.28389735 0.48863068 0.43871377 0.33991800 0.16056149
#> [3013] 0.33271042 0.60163258 0.22724945 0.19716599 0.20385293 0.38402601
#> [3019] 0.56581559 0.12461858 0.33029945 0.10768064 0.57937013 0.86768193
#> [3025] 0.21357726 0.71173937 0.22401959 0.28584409 0.89036051 0.31212596
#> [3031] 0.18967841 0.37688347 0.27339463 0.53790551 0.16941178 0.16108138
#> [3037] 0.20545230 0.20722680 0.37189826 0.59898361 0.18835760 0.53083408
#> [3043] 0.36606690 0.22535136 0.31526942 0.42195085 0.39494005 0.35407292
#> [3049] 0.28839702 0.51690364 0.23848827 0.25350593 0.24020269 0.49451643
#> [3055] 0.54472012 0.26940543 0.95719563 0.71134084 0.22102915 0.41971505
#> [3061] 0.82649490 0.10346551 0.66616486 0.22073708 0.22919574 0.35366142
#> [3067] 0.42628613 0.22639669 0.45918548 0.16747689 0.26838142 0.06651256
#> [3073] 0.14584550 0.54595126 0.23042542 0.30312304 0.46387486 0.40466101
#> [3079] 0.18258205 0.32217994 0.11500887 0.62282141 0.58345705 0.22802205
#> [3085] 0.12429653 0.85039941 0.41265559 0.08978784 0.33501596 0.28603624
#> [3091] 0.22315258 0.39636560 0.24274108 0.74717895 0.25903850 0.21483713
#> [3097] 0.72831903 0.11023573 0.39241206 0.68606647 0.22715149 0.31315616
#> [3103] 0.25838549 0.47441994 0.38260721 0.26610394 0.42435201 0.18369711
#> [3109] 0.61355709 0.11129343 0.88017236 0.24992948 0.34505371 0.33499751
#> [3115] 0.60146518 0.14456827 0.21878103 0.18584516 0.19904656 0.24475096
#> [3121] 0.59605601 0.57874110 0.10450900 0.36430320 0.20690025 0.37000855
#> [3127] 0.68058549 0.21840189 0.31612475 0.21090763 0.36815766 0.20824262
#> [3133] 0.46798165 0.54414249 0.78305896 0.27942965 0.47736460 0.76132505
#> [3139] 0.45591245 0.30544658 0.07745144 0.13131440 0.57582373 0.35963087
#> [3145] 0.12278719 0.84344611 0.37315470 0.52495066 0.12720073 0.29133494
#> [3151] 0.25562897 0.27115073 0.54473845 0.20031505 0.15715355 0.64281197
#> [3157] 0.52160622 0.22458063 0.14438239 0.34403638 0.43143076 0.16752562
#> [3163] 0.37202510 0.65712248 0.15766702 0.12702242 0.55095892 0.17200579
#> [3169] 0.48456805 0.08528133 0.47964454 0.30619419 0.11469113 0.21847033
#> [3175] 0.21483474 0.30171759 0.79074070 0.89334383 0.26481290 0.81897912
#> [3181] 0.54319838 0.31594527 0.28263639 0.19422737 0.19501048 0.12968590
#> [3187] 0.22299743 0.10971483 0.37801854 0.29850638 0.20937508 0.26476470
#> [3193] 0.53687024 0.32740104 0.22080577 0.35619996 0.20038474 0.25796605
#> [3199] 0.24233816 0.30233084 0.34380657 0.19026709 0.50818527 0.24496178
#> [3205] 0.34606894 0.45290594 0.09063125 0.16830632 0.15403194 0.11120773
#> [3211] 0.32372942 0.30661328 0.23910082 0.19984075 0.72590184 0.74835502
#> [3217] 0.14908488 0.61776502 0.10240059 0.57136048 0.36140830 0.18857698
#> [3223] 0.09298892 0.62033555 0.49295605 0.53658879 0.38663101 0.28446208
#> [3229] 0.50696927 0.19671652 0.20082759 0.48948185 0.13509848 0.32895682
#> [3235] 0.18435356 0.45272781 0.09257460 0.18488744 0.18677629 0.18906766
#> [3241] 0.46728200 0.57470335 0.30530975 0.16034313 0.34713985 0.28647141
#> [3247] 0.25594817 0.47414297 0.22781621 0.51145010 0.21542262 0.43143294
#> [3253] 0.24601615 0.26758740 0.46659683 0.31993639 0.57423203 0.95029254
#> [3259] 0.39418771 0.31533936 0.59583006 0.14332384 0.08734141 0.13137254
#> [3265] 0.10743968 0.17244973 0.24518040 0.15145133 0.17491226 0.84856322
#> [3271] 0.42356746 0.35383821 0.39535639 0.38931126 0.38813132 0.20291986
#> [3277] 0.64718402 0.48542613 0.12598409 0.12612165 0.34981242 0.16758190
#> [3283] 0.17339160 0.14223121 0.22304207 0.53630688 0.22876047 0.13182433
#> [3289] 0.20784567 0.23273003 0.13246836 0.18934399 0.12678251 0.39661920
#> [3295] 0.08842457 0.88843827 0.17172273 0.37658052 0.35166260 0.40709724
#> [3301] 0.52573835 0.57875203 0.35178704 0.32336238 0.42043152 0.17949187
#> [3307] 0.48170127 0.08354337 0.79174667 0.86847659 0.67119243 0.25358604
#> [3313] 0.40226011 0.21220112 0.18914171 0.27516947 0.35561168 0.21945399
#> [3319] 0.42438038 0.21320768 0.55894032 0.28661599 0.12256243 0.74307084
#> [3325] 0.55960669 0.41506930 0.45244772 0.28845441 0.28652692 0.47097408
#> [3331] 0.15357840 0.21686485 0.33407919