Tukey Depth

Computes the Tukey depth of a point or vectors of points with respect to a multivariate data set.

depthTukey(u, X, ndir = 1000, threads = -1, exact = FALSE)

Arguments

u
Numerical vector or matrix whose depth is to be calculated. Dimension has to be the same as that of the observations.

X
The data as a matrix, data frame or list. If it is a matrix or data frame, then each row is viewed as one multivariate observation. If it is a list, all components must be numerical vectors of equal length (coordinates of observations).

ndir
number of directions used in computations

threads
number of threads used in parallel computations. Default value -1 means that all possible cores will be used.

exact
if TRUE exact alhorithm will be used . Currently it works only for 2 dimensional data set.

…
currently not supported.

Details

Irrespective of dimension, Projection and Tukey's depth is obtained by approximate calculation.

Returns the depth of multivariate point u with respect to data set X.

Examples

# NOT RUN {
x <- matrix(rnorm(3000), nc = 3)
depthTukey(x, ndir = 2000)
# }

# Exact algorithm in 2d
x <- matrix(rnorm(2000), nc = 2)
depthTukey(x, exact = TRUE)

#> Depth method:  Tukey 
#>    [1] 0.066 0.386 0.106 0.113 0.010 0.391 0.422 0.094 0.242 0.002 0.224 0.217
#>   [13] 0.409 0.007 0.073 0.014 0.076 0.095 0.133 0.074 0.285 0.039 0.034 0.140
#>   [25] 0.005 0.230 0.103 0.030 0.182 0.094 0.052 0.016 0.150 0.110 0.252 0.109
#>   [37] 0.044 0.288 0.099 0.199 0.277 0.004 0.144 0.062 0.142 0.104 0.233 0.143
#>   [49] 0.100 0.008 0.217 0.275 0.199 0.203 0.168 0.155 0.053 0.177 0.053 0.256
#>   [61] 0.069 0.036 0.110 0.169 0.030 0.036 0.087 0.222 0.011 0.027 0.152 0.020
#>   [73] 0.127 0.301 0.111 0.394 0.109 0.183 0.001 0.173 0.403 0.034 0.091 0.280
#>   [85] 0.033 0.062 0.087 0.013 0.018 0.279 0.160 0.025 0.058 0.268 0.227 0.121
#>   [97] 0.041 0.151 0.295 0.307 0.428 0.082 0.042 0.265 0.015 0.091 0.247 0.047
#>  [109] 0.026 0.031 0.084 0.224 0.265 0.189 0.058 0.031 0.324 0.288 0.251 0.117
#>  [121] 0.377 0.323 0.226 0.418 0.300 0.004 0.325 0.062 0.044 0.089 0.025 0.302
#>  [133] 0.013 0.252 0.386 0.271 0.277 0.244 0.069 0.092 0.407 0.035 0.226 0.125
#>  [145] 0.431 0.247 0.179 0.081 0.108 0.205 0.013 0.267 0.198 0.054 0.010 0.007
#>  [157] 0.261 0.002 0.013 0.177 0.107 0.329 0.122 0.350 0.113 0.346 0.239 0.045
#>  [169] 0.350 0.136 0.010 0.181 0.028 0.035 0.116 0.148 0.116 0.121 0.110 0.190
#>  [181] 0.035 0.241 0.004 0.375 0.204 0.187 0.003 0.074 0.081 0.102 0.228 0.036
#>  [193] 0.225 0.022 0.021 0.034 0.020 0.097 0.135 0.150 0.083 0.306 0.287 0.142
#>  [205] 0.446 0.162 0.181 0.061 0.101 0.158 0.147 0.170 0.227 0.114 0.016 0.155
#>  [217] 0.129 0.298 0.163 0.162 0.008 0.019 0.015 0.046 0.041 0.079 0.267 0.064
#>  [229] 0.007 0.272 0.070 0.036 0.005 0.103 0.008 0.389 0.002 0.030 0.060 0.378
#>  [241] 0.259 0.032 0.092 0.306 0.309 0.127 0.051 0.024 0.015 0.363 0.278 0.077
#>  [253] 0.204 0.126 0.178 0.004 0.067 0.251 0.003 0.312 0.124 0.193 0.095 0.063
#>  [265] 0.036 0.023 0.129 0.375 0.160 0.308 0.255 0.095 0.418 0.140 0.105 0.022
#>  [277] 0.045 0.208 0.033 0.394 0.389 0.321 0.240 0.117 0.074 0.019 0.026 0.239
#>  [289] 0.037 0.010 0.152 0.117 0.100 0.202 0.183 0.064 0.212 0.051 0.044 0.390
#>  [301] 0.154 0.036 0.367 0.067 0.135 0.089 0.016 0.396 0.225 0.010 0.408 0.270
#>  [313] 0.431 0.289 0.160 0.294 0.075 0.066 0.005 0.044 0.067 0.048 0.461 0.253
#>  [325] 0.056 0.074 0.121 0.120 0.395 0.013 0.171 0.127 0.120 0.003 0.082 0.149
#>  [337] 0.108 0.009 0.151 0.013 0.047 0.177 0.035 0.002 0.005 0.054 0.068 0.121
#>  [349] 0.144 0.010 0.055 0.069 0.194 0.080 0.210 0.429 0.050 0.300 0.072 0.185
#>  [361] 0.016 0.076 0.031 0.062 0.234 0.194 0.117 0.226 0.177 0.056 0.338 0.064
#>  [373] 0.131 0.095 0.013 0.247 0.034 0.002 0.225 0.350 0.004 0.459 0.098 0.051
#>  [385] 0.280 0.125 0.012 0.052 0.015 0.046 0.218 0.191 0.084 0.180 0.156 0.024
#>  [397] 0.131 0.281 0.044 0.062 0.451 0.211 0.099 0.022 0.025 0.006 0.306 0.305
#>  [409] 0.021 0.007 0.054 0.024 0.253 0.145 0.283 0.112 0.287 0.013 0.136 0.097
#>  [421] 0.028 0.110 0.092 0.187 0.035 0.043 0.242 0.060 0.100 0.045 0.366 0.222
#>  [433] 0.001 0.096 0.086 0.289 0.432 0.095 0.094 0.090 0.207 0.470 0.186 0.033
#>  [445] 0.104 0.055 0.002 0.160 0.024 0.092 0.008 0.100 0.014 0.005 0.083 0.112
#>  [457] 0.116 0.022 0.080 0.138 0.227 0.075 0.305 0.048 0.118 0.175 0.025 0.238
#>  [469] 0.051 0.176 0.046 0.186 0.017 0.174 0.091 0.199 0.082 0.261 0.237 0.168
#>  [481] 0.065 0.102 0.141 0.034 0.218 0.191 0.063 0.228 0.405 0.258 0.281 0.070
#>  [493] 0.310 0.080 0.029 0.017 0.226 0.191 0.161 0.153 0.252 0.105 0.107 0.058
#>  [505] 0.224 0.006 0.433 0.083 0.313 0.163 0.061 0.044 0.344 0.207 0.135 0.134
#>  [517] 0.150 0.094 0.162 0.092 0.017 0.116 0.431 0.241 0.035 0.101 0.246 0.281
#>  [529] 0.195 0.105 0.004 0.412 0.174 0.009 0.063 0.424 0.353 0.375 0.144 0.015
#>  [541] 0.304 0.042 0.001 0.006 0.192 0.032 0.127 0.018 0.009 0.077 0.007 0.129
#>  [553] 0.002 0.151 0.099 0.237 0.024 0.367 0.208 0.020 0.244 0.014 0.264 0.046
#>  [565] 0.025 0.016 0.057 0.031 0.162 0.005 0.078 0.015 0.072 0.075 0.104 0.360
#>  [577] 0.081 0.156 0.069 0.213 0.292 0.020 0.147 0.118 0.135 0.004 0.089 0.058
#>  [589] 0.241 0.101 0.062 0.020 0.096 0.099 0.175 0.420 0.070 0.182 0.392 0.033
#>  [601] 0.014 0.029 0.281 0.119 0.074 0.122 0.055 0.385 0.157 0.024 0.018 0.031
#>  [613] 0.170 0.118 0.067 0.159 0.044 0.041 0.394 0.184 0.027 0.092 0.046 0.074
#>  [625] 0.040 0.222 0.020 0.012 0.022 0.099 0.135 0.282 0.029 0.015 0.020 0.062
#>  [637] 0.155 0.179 0.081 0.068 0.214 0.345 0.201 0.048 0.293 0.041 0.160 0.249
#>  [649] 0.173 0.023 0.143 0.020 0.327 0.301 0.062 0.124 0.128 0.054 0.292 0.238
#>  [661] 0.001 0.110 0.013 0.059 0.114 0.042 0.181 0.363 0.272 0.144 0.012 0.017
#>  [673] 0.355 0.154 0.361 0.103 0.114 0.127 0.017 0.199 0.004 0.032 0.418 0.001
#>  [685] 0.205 0.099 0.055 0.061 0.254 0.093 0.302 0.314 0.111 0.207 0.173 0.009
#>  [697] 0.416 0.333 0.020 0.050 0.013 0.008 0.027 0.148 0.112 0.340 0.259 0.106
#>  [709] 0.074 0.192 0.313 0.222 0.235 0.044 0.157 0.157 0.103 0.073 0.123 0.083
#>  [721] 0.113 0.068 0.260 0.029 0.123 0.256 0.077 0.120 0.168 0.001 0.031 0.002
#>  [733] 0.343 0.319 0.107 0.104 0.126 0.044 0.142 0.078 0.276 0.128 0.091 0.010
#>  [745] 0.001 0.222 0.059 0.038 0.062 0.070 0.218 0.406 0.181 0.076 0.065 0.127
#>  [757] 0.077 0.007 0.010 0.153 0.115 0.036 0.211 0.011 0.009 0.134 0.098 0.245
#>  [769] 0.036 0.027 0.086 0.058 0.029 0.210 0.299 0.166 0.096 0.023 0.197 0.005
#>  [781] 0.036 0.060 0.329 0.045 0.101 0.031 0.006 0.186 0.103 0.050 0.039 0.115
#>  [793] 0.389 0.029 0.142 0.438 0.241 0.246 0.200 0.004 0.360 0.139 0.008 0.003
#>  [805] 0.441 0.055 0.427 0.002 0.196 0.302 0.246 0.162 0.004 0.201 0.059 0.057
#>  [817] 0.027 0.086 0.066 0.004 0.221 0.011 0.074 0.267 0.323 0.022 0.046 0.140
#>  [829] 0.306 0.296 0.109 0.048 0.049 0.123 0.169 0.098 0.315 0.081 0.047 0.168
#>  [841] 0.253 0.267 0.219 0.222 0.049 0.094 0.360 0.139 0.358 0.042 0.287 0.200
#>  [853] 0.057 0.229 0.001 0.183 0.162 0.131 0.087 0.343 0.089 0.016 0.473 0.359
#>  [865] 0.239 0.133 0.217 0.018 0.033 0.268 0.136 0.042 0.105 0.332 0.212 0.267
#>  [877] 0.315 0.347 0.264 0.337 0.341 0.007 0.006 0.127 0.164 0.300 0.163 0.100
#>  [889] 0.406 0.042 0.151 0.112 0.188 0.038 0.045 0.138 0.308 0.177 0.304 0.227
#>  [901] 0.030 0.209 0.130 0.207 0.012 0.021 0.036 0.214 0.382 0.044 0.339 0.192
#>  [913] 0.367 0.136 0.088 0.231 0.030 0.209 0.257 0.022 0.009 0.327 0.012 0.117
#>  [925] 0.408 0.171 0.036 0.056 0.011 0.161 0.229 0.290 0.304 0.132 0.206 0.079
#>  [937] 0.109 0.425 0.050 0.109 0.145 0.100 0.065 0.093 0.232 0.150 0.102 0.139
#>  [949] 0.002 0.086 0.084 0.107 0.078 0.073 0.017 0.267 0.023 0.105 0.006 0.069
#>  [961] 0.178 0.033 0.021 0.204 0.027 0.089 0.414 0.162 0.022 0.094 0.112 0.062
#>  [973] 0.267 0.130 0.202 0.013 0.150 0.210 0.116 0.403 0.096 0.025 0.118 0.121
#>  [985] 0.236 0.066 0.148 0.266 0.237 0.032 0.169 0.035 0.100 0.147 0.029 0.285
#>  [997] 0.107 0.001 0.286 0.181